% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{aicc}
\alias{aicc}
\title{New AICc-function. The X here needs to include a .id column as the first column in X, which now is a data.table object, identifying how the diagonal should be in the H-matrix}
\usage{
aicc(h.vec, y, X, negative = FALSE, kernel = "Euclidean",
  scale_var = T, S_scale_dist = F, idcol = T)
}
\arguments{
\item{h.vec}{Vector of bandwidths of size \code{ncol(X})}

\item{X}{Needs to include a .id column as the first column in X, which now is a data.table object, identifying how the diagonal should be in the H-matrix}
}
\value{
H-function output.
}
\description{
Need to improve this help file.
}
\author{
Martin Jullum
}
