% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aicc_full_cpp}
\alias{aicc_full_cpp}
\title{AICc formula for several sets, alternative definition}
\usage{
aicc_full_cpp(h, x_list, mcov_list, S_scale_dist, y_list, negative)
}
\arguments{
\item{h}{numeric specifying the scaling (sigma)}

\item{S_scale_dist}{logical indicating whether the Mahalanobis distance should be scaled with the number of variables}

\item{x}{matrix with "covariates"}

\item{mcov}{covariance matrix}

\item{y}{vector with the "response variable"}
}
\value{
Scalar with the numeric value of the AICc formula
}
\description{
AICc formula for several sets, alternative definition
}
\author{
Martin Jullum
}
