% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AICc_full_cpp_alt}
\alias{AICc_full_cpp_alt}
\title{AICc formula for several sets, alternative definition}
\usage{
AICc_full_cpp_alt(h, X_list, mcov_list, S_scale_dist, y_list, negative)
}
\arguments{
\item{h}{numeric specifying the scaling (sigma)}

\item{S_scale_dist}{logical indicating whether the Mahlanobis distance should be scaled with the number of variables}

\item{X}{matrix with "covariates"}

\item{mcov}{covariance matrix}

\item{y}{vector with the "response variable"}
}
\value{
Scalar with the numeric value of the AICc formula
}
\description{
AICc formula for several sets, alternative definition
}
\author{
Martin Jullum
}
