% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aicc_full_single_cpp}
\alias{aicc_full_single_cpp}
\title{Temp-function for computing the full AICc with several X's etc}
\usage{
aicc_full_single_cpp(X, mcov, S_scale_dist, h, y)
}
\arguments{
\item{X}{matrix.}

\item{mcov}{matrix
The covariance matrix of X.}

\item{S_scale_dist}{logical.
Indicating whether the Mahalanobis distance should be scaled with the number of variables}

\item{h}{numeric specifying the scaling (sigma)}

\item{y}{Vector
Representing the (temporary) response variable}
}
\value{
Scalar with the numeric value of the AICc formula.
}
\description{
Temp-function for computing the full AICc with several X's etc
}
\author{
Martin Jullum
}
\keyword{internal}
