% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vS.R
\name{append_vS_list}
\alias{append_vS_list}
\title{Append the New \code{vS_list} to the Previous \code{vS_list}}
\usage{
append_vS_list(vS_list, internal)
}
\arguments{
\item{vS_list}{List.
Output from \code{\link[=compute_vS]{compute_vS()}}.}

\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
The vS_list after being merged with previously computed vS_lists (stored in internal)
}
\description{
Append the New \code{vS_list} to the Previous \code{vS_list}
}
\keyword{internal}
