% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vS.R
\name{batch_compute_vS}
\alias{batch_compute_vS}
\title{Batch computing \code{v(S)} for all features subsets \code{S}.}
\usage{
batch_compute_vS(S, internal, model, predict_model, p = NULL, output_size = 1)
}
\arguments{
\item{output_size}{The number of outputs per row of new data from predict_model}
}
\description{
Batch computing \code{v(S)} for all features subsets \code{S}.
}
\keyword{internal}
