% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_and_casual_Shapley.R
\name{check_coalitions_respect_order}
\alias{check_coalitions_respect_order}
\title{Auxiliary function that verifies that the coalitions respect the causal order}
\usage{
check_coalitions_respect_order(coalitions, causal_ordering)
}
\arguments{
\item{coalitions}{List of integer vectors containing the coalitions indicating which
features to conditioning on that we are to check against the causal ordering.}

\item{causal_ordering}{List of vectors containing the partial causal ordering.
The elements in the list represents the components in the causal ordering and can either
be a single feature index or several, that is, a vector. For example, we can have
\code{list(c(1,2), c(3, 4))}, which means that \verb{1,2 -> 3} and \verb{1,2 -> 4}, i.e., one and
two are the ancestors of three and four, but three and four are not related.}
}
\value{
Logical array indicating whether the coalitions respect the causal order or not.
}
\description{
Auxiliary function that verifies that the coalitions respect the causal order
}
\examples{
coalitions <- list(c(1, 2, 3, 5), c(1, 2, 4), c(1, 4))
causal_ordering <- list(1:3, 4:7, 8:10)
check_coalitions_respect_order(coalitions, causal_ordering) # c(TRUE, FALSE, FALSE)
check_coalitions_respect_order(c(1, 2, 3, 5), causal_ordering) # TRUE
check_coalitions_respect_order(list(c(1, 2, 3, 5)), causal_ordering) # TRUE
check_coalitions_respect_order(list(c(1, 2, 5)), causal_ordering) # FALSE
check_coalitions_respect_order(list(c(1:7, 10)), causal_ordering) # TRUE
check_coalitions_respect_order(list(c(1:3, 5:6, 10)), causal_ordering) # FALSE

}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
