% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Checks the convergence according to the convergence threshold}
\usage{
check_convergence(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
The (updated) internal list
}
\description{
Checks the convergence according to the convergence threshold
}
\keyword{internal}
