% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{check_features}
\alias{check_features}
\title{Checks that two extracted feature lists have exactly the same properites}
\usage{
check_features(
  f_list_1,
  f_list_2,
  name_1 = "model",
  name_2 = "data",
  use_1_as_truth = T
)
}
\arguments{
\item{f_list_1, f_list_2}{List. As extracted from either \code{get_data_features} or \code{get_model_features}.}

\item{name_1, name_2}{Character. Names corresponding to the input of respectively \code{f_list_1} and \code{f_list_2},
used to provide informative error messages.}

\item{use_1_as_truth}{Logical. If TRUE, \code{f_list_2} is compared to \code{f_list_1}, i.e. additional elements
is allowed in \code{f_list_2}, and if \code{f_list_1}'s feature classes contains NA's, feature class check is
ignored regardless of what is specified in \code{f_list_1}. If FALSE, \code{f_list_1} and \code{f_list_2} are
equated and they need to contain exactly the same elements. Set to TRUE when comparing a model and data, and FALSE
when comparing two data sets.}
}
\value{
List. The \code{f_list_1} is returned as inserted, but with an additional sorted factor list. This could be
used to reorder the data.
}
\description{
Checks that two extracted feature lists have exactly the same properites
}
\examples{
Load example data
data("Boston", package = "MASS")
# Split data into test- and training data
x_train <- as.data.table(head(Boston))
x_train[,rad:=as.factor(rad)]
data_features <- get_data_features(x_train)

model <- lm(medv ~ lstat + rm + dis + indus, data = x_train)

cnms <- c("lstat", "rm", "dis", "indus")

# Checking that features used by the model corresponds to cnms
model_features <- get_model_features(x = model, feature_labels = NULL)

check_features(model_features,data_features)
}
\keyword{internal}
