% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_and_casual_Shapley.R
\name{check_n_coalitions_causal}
\alias{check_n_coalitions_causal}
\title{Auxiliary function that verifies that the number of coalitions is valid}
\usage{
check_n_coalitions_causal(n_coalitions, causal_ordering)
}
\arguments{
\item{causal_ordering}{List of vectors specifying the components of the partial causal ordering.
The features in each component are given by the feature/group names or their indices.
If \code{causal_ordering} is \code{NULL} (default), then no causal ordering is assumed, i.e., a causal ordering with
a single component containing all features \code{list(1:n_features)} or groups \code{list(1:n_groups)} for
feature-wise and group-wise Shapley values, respectively. If the user specifies the features/groups
by their names (strings), then \code{shapr} will convert them to the feature/group indices specified by the
order in \code{get_model_specs} for features and \code{groups} for groups. For feature-wise Shapley values,
the elements in the \code{causal_ordering} list represents the components in the causal ordering and can either
be a single feature index or a vector of feature indices. For example, \code{list(c(1, 2), c(3, 4))} implies that
\verb{1,2 -> 3} and \verb{1,2 -> 4}, i.e., feature one and two are the ancestors of feature three and four, but feature
three and four are not related. All features/groups must be included in \code{causal_ordering} without any duplicates.}
}
\description{
Auxiliary function that verifies that the number of coalitions is valid
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
