% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prediction.R
\name{check_prediction_R}
\alias{check_prediction_R}
\title{Model testing function}
\usage{
check_prediction_R(x_test, predict_model, model)
}
\arguments{
\item{model}{Objects.
The model object that ought to be explained}
}
\description{
Model testing function
}
\keyword{internal}
