% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{check_regression_parameters}
\alias{check_regression_parameters}
\title{Check regression parameters}
\usage{
check_regression_parameters(internal)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}
}
\value{
The same \code{internal} list, but added logical indicator \code{internal$parameters$regression_tune}
if we are to tune the regression model/models.
}
\description{
Check regression parameters
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
