% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_sage_args}
\alias{check_sage_args}
\title{Check the SAGE Arguments}
\usage{
check_sage_args(sage_args)
}
\arguments{
\item{sage_args}{List of SAGE arguments, see \code{\link[=get_sage_args_default]{get_sage_args_default()}}.}
}
\value{
Nothing. Called for its side effect of aborting on invalid input.
}
\description{
Check the SAGE Arguments
}
\author{
Martin Jullum
}
\keyword{internal}
