% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setup.R
\name{check_setup}
\alias{check_setup}
\title{check_setup}
\usage{
check_setup(
  x_train,
  x_explain,
  model,
  approach,
  prediction_zero,
  n_combinations,
  group,
  n_samples,
  n_batches,
  seed,
  keep_samp_for_vS,
  predict_model,
  get_model_specs,
  ...
)
}
\arguments{
\item{x_train}{...}
}
\description{
check_setup
}
