% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{check_vaeac_extra_named_list}
\alias{check_vaeac_extra_named_list}
\title{Check vaeac.extra_parameters list}
\usage{
check_vaeac_extra_named_list(vaeac.extra_parameters)
}
\arguments{
\item{vaeac.extra_parameters}{List containing the extra parameters to the \code{vaeac} approach}
}
\description{
Check vaeac.extra_parameters list
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
