% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_compute_vS}
\alias{cli_compute_vS}
\title{Printing messages in compute_vS with cli}
\usage{
cli_compute_vS(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\description{
Printing messages in compute_vS with cli
}
\keyword{internal}
