% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coalition_matrix_cpp}
\alias{coalition_matrix_cpp}
\title{Get coalition matrix}
\usage{
coalition_matrix_cpp(coalitions, m)
}
\arguments{
\item{coalitions}{List}

\item{m}{Positive integer. Total number of coalitions}
}
\value{
Matrix
}
\description{
Get coalition matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
