% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{col_fun}
\alias{col_fun}
\title{Function to extract the column number of the conditional expectation matrix as a function of
all the possible x_test values}
\usage{
col_fun(tbl, S_dt)
}
\arguments{
\item{tbl}{Data.table. Consists of all possible x_test values.}

\item{S_dt}{??}
}
\value{
list of column numbers
}
\description{

}
