% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_estimates}
\alias{compute_estimates}
\title{Compute the Shapley Values and Their Standard Deviation Given \code{v(S)}}
\usage{
compute_estimates(internal, vS_list)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}

\item{vS_list}{List.
Output from \code{\link[=compute_vS]{compute_vS()}}.}
}
\value{
The (updated) internal list
}
\description{
Compute the Shapley Values and Their Standard Deviation Given \code{v(S)}
}
\keyword{internal}
