% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{compute_kernelShap}
\alias{compute_kernelShap}
\title{Computes the kernelShap values for the test data given to prepare_kernelShao}
\usage{
compute_kernelShap(model, l, noSamp_MC = 1000, verbose = FALSE,
  cond_approach = "empirical", empirical_settings = list(type =
  "fixed_sigma", fixed_sigma_vec = 0.1, AICc_no_samp_per_optim = 1000,
  AIC_optim_max_eval = 20, AIC_optim_startval = 0.1, w_threshold = 0.95),
  pred_zero, mu = NULL, Sigma = NULL, ensure_condcov_symmetry = F)
}
\arguments{
\item{model}{Model object. Fitted model that is used to produce the predictions}

\item{l}{List. The output from the prepare_kernelShap function}

\item{noSamp_MC}{Postive integer. Indicating the maximum number of samples to use in the Monte Carlo integration for every conditional expectation (previously called n_threshold)}

\item{verbose}{Integer. How much information to print during function execution (in development)}

\item{cond_approach}{String or list. When being a list, the elements in the list refers to the rows in l$X that ought to be included in each of the approaches!}

\item{empirical_settings}{List. Specifying the settings when using the empirial method to compute the conditional expectations.}

\item{pred_zero}{The prediction value for unseen data, typically equal to the mean of the response}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{Sigma}{Numeric matrix. (Optional) Containing the covariance matrix of the data generating distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{ensure_condcov_symmetry}{Logical. Whether to an apply a symmetry ensurance mechanism for the conditional covariance matrices in the Gaussian and copula approaches.
Typically only needed if theoriginal covariance is just barely positive definite.}
}
\value{
List with kernel Shap values (Kshap) and other object used to perform the computation (helpful for debugging etc.)
}
\description{
Computes the kernelShap values for the test data given to prepare_kernelShao
}
\author{
Martin Jullum
}
