% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingauss_extra.R
\name{compute_lingauss_Tmu_Tx}
\alias{compute_lingauss_Tmu_Tx}
\title{Compute Tmu and Tx for lingauss approach}
\usage{
compute_lingauss_Tmu_Tx(
  internal,
  gaussian.mu = NULL,
  gaussian.cov_mat = NULL,
  ...
)
}
\arguments{
\item{internal}{Not used.}

\item{gaussian.mu}{Numeric vector. (Optional)
Containing the mean of the data generating distribution.
\code{NULL} means it is estimated from the \code{x_train}.}

\item{gaussian.cov_mat}{Numeric matrix. (Optional)
Containing the covariance matrix of the data generating distribution.
\code{NULL} means it is estimated from the \code{x_train}.}

\item{...}{\code{approach}-specific arguments. See below.}
}
\description{
Compute Tmu and Tx for lingauss approach
}
