% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{compute_normalization}
\alias{compute_normalization}
\title{Compute Featurewise Means and Standard Deviations}
\usage{
compute_normalization(data, one_hot_max_sizes)
}
\arguments{
\item{data}{A torch_tensor of dimension N x p containing the data.}

\item{one_hot_max_sizes}{A torch tensor of dimension p containing the one hot sizes of the p features.
The sizes for the continuous features can either be '0' or '1'.}
}
\value{
List containing the means and the standard deviations of the different features.
}
\description{
Returns the means and standard deviations for all continuous features in the data set.
Categorical features get mean=0 and sd=1 by default.
}
