% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{compute_time}
\alias{compute_time}
\title{Gathers and computes the timing of the different parts of the explain function.}
\usage{
compute_time(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
List of reformatted timing information
}
\description{
Gathers and computes the timing of the different parts of the explain function.
}
\keyword{internal}
