% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_vS_loss}
\alias{compute_vS_loss}
\title{Compute the SAGE Value Function}
\usage{
compute_vS_loss(dt_vS, internal)
}
\arguments{
\item{dt_vS}{Data.table with the contribution function \code{v(S)} estimates for each coalition.}

\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}
}
\value{
A single-column numeric matrix with one row per coalition in \code{dt_vS}.
}
\description{
Compute the SAGE Value Function
}
\details{
For each coalition (row of \code{dt_vS}), the value function used for the SAGE values is the negative model
loss \code{-loss_func(y_explain, v(S))}, where \code{v(S)} is the vector of conditional expectations across the explained
observations.
}
\author{
Martin Jullum
}
\keyword{internal}
