% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R, R/observations.R
\name{cond_expec_new}
\alias{cond_expec_new}
\title{Function to calculate conditional expectations of the cutoff jointly Normal random variables
for the x_test observations. I.e. doing what cond_expec + extract_cond_expec does together,
just much faster.}
\usage{
cond_expec_new(cond_list, explainer, x_test)

cond_expec_new(cond_list, explainer, x_test)
}
\arguments{
\item{cond_list}{List. Calculated using the \code{cond_prob} function.}

\item{explainer}{explainer object from \code{shapr} package.}

\item{x_test}{Matrix. Consists of all the test observations. Has the same dimension
as the number of joint Normal random variables calculated in \code{sim_true_Normal} function.}

\item{cond_expec_dt}{data.table. Calculated using the \code{cond_expec} function.}

\item{prediction_zero}{Numeric. Number to assigned to phi_0 in Shapley framework.}

\item{joint_prob_dt}{data.table The first element in the list calculated using the \code{sim_true_Normal} function.}
}
\value{
data.table

data.table
}
\description{



}
