% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{condMVN_modified}
\alias{condMVN_modified}
\title{Calculate conditional mu and Sigma}
\usage{
condMVN_modified(mean, sigma, dependent.ind, given.ind, X.given,
  check.sigma = TRUE)
}
\arguments{
\item{sigma}{Numeric}
}
\value{
Numeric
}
\description{
Copied from the \link[condMVNorm]{condMVN} package, but added a line to
ensure symmetry of resulting conditional covariance matrix
}
\author{
Martin Jullum
}
