% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{continue_train_VAEAC_model_shapr}
\alias{continue_train_VAEAC_model_shapr}
\title{Continue to Train the VAEAC model}
\usage{
continue_train_VAEAC_model_shapr(
  explanation,
  epochs_new,
  lr_new = NULL,
  training_data = NULL,
  save_data = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{epochs_new}{Integer. The number of extra epochs to conduct.}

\item{lr_new}{Numeric. If we are to overwrite the old learning rate in the adam optimizer.}

\item{training_data}{Matrix/data.frame containing new training data. If not present, then we try to load training data from the VAEAC_model.}

\item{save_data}{Boolean. If we are to save the training data.}

\item{verbose}{Boolean. If we are to print out information to the user.}

\item{VAEAC_model}{The output revived from the 'train_VAEAC_model' function.}
}
\value{
A list containing the training/validation errors and paths to where the VAEAC models are saved on the disk.
}
\description{
Function that loads a previously trained VAEAC model and continue the training, either
on new data or on the same dataset as it was trained on before. If we are given a new dataset, then
we assume that new dataset has the same distribution and one_hot_max_sizes as the original dataset.
}
