% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_and_casual_Shapley.R
\name{convert_feature_name_to_idx}
\alias{convert_feature_name_to_idx}
\title{Convert feature names into feature indices}
\usage{
convert_feature_name_to_idx(causal_ordering, labels, feat_group_txt)
}
\arguments{
\item{causal_ordering}{List of vectors specifying the components of the partial causal ordering.
The features in each component are given by the feature/group names or their indices.
If \code{causal_ordering} is \code{NULL} (default), then no causal ordering is assumed, i.e., a causal ordering with
a single component containing all features \code{list(1:n_features)} or groups \code{list(1:n_groups)} for
feature-wise and group-wise Shapley values, respectively. If the user specifies the features/groups
by their names (strings), then \code{shapr} will convert them to the feature/group indices specified by the
order in \code{get_model_specs} for features and \code{groups} for groups. For feature-wise Shapley values,
the elements in the \code{causal_ordering} list represents the components in the causal ordering and can either
be a single feature index or a vector of feature indices. For example, \code{list(c(1, 2), c(3, 4))} implies that
\verb{1,2 -> 3} and \verb{1,2 -> 4}, i.e., feature one and two are the ancestors of feature three and four, but feature
three and four are not related. All features/groups must be included in \code{causal_ordering} without any duplicates.}

\item{labels}{Vector of strings containing (the order of) the feature names.}

\item{feat_group_txt}{String that is either "feature" or "group" based on
if \code{shapr} is computing feature- or group-wise Shapley values}
}
\value{
The \code{causal_ordering} list, but with feature indices (w.r.t. \code{labels}) instead of feature names.
}
\description{
Functions that takes a \code{causal_ordering} specified using strings and convert these strings to feature indices.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
