% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupVariables.R
\name{corr_rect_hclust}
\alias{corr_rect_hclust}
\title{Draw rectangles on the correlation matrix graph}
\usage{
corr_rect_hclust(corr, cluster, k = 2, col = "yellow", lwd = 6)
}
\arguments{
\item{corr}{correlation matrix}

\item{cluster}{cluster object from hclust}

\item{k}{number of clusters}

\item{col}{box colour}

\item{lwd}{box line width}
}
\description{
Draw rectangles on the correlation matrix graph
}
\author{
Anders Løland
}
