% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correction_matrix_cpp}
\alias{correction_matrix_cpp}
\title{Correction term with trace_input in AICc formula}
\usage{
correction_matrix_cpp(tr_H, n)
}
\arguments{
\item{tr_H}{numeric
The trace of H}

\item{n}{numeric
The number of rows in H}
}
\value{
Scalar
}
\description{
Correction term with trace_input in AICc formula
}
\author{
Martin Jullum
}
\keyword{internal}
