% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{default_doc}
\alias{default_doc}
\title{Unexported documentation helper function}
\usage{
default_doc()
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within
\code{\link[shapr:explain]{shapr:::explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data}, \code{functions}, \code{output}.}

\item{model}{Objects.
The model object that ought to be explained}
}
\value{
List \code{internal}
It holds all parameters, data, functions and computed objects used within
\code{\link[shapr:explain]{shapr:::explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data}, \code{functions}, \code{output}. [explain()]
[roxygen2::roxygenize()]
[explain()]
}
\description{
Unexported documentation helper function
}
\keyword{internal}
