% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{default_doc_explain}
\alias{default_doc_explain}
\title{Exported documentation helper function.}
\usage{
default_doc_explain()
}
\arguments{
\item{internal}{Not used.}
}
\description{
Exported documentation helper function.
}
\keyword{internal}
