% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{default_doc_export}
\alias{default_doc_export}
\title{Exported documentation helper function.}
\usage{
default_doc_export(internal, iter, index_features, digits, nsmall)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}

\item{iter}{Integer.
The iteration number. Only used internally.}

\item{index_features}{Positive integer vector. Specifies the id_coalition to
apply to the present method. \code{NULL} means all coalitions. Only used internally.}

\item{digits}{Integer.
Number of significant digits to display.
Defaults to 2.}

\item{nsmall}{Integer.
Minimum digits to the right of the decimal point.
Defaults to 0.}
}
\description{
Exported documentation helper function.
}
\keyword{internal}
