% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{evaluate_approach}
\alias{evaluate_approach}
\title{MSE of Frye and Covert.}
\usage{
evaluate_approach(
  explanation,
  return_individual = TRUE,
  return_df = FALSE,
  return_p_hat_avg = TRUE,
  return_coalition = TRUE
)
}
\arguments{
\item{explanation}{object returned from explain() function in the shapr package.
Note that the 'prediction' function must have been updated to the version in this file.
I.e., that explanation keeps the simulated data.}

\item{return_individual}{Boolean. If we are to return the error on individual basis.}

\item{return_df}{Boolean. If we are to return the data frame.}

\item{return_p_hat_avg}{Boolean. If we are to return the average prediction.}

\item{return_coalition}{Boolean. If we are to return the results for the different coalitions.}

\item{explainer}{object returned from the shapr() function in the shapr package.}
}
\description{
Function that computes the MSE's proposed in Frye and Covert.
So it does not need the true Shapley values.
}
