% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forecast.R
\name{explain_forecast}
\alias{explain_forecast}
\title{Explain a forecast.
TODO: Write documentation.}
\usage{
explain_forecast(
  model,
  data,
  reg = NULL,
  train_idx,
  explain_idx,
  lags,
  horizon,
  approach,
  prediction_zero,
  n_combinations = NULL,
  group_lags = TRUE,
  n_samples = 1000,
  n_batches = 1,
  seed = 1,
  keep_samp_for_vS = FALSE,
  predict_model = NULL,
  get_model_specs = NULL,
  ...
)
}
\description{
Explain a forecast.
TODO: Write documentation.
}
