% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_lingauss.R
\name{explain_lingauss}
\alias{explain_lingauss}
\title{Explain the output of a linear model with Gaussian distributed features, using Shapley values values}
\usage{
explain_lingauss(
  model,
  x_explain,
  x_train,
  n_permutations = NULL,
  group = NULL,
  seed = 1,
  predict_model = NULL,
  get_model_specs = NULL,
  timing = TRUE,
  ...
)
}
\arguments{
\item{model}{The model whose predictions we want to explain.
Run \code{\link[=get_supported_models]{get_supported_models()}}
for a table of which models \code{explain} supports natively. Unsupported models
can still be explained by passing \code{predict_model} and (optionally) \code{get_model_specs},
see details for more information.}

\item{x_explain}{A matrix or data.frame/data.table.
Contains the the features, whose predictions ought to be explained.}

\item{x_train}{Matrix or data.frame/data.table.
Contains the data used to estimate the (conditional) distributions for the features
needed to properly estimate the conditional expectations in the Shapley formula.}

\item{n_permutations}{Integer. Number of permutations to sample when estimating the Shapley values with
the permutation approach. If \code{NULL}, all permutations are used, which corresponds to exact computation of the
Shapley values (under the linear + Gaussian assumption).
The maximum number of permutations equals \verb{m!}, where \code{m} is the number of features.}

\item{group}{List.
If \code{NULL} regular feature wise Shapley values are computed.
If provided, group wise Shapley values are computed. \code{group} then has length equal to
the number of groups. The list element contains character vectors with the features included
in each of the different groups.}

\item{seed}{Positive integer.
Specifies the seed before any randomness based code is being run.
If \code{NULL} the seed will be inherited from the calling environment.}

\item{predict_model}{Function.
The prediction function used when \code{model} is not natively supported.
(Run \code{\link[=get_supported_models]{get_supported_models()}} for a list of natively supported
models.)
The function must have two arguments, \code{model} and \code{newdata} which specify, respectively, the model
and a data.frame/data.table to compute predictions for. The function must give the prediction as a numeric vector.
\code{NULL} (the default) uses functions specified internally.
Can also be used to override the default function for natively supported model classes.}

\item{get_model_specs}{Function.
An optional function for checking model/data consistency when \code{model} is not natively supported.
(Run \code{\link[=get_supported_models]{get_supported_models()}} for a list of natively supported
models.)
The function takes \code{model} as argument and provides a list with 3 elements:
\describe{
\item{labels}{Character vector with the names of each feature.}
\item{classes}{Character vector with the classes of each features.}
\item{factor_levels}{Character vector with the levels for any categorical features.}
}
If \code{NULL} (the default) internal functions are used for natively supported model classes, and the checking is
disabled for unsupported model classes.
Can also be used to override the default function for natively supported model classes.}

\item{timing}{Logical.
Whether the timing of the different parts of the \code{explain()} should saved in the model object.}

\item{...}{Further arguments passed to specific approaches}
}
\description{
Explain the output of a linear model with Gaussian distributed features, using Shapley values values
}
\author{
Martin Jullum
}
