% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_combinations}
\alias{feature_combinations}
\title{Define feature combinations, and fetch additional information about each unique combination}
\usage{
feature_combinations(m, exact = TRUE, n_combinations = 200,
  shapley_weight_inf_replacement = 10^6, reduce_dim = TRUE)
}
\arguments{
\item{m}{Positive integer. Total number of features.}

\item{exact}{Logical. If \code{TRUE} we generate all \code{2^m} combinations, otherwise a
subsample of the combinations.}

\item{n_combinations}{Positive integer.}

\item{shapley_weight_inf_replacement}{Numeric.}

\item{reduce_dim}{Logical.}
}
\value{
A data.table that contains the following columns:
\describe{
\item{ID}{Positive integer. Represents a unique key for each combination. The table
is sorted by \code{ID}, so \code{x[[ID]] = 1:nrow(x)} }
\item{features}{List. If combination \code{i} is represented by \code{features[[i]]} is an
integer vector of representing feature combination \code{i}}
\item{nfeatures}{Positive integer. \code{nfeatures[i]} equals the number of features in combination
\code{i}, i.e. \code{nfeatures[i] = length(features[[i]])}}
\item{N}{Positive integer}
}
}
\description{

}
\details{
TODO: Write some additional information about exact
}
\examples{
# All combinations
x <- feature_combinations(m = 5)
nrow(x) # Equals 2^5 = 32

# Subsample of combinations
x <- feature_combinations(m = 13, n_combinations = 1e3)
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
