% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_group}
\alias{feature_group}
\title{Analogue to feature_exact, but for groups instead.}
\usage{
feature_group(group_num, weight_zero_m = 10^6)
}
\arguments{
\item{group_num}{List. Contains vector of integers indicating the feature numbers for the
different groups.}

\item{weight_zero_m}{Positive integer. Represents the Shapley weight for two special
cases, i.e. the case where you have either \code{0} or \code{m} features/feature groups.}
}
\value{
data.table with all feature group combinations, shapley weights etc.
}
\description{
Analogue to feature_exact, but for groups instead.
}
\keyword{internal}
