% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{feature_matrix_cpp}
\alias{feature_matrix_cpp}
\title{Get feature matrix}
\usage{
feature_matrix_cpp(features, m)
}
\arguments{
\item{features}{List}

\item{m}{Positive integer. Total number of features}
}
\value{
Matrix
}
\description{
Get feature matrix
}
\author{
Nikolai Sellereite
}
