% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_computation.R
\name{feature_not_exact}
\alias{feature_not_exact}
\title{Sample feature combinations when not using all possible combinations}
\usage{
feature_not_exact(
  m,
  n_combinations = 200,
  weight_zero_m = 10^6,
  sampling_method = c("unique", "unique_SW", "unique_paired", "unique_paired_SW",
    "non_unique", "non_unique_SW", "chronological_order_increasing",
    "chronological_order_decreasing", "largest_weights",
    "largest_weights_combination_size", "smallest_weights",
    "smallest_weights_constant_SW", "smallest_weights_combination_size",
    "specific_coalition_set", "pilot_estimates_paired"),
  pilot_estimates_vS = NULL,
  specific_coalition_set = NULL,
  specific_coalition_set_weights = NULL
)
}
\arguments{
\item{m}{Positive integer. Total number of features.}

\item{n_combinations}{Positive integer. Note that if \code{exact = TRUE},
\code{n_combinations} is ignored. However, if \code{m > 12} you'll need to add a positive integer
value for \code{n_combinations}.}

\item{weight_zero_m}{Positive integer. Represents the Shapley weight for two special
cases, i.e. the case where you have either \code{0} or \code{m} features/feature groups.}

\item{sampling_method}{String. Specifying which of the coalition sampling methods to use.}

\item{pilot_estimates_vS}{LARS ADD DESCRIPTION}

\item{specific_coalition_set}{Integers. Array of integers of length \code{n_combinations}, where the first and last
entries are 1 and 2^m, respectively.}

\item{specific_coalition_set_weights}{Numerics. Array of numerics of length \code{n_combinations} where the i'th entry
represents the weight of the i'th coalitions specified in \code{specific_coalition_set}. If \code{NULL}, then the function
use the Shapley kernel weights. Note that the weights for the empty and full coalitions are determined by
\code{weight_zero_m}.}
}
\description{
Sample feature combinations when not using all possible combinations
}
\keyword{internal}
