% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{features}
\alias{features}
\alias{features.default}
\alias{features.lm}
\alias{features.glm}
\alias{features.ranger}
\alias{features.gam}
\alias{features.xgb.Booster}
\title{Fetch feature labels from model object}
\usage{
features(x, feature_labels = NULL)

\method{features}{default}(x, feature_labels = NULL)

\method{features}{lm}(x, feature_labels = NULL)

\method{features}{glm}(x, feature_labels = NULL)

\method{features}{ranger}(x, feature_labels = NULL)

\method{features}{gam}(x, feature_labels = NULL)

\method{features}{xgb.Booster}(x, feature_labels = NULL)
}
\arguments{
\item{x}{Numeric matrix or data.frame. Contains the data used for training the model.}

\item{feature_labels}{Character vector. The labels/names of the features used for training the model.
Only applicable if you are using a custom model. Otherwise the features in use are extracted from \code{model}.}
}
\description{
Fetch feature labels from model object
}
\keyword{internal}
