% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_explanation.R
\name{finalize_explanation}
\alias{finalize_explanation}
\title{finalize_explanation}
\usage{
finalize_explanation(vS_list, internal)
}
\arguments{
\item{vS_list}{...}
}
\description{
finalize_explanation
}
