% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_convergence_info}
\alias{format_convergence_info}
\title{Internal function to extract formatted info about the (current) convergence state of the shapr call}
\usage{
format_convergence_info(internal, iter)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}

\item{iter}{Integer.
The iteration number. Only used internally.}
}
\description{
To be used by both \code{\link[=print_iter]{print_iter()}} and \link{summary.shapr}
}
\keyword{internal}
