% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_info_basic}
\alias{format_info_basic}
\title{Internal function to extract a vector with formatted info about the shapr call}
\usage{
format_info_basic(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\description{
To be used by both \code{\link[=cli_startup]{cli_startup()}} and \code{\link[=shapr.summary]{shapr.summary()}}
}
\keyword{internal}
