% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_info_extra}
\alias{format_info_extra}
\title{Internal function to extract some extra formatted info about the shapr call}
\usage{
format_info_extra(internal)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}
}
\description{
To be used in \code{\link[=summary.shapr]{summary.shapr()}}
}
\keyword{internal}
