% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_round}
\alias{format_round}
\title{Format numbers with rounding}
\usage{
format_round(x, digits)
}
\arguments{
\item{x}{Numeric vector. The numbers to format.}

\item{digits}{Integer. The number of digits to round to.}
}
\value{
Character vector. The formatted numbers.
}
\description{
Format numbers with rounding
}
\keyword{internal}
