% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\docType{data}
\name{gauss_cat_loss}
\alias{gauss_cat_loss}
\title{A \code{\link[torch:nn_module]{torch::nn_module()}} Representing a \code{gauss_cat_loss}}
\format{
An object of class \code{logical} of length 1.
}
\usage{
gauss_cat_loss
}
\description{
The \verb{gauss_cat_loss module} layer computes the log probability of the \code{groundtruth} for each object
given the mask and the distribution parameters. That is, the log-likelihoods of the true/full training observations
based on the generative distributions parameters \code{distr_params} inferred by the masked versions of the observations.
}
\details{
Note that the module works with mixed data represented as 2-dimensional inputs and it
works correctly with missing values in \code{groundtruth} as long as they are represented by NaNs.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
