% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\docType{data}
\name{gauss_cat_sampler_most_likely}
\alias{gauss_cat_sampler_most_likely}
\title{A \code{\link[torch:nn_module]{torch::nn_module()}} Representing a \code{gauss_cat_sampler_most_likely}}
\format{
An object of class \code{logical} of length 1.
}
\usage{
gauss_cat_sampler_most_likely
}
\value{
A \code{gauss_cat_sampler_most_likely} object.
}
\description{
The \code{gauss_cat_sampler_most_likely} generates the most likely samples from the generative distribution
defined by the output of the vaeac. I.e., the layer will return the mean and most probable class for the Gaussian
(continuous features) and categorical (categorical features) distributions, respectively.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
