% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\docType{data}
\name{gauss_cat_sampler_random}
\alias{gauss_cat_sampler_random}
\title{A \code{\link[torch:nn_module]{torch::nn_module()}} Representing a gauss_cat_sampler_random}
\format{
An object of class \code{logical} of length 1.
}
\usage{
gauss_cat_sampler_random
}
\description{
The \code{gauss_cat_sampler_random} generates random samples from the generative distribution defined by the
output of the vaeac. The random sample is generated by sampling from the inferred Gaussian and categorical
distributions for the continuous and categorical features, respectively.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
