% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{Gauss_trans_func}
\alias{Gauss_trans_func}
\title{Transforms a sample to standardized normal (dimension 1)}
\usage{
Gauss_trans_func(x)
}
\arguments{
\item{x}{Vector of data to transform}
}
\value{
Vector of transformed data
}
\description{
Transforms a sample to standardized normal (dimension 1)
}
\author{
Martin Jullum
}
