% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{Gauss_trans_func_seperate}
\alias{Gauss_trans_func_seperate}
\title{Transforms new data to standardized normal (dimension 1) based on other datas transformation}
\usage{
Gauss_trans_func_seperate(yx, n_y)
}
\arguments{
\item{yx}{Vector where the first part is the new data to transform,
and last part is the data with the original transformation}

\item{n_z}{How many elements of the zx vector that belongs to the z-part (Gaussian data)}
}
\value{
Vector of back-transformed Gaussian data
}
\description{
Handled in this way in order to allow using the apply function over this function
}
\author{
Martin Jullum
}
