% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{GaussCatLoss}
\alias{GaussCatLoss}
\title{A \code{\link[torch:nn_module]{torch::nn_module()}} Representing a GaussCatLoss}
\usage{
GaussCatLoss(one_hot_max_sizes, min_sigma = 1e-04, min_prob = 1e-04)
}
\arguments{
\item{one_hot_max_sizes}{A vector of integers where the i-th entry is the number of
one-hot encoding for the i-th feature.
I.e., a categorical feature with 5 levels will have a one_hot_max_size of 5.
A feature with a one_hot_max_size of either 0 or 1 will be treated as a continuous feature.}

\item{min_sigma}{For stability it might be desirable that the minimal sigma is not too close to zero.}

\item{min_prob}{For stability it might be desirable that the minimal probability is not too close to zero.}
}
\description{
The GaussCatLoss module/layer computes the log probability
of the \code{groundtruth} for each object given the mask and the distribution parameters.
That is, the log-likelihoods of the true/full training observations based on the
generative distributions parameters \code{distr_params} inferred by the masked versions of the observations.
}
\details{
Note that the module works with mixed data represented as 2-dimensional inputs and it
works correctly with missing values in \code{groundtruth} as long as they are represented by NaNs.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
