% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{GaussCatSampler}
\alias{GaussCatSampler}
\title{A \link[torch:nn_module]{torch::nn_module} Representing a GaussCatSampler}
\usage{
GaussCatSampler(
  one_hot_max_sizes,
  sample_most_probable = FALSE,
  min_sigma = 1e-04,
  min_prob = 1e-04
)
}
\arguments{
\item{one_hot_max_sizes}{A vector of integers where the i-th entry is
the number of one-hot encoding for the i-th feature.
I.e., a categorical feature with 5 levels will have a one_hot_max_size of 5.
A feature with a one_hot_max_size of either 0 or 1 will be treated as a continuous feature.}

\item{sample_most_probable}{A boolean indicating if we are to return the mean and most probable class of
the Gaussian and categorical distributions, respectively (\code{TRUE}), or if sampling from the Gaussian and categorical
distributions is to be performed (\code{FALSE}).}

\item{min_sigma}{For stability it might be desirable that the minimal sigma is not too close to zero.}

\item{min_prob}{For stability it might be desirable that the minimal probability is not too close to zero.}
}
\value{
A \code{GaussCatSampler} object.
}
\description{
The GaussCatSampler generates a sample from the generative distribution defined by
the output of the vaeac. It can either return/generate the most likely values or use random sampling.
}
\section{Fields}{

\describe{
\item{\code{classname}}{Type of torch::nn_module}
}}

\author{
Lars Henry Berge Olsen
}
\keyword{internal}
