% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{GaussCatSamplerRandom}
\alias{GaussCatSamplerRandom}
\title{A \link[torch:nn_module]{torch::nn_module} Representing a GaussCatSamplerRandom}
\usage{
GaussCatSamplerRandom(one_hot_max_sizes, min_sigma = 1e-04, min_prob = 1e-04)
}
\arguments{
\item{one_hot_max_sizes}{A vector of integers where the i-th entry is the number of one-hot encoding
for the i-th feature. I.e., a categorical feature with 5 levels will have a one_hot_max_size of 5.
A feature with a one_hot_max_size of either 0 or 1 will be treated as a continuous feature.}

\item{min_sigma}{For stability it might be desirable that the minimal sigma is not too close to zero.}

\item{min_prob}{For stability it might be desirable that the minimal probability is not too close to zero.}
}
\description{
The GaussCatSamplerRandom generates random samples from the generative
distribution defined by the output of the vaeac. The random sample is generated by
sampling from the inferred Gaussian and categorical distributions for the
continuous and categorical features, respectively.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
