% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{gaussian_transform}
\alias{gaussian_transform}
\title{Transforms a sample to standardized normal (dimension 1)}
\usage{
gaussian_transform(x)
}
\arguments{
\item{x}{Vector of data to transform}
}
\value{
Vector of transformed data
}
\description{
Transforms a sample to standardized normal (dimension 1)
}
\author{
Martin Jullum
}
