% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_transform_cpp}
\alias{gaussian_transform_cpp}
\title{Transforms a sample to standardized normal distribution}
\usage{
gaussian_transform_cpp(x)
}
\arguments{
\item{x}{Numeric matrix. The data which should be transformed to a standard normal distribution.}
}
\value{
Numeric matrix of dimension \code{dim(x)}
}
\description{
Transforms a sample to standardized normal distribution
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
