% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{gaussian_transform_separate}
\alias{gaussian_transform_separate}
\title{Transforms new data to standardized normal (dimension 1) based on other data transformations}
\usage{
gaussian_transform_separate(yx, n_y)
}
\arguments{
\item{yx}{Vector where the first part is the new data to transform,
and last part is the data with the original transformation}

\item{n_y}{Postive integer}
}
\value{
Vector of back-transformed Gaussian data
}
\description{
Handled in this way in order to allow using the apply function over this function
}
\author{
Martin Jullum
}
