% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_transform_separate_cpp}
\alias{gaussian_transform_separate_cpp}
\title{Transforms new data to standardized normal (column-wise) based on other data transformations}
\usage{
gaussian_transform_separate_cpp(y, x)
}
\arguments{
\item{y}{arma::mat. A numeric matrix containing the data that is to be transformed.}

\item{x}{arma::mat. A numeric matrix containing the data of the original transformation.}
}
\value{
An arma::mat matrix of the same dimension as \code{y} containing the back-transformed Gaussian data.
}
\description{
Transforms new data to standardized normal (column-wise) based on other data transformations
}
\author{
Lars Henry Berge Olsen, Martin Jullum
}
\keyword{internal}
