% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_adaptive_arguments_default}
\alias{get_adaptive_arguments_default}
\title{Function to specify arguments of the adaptive estimation procedure}
\usage{
get_adaptive_arguments_default(
  internal,
  max_iter = 20,
  initial_n_combinations = min(200, ceiling((2^internal$parameters$n_features)/10)),
  convergence_tolerance = 0.02,
  reduction_factor_vec = c(seq(0.1, 1, by = 0.1), rep(1, max_iter - 10)),
  n_boot_samps = 100,
  compute_sd = ifelse(internal$parameters$exact, FALSE, TRUE)
)
}
\arguments{
\item{internal}{Not used.}

\item{max_iter}{TODO: Specify}

\item{initial_n_combinations}{TODO: Specify}

\item{convergence_tolerance}{TODO: Specify}

\item{reduction_factor_vec}{TODO: Specify}

\item{n_boot_samps}{TODO: Specify}

\item{compute_sd}{TODO: Specify}
}
\description{
Function to specify arguments of the adaptive estimation procedure
}
\details{
The functions sets default values for the adaptive estimation procedure, according to the function defaults.
If the argument \code{adaptive} of \code{\link[=explain]{explain()}} is FALSE, it sets parameters corresponding to the use of a
non-adaptive estimation procedure
}
\author{
Martin Jullum
}
