% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_gaussian.R
\name{get_cov_mat}
\alias{get_cov_mat}
\title{get_cov_mat}
\usage{
get_cov_mat(x_train, min_eigen_value = 1e-06)
}
\arguments{
\item{x_train}{Matrix or data.frame/data.table.
Data used to estimate the (conditional) feature distributions
needed to properly estimate the conditional expectations in the Shapley formula.}

\item{min_eigen_value}{Numeric
Specifies the smallest allowed eigen value before the covariance matrix of \code{x_train} is assumed to not be
positive definite, and \code{\link[Matrix:nearPD]{Matrix::nearPD()}} is used to find the nearest one.}
}
\description{
get_cov_mat
}
\keyword{internal}
