% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_data_features}
\alias{get_data_features}
\title{Fetches feature labels from a given model object}
\usage{
get_data_features(x)
}
\arguments{
\item{x}{data.table The data to extract feature information from.}
}
\description{
Fetches feature labels from a given model object
}
\examples{
# Load example data
data("Boston", package = "MASS")
# Split data into test- and training data
x_train <- as.data.table(head(Boston))
x_train[,rad:=as.factor(rad)]
get_data_features(x_train)
}
\keyword{internal}
