% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forecast.R
\name{get_data_forecast}
\alias{get_data_forecast}
\title{Set up data for explain_forecast}
\usage{
get_data_forecast(
  y,
  xreg,
  train_idx,
  explain_idx,
  explain_y_lags,
  explain_xreg_lags,
  horizon
)
}
\arguments{
\item{y}{A matrix or numeric vector containing the endogenous variables for the model.
One variable per column, one observation per row.}

\item{xreg}{A matrix containing exogenous regressors for the model.
One variable per column, one observation per row. Should have nrow(data) + horizon rows.}

\item{train_idx}{The observations indices in data to use as training examples.}

\item{explain_idx}{The observations indices in data to explain.}

\item{explain_y_lags}{Numeric vector
Indicates the number of lags of y to include in the explanation.}

\item{explain_xreg_lags}{Numeric vector
Indicates the number of lags of xreg to include in the explanation.}

\item{horizon}{The forecast horizon to explain.}
}
\value{
A list containing
\itemize{
\item The data.frames x_train and x_explain which holds the lagged data examples.
\item A numeric, n_endo denoting how many columns are endogenous in x_train and x_explain.
\item A list, group with groupings of each variable to explain per variable and not per variable and lag.
}
}
\description{
Set up data for explain_forecast
}
