% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_extra_comp_args_default}
\alias{get_extra_comp_args_default}
\title{Gets the default values for the extra estimation arguments}
\usage{
get_extra_comp_args_default(
  internal,
  compute_sd = isFALSE(internal$parameters$exact),
  n_boot_samps = 100,
  max_batch_size = 10,
  min_n_batches = 10
)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}

\item{compute_sd}{Logical. Whether to estimate the standard deviations of the Shapley value estimates. This is TRUE
whenever sampling based kernelSHAP is applied (either iteratively or with a fixed number of coalitions).}

\item{n_boot_samps}{Integer. The number of bootstrapped samples (i.e. samples with replacement) from the set of all
coalitions used to estimate the standard deviations of the Shapley value estimates.}

\item{max_batch_size}{Integer. The maximum number of coalitions to estimate simultaneously within each iteration.
A larger numbers requires more memory, but may have a slight computational advantage.}

\item{min_n_batches}{Integer. The minimum number of batches to split the computation into within each iteration.
Larger numbers gives more frequent progress updates. If parallelization is applied, this should be set no smaller
than the number of parallel workers.}
}
\description{
Gets the default values for the extra estimation arguments
}
\author{
Martin Jullum
}
