% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_extra_parameters}
\alias{get_extra_parameters}
\title{This includes both extra parameters and other objects}
\usage{
get_extra_parameters(internal, type)
}
\description{
This includes both extra parameters and other objects
}
\keyword{internal}
