% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_R.R
\name{get_feature_specs}
\alias{get_feature_specs}
\title{Gets the feature specifications form the model}
\usage{
get_feature_specs(get_model_specs, model)
}
\arguments{
\item{get_model_specs}{Function.
An optional function for checking model/data consistency when \code{model} is not natively supported.
(Run \code{\link[=get_supported_models]{get_supported_models()}} for a list of natively supported models.)
The function takes \code{model} as argument and provides a list with 3 elements:
\describe{
\item{labels}{Character vector with the names of each feature.}
\item{classes}{Character vector with the classes of each features.}
\item{factor_levels}{Character vector with the levels for any categorical features.}
}
If \code{NULL} (the default) internal functions are used for natively supported model classes, and the checking is
disabled for unsupported model classes.
Can also be used to override the default function for natively supported model classes.}

\item{model}{Model object.
Specifies the model whose predictions we want to explain.
Run \code{\link[=get_supported_models]{get_supported_models()}}
for a table of which models \code{explain} supports natively. Unsupported models
can still be explained by passing \code{predict_model} and (optionally) \code{get_model_specs},
see details for more information.}
}
\description{
Gets the feature specifications form the model
}
\keyword{internal}
