% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup_R.R
\name{get_linear_coef}
\alias{get_linear_coef}
\title{Extracts the (purely) linear model coefficients}
\usage{
get_linear_coef(model, feature_specs)
}
\description{
Extracts the (purely) linear model coefficients
}
\keyword{internal}
