% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_and_casual_Shapley.R
\name{get_max_n_coalitions_causal}
\alias{get_max_n_coalitions_causal}
\title{Get the number of coalitions that respects the causal ordering}
\usage{
get_max_n_coalitions_causal(causal_ordering)
}
\arguments{
\item{causal_ordering}{List of vectors containing the partial causal ordering.
The elements in the list represents the components in the causal ordering and can either
be a single feature index or several, that is, a vector. For example, we can have
\code{list(c(1,2), c(3, 4))}, which means that \verb{1,2 -> 3} and \verb{1,2 -> 4}, i.e., one and
two are the ancestors of three and four, but three and four are not related.}
}
\value{
Integer. The (maximum) number of coalitions that respects the causal ordering.
}
\description{
Get the number of coalitions that respects the causal ordering
}
\details{
The function computes the number of coalitions that respects the causal ordering by computing the number
of coalitions in each partial causal component and then summing these. We compute
the number of coalitions in the \eqn{i}th a partial causal component by \eqn{2^n - 1},
where \eqn{n} is the number of features in the the \eqn{i}th partial causal component
and we subtract one as we do not want to include the situation where no features in
the \eqn{i}th partial causal component are present. In the end, we add 1 for the
empty coalition.
}
\examples{
get_max_n_coalitions_causal(list(1:10)) # 2^10 = 1024 (no causal order)
get_max_n_coalitions_causal(list(1:3, 4:7, 8:10)) # 30
get_max_n_coalitions_causal(list(1:3, 4:5, 6:7, 8, 9:10)) # 18
get_max_n_coalitions_causal(list(1:3, c(4, 8), c(5, 7), 6, 9:10)) # 18
get_max_n_coalitions_causal(list(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)) # 11

}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
