% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_gaussian.R
\name{get_mu_vec}
\alias{get_mu_vec}
\title{get_mu_vec}
\usage{
get_mu_vec(x_train)
}
\arguments{
\item{x_train}{Matrix or data.frame/data.table.
Data used to estimate the (conditional) feature distributions
needed to properly estimate the conditional expectations in the Shapley formula.}
}
\description{
get_mu_vec
}
\keyword{internal}
