% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict_model.R
\name{get_predict_model}
\alias{get_predict_model}
\title{Model testing function}
\usage{
get_predict_model(x_test, predict_model, model)
}
\arguments{
\item{model}{Objects.
The model object that ought to be explained}
}
\description{
Model testing function
}
\keyword{internal}
