% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{get_regression_tune}
\alias{get_regression_tune}
\title{Get if model is to be tuned}
\usage{
get_regression_tune(regression_model, regression_tune_values, x_train)
}
\arguments{
\item{regression_model}{A \code{tidymodels} object of class \code{model_specs}. Default is a linear regression model
(\code{\link[parsnip:linear_reg]{parsnip::linear_reg()}}). See \href{https://www.tidymodels.org/find/parsnip/}{tidymodels} for all possible models,
and see the vignette for how to add new/own models.}

\item{regression_tune_values}{Either \code{NULL} (default), a data.frame/data.table/tibble, or a function.
The data.frame must contain the possible hyperparameter value combinations to try.
The column names must match the names of the tuneable parameters specified in \code{regression_model}.
If \code{regression_tune_values} is a function, then it should take one argument \code{x} which is the training data
for the current combination/coalition and returns a data.frame/data.table/tibble with the properties described above.
Using a function allows the hyperparameter values to change based on the size of the combination.}

\item{x_train}{Matrix or data.frame/data.table.
Contains the data used to estimate the (conditional) distributions for the features
needed to properly estimate the conditional expectations in the Shapley formula.}
}
\description{
That is, if the regression model contains hyperparameters we are to tune using cross validation.
See \href{https://www.tidymodels.org/find/parsnip/#model-args}{tidymodels} for default model hyperparameters.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
