% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_results}
\alias{get_results}
\title{Extract components from a shapr object}
\usage{
get_results(
  x,
  what = c("shapley_est", "shapley_sd", "pred_explain", "MSEv", "MSEv_explicand",
    "MSEv_coalition", "iterative_info", "iterative_shapley_est", "iterative_shapley_sd",
    "saving_path", "timing", "parameters", "x_train", "x_explain", "model", "dt_vS",
    "dt_samp_for_vS"),
  ...
)
}
\arguments{
\item{x}{A \code{shapr} object}

\item{what}{Character vector specifying one or more components to extract.
Options:
"shapley_est", "shapley_sd", "pred_explain",
"MSEv", "MSEv_explicand", "MSEv_coalition",
"iterative_info", "iterative_shapley_est", "iterative_shapley_sd",
"saving_path", "timing", "parameters",
"x_train", "x_explain", "model", "dt_vS", "dt_samp_for_vS"}

\item{...}{Not used}
}
\value{
If a single component is requested, returns that object.
If multiple are requested, returns a named list.
}
\description{
Extract components from a shapr object
}
