% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_sage_args_default}
\alias{get_sage_args_default}
\title{Get the Default Values for the SAGE Arguments}
\usage{
get_sage_args_default(loss_func = NULL)
}
\arguments{
\item{loss_func}{Function or \code{NULL}.
The loss function used to measure the model loss when computing the SAGE values.
Must take two arguments, the true response and the model prediction (in that order), and return a single numeric
loss value.
If \code{NULL} (default), logistic (cross-entropy) loss is used for binary responses (values in 0/1) and mean squared
error loss otherwise.}
}
\value{
A list with the default values for the SAGE arguments.
}
\description{
Get the Default Values for the SAGE Arguments
}
\author{
Martin Jullum
}
