% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_supported_approaches}
\alias{get_supported_approaches}
\title{Get the Implemented Approaches}
\usage{
get_supported_approaches()
}
\value{
Character vector.
The names of the implemented approaches that can be passed to argument \code{approach} in \code{\link[=explain]{explain()}}.
}
\description{
Get the Implemented Approaches
}
