% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_supported_models}
\alias{get_supported_models}
\title{Provide a Data.table with the Supported Models}
\usage{
get_supported_models()
}
\value{
A data.table with the supported models.
}
\description{
Provide a Data.table with the Supported Models
}
