% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_and_casual_Shapley.R
\name{get_valid_causal_coalitions}
\alias{get_valid_causal_coalitions}
\title{Get all coalitions satisfying the causal ordering}
\usage{
get_valid_causal_coalitions(
  causal_ordering,
  sort_features_in_coalitions = TRUE
)
}
\arguments{
\item{causal_ordering}{List of vectors containing the partial causal ordering.
The elements in the list represents the components in the causal ordering and can either
be a single feature index or several, that is, a vector. For example, we can have
\code{list(c(1,2), c(3, 4))}, which means that \verb{1,2 -> 3} and \verb{1,2 -> 4}, i.e., one and
two are the ancestors of three and four, but three and four are not related.}

\item{sort_features_in_coalitions}{Boolean. If \code{TRUE}, then the feature indices in the
coalitions are sorted in increasing order. If \code{FALSE}, then the function maintains the
order of features within each group given in \code{causal_ordering}.}
}
\value{
List of vectors containing all coalitions that respects the causal ordering.
}
\description{
Get all coalitions satisfying the causal ordering
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
