% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{get_validation_iwae}
\alias{get_validation_iwae}
\title{Compute the Importance Sampling Estimator (Validation Error)}
\usage{
get_validation_iwae(
  val_dataloader,
  mask_generator,
  batch_size,
  model,
  num_samples,
  verbose = FALSE
)
}
\arguments{
\item{val_dataloader}{A torch dataloader which loads the validation data.}

\item{mask_generator}{A mask generator object that generates the masks.}

\item{batch_size}{Integer. The number of samples to include in each batch.}

\item{model}{The vaeac model.}

\item{num_samples}{Number of samples to generate for computing the IWAE for each validation sample.}

\item{verbose}{Boolean. If we are to print results to the user.}
}
\value{
The average iwae over all instances in the validation dataset.
}
\description{
Compute the Importance Sampling Estimator which the vaeac model
uses to evaluate its performance on the validation data.
}
\details{
Compute mean IWAE log likelihood estimation of the validation set.
Takes validation data loader, mask generator, batch size, model (vaeac)
and number of IWAE latent samples per object.Returns one the estimation (float).
IWAE is an abbreviation for Importance Sampling Estimator
\deqn{\log p_{\theta, \psi}(x|y) \approx
\log {\frac{1}{S}\sum_{i=1}^S p_\theta(x|z_i, y) p_\psi(z_i|y) \big/ q_\phi(z_i|x,y),}}
where \eqn{z_i \sim q_\phi(z|x,y)}.
For more details, see \href{https://www.jmlr.org/papers/volume23/21-1413/21-1413.pdf}{Olsen et al. (2022)}.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
