% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{get_weighted_matrix}
\alias{get_weighted_matrix}
\title{Get weighted matrix}
\usage{
get_weighted_matrix(X, use_shapley_weights_in_W = T,
  normalize_W_weights = T)
}
\arguments{
\item{X}{data.table}
}
\value{
Matrix
}
\description{
Get weighted matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
