% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{get_weights}
\alias{get_weights}
\title{Calculate shapley weights}
\usage{
get_weights(X, m)
}
\arguments{
\item{X}{data.table}
}
\value{
data.table
}
\description{
Calculate shapley weights
}
\author{
Nikolai Sellereite
}
