% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{ggpairs_plot_imputed_and_true_data_shapr}
\alias{ggpairs_plot_imputed_and_true_data_shapr}
\title{Plot Pairwise Plots for Imputed and True Data}
\usage{
ggpairs_plot_imputed_and_true_data_shapr(
  explanation,
  which_vaeac_model = "best",
  true_data = NULL,
  return_figures = TRUE,
  save_figures = FALSE,
  show_figures = FALSE,
  upper_cont = c("cor", "points", "smooth", "smooth_loess", "density", "blank"),
  upper_cat = c("count", "cross", "ratio", "facetbar", "blank"),
  upper_mix = c("box", "box_no_facet", "dot", "dot_no_facet", "facethist",
    "facetdensity", "denstrip", "blank"),
  lower_cont = c("points", "smooth", "smooth_loess", "density", "cor", "blank"),
  lower_cat = c("facetbar", "ratio", "count", "cross", "blank"),
  lower_mix = c("facetdensity", "box", "box_no_facet", "dot", "dot_no_facet",
    "facethist", "denstrip", "blank"),
  diag_cont = c("densityDiag", "barDiag", "blankDiag"),
  diag_cat = c("barDiag", "blankDiag"),
  cor_method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{explanation}{List. The output list from the shapr::explanation function.}

\item{which_vaeac_model}{String. Indicating which VAEAC model to use when generating the samples.}

\item{true_data}{Matrix/data.frame containing the data from the distribution that the VAEAC model is fitted to.}

\item{return_figures}{Boolean. If we are to return the figures in a list.}

\item{save_figures}{Boolean. If we are to save the figures as png files.}

\item{show_figures}{Boolean. If we are to show the figures while making them.
If TRUE, then we create the figures twice, so the run time will double.}

\item{upper_cont}{Type of plot to use in upper triangle for continuous features.}

\item{upper_cat}{Type of plot to use in upper triangle for categorical features.}

\item{upper_mix}{Type of plot to use in upper triangle for mixed features.}

\item{lower_cont}{Type of plot to use in lower triangle for continuous features.}

\item{lower_cat}{Type of plot to use in lower triangle for categorical features.}

\item{lower_mix}{Type of plot to use in lower triangle for mixed features.}

\item{diag_cont}{Type of plot to use on the diagonal for continuous features.}

\item{diag_cat}{Type of plot to use on the diagonal for categorical features.}

\item{cor_method}{Type of correlation measure.}

\item{...}{Extra parameters sent to ggsave function.}
}
\value{
A list containing the figures if \code{return_figures} = TRUE.
}
\description{
A function that either takes the output of the train_VAEAC_model or continue_train_VAEAC_model function
and iterate over the VAEAC models. It generates imputations from the unconditioned distribution p(x) estimated by
the VAEAC models and then compares them with data from the true distribution (if provided).
}
\details{
See https://www.blopig.com/blog/2019/06/a-brief-introduction-to-ggpairs/ for introduction to ggpairs.
Each combination of variables are plotted according to whether they are:
continuous: e.g. "points" (lower default), "smooth", "smooth_loess", "density", "cor" (upper default), or "blank"
combo: e.g. "box", "box_no_facet" (upper default), "dot", "dot_no_facet", "facethist", "facetdensity", "denstrip", or "blank"
discrete: e.g. "ratio", "facetbar", or "blank"
Plots on the diagonal can either be:
continuous: "densityDiag", "barDiag", "blankDiag"
discrete: "barDiag", "blankDiag"
Also call vig_ggally("ggally_plots").
}
