% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{global_arguments}
\alias{global_arguments}
\title{Set of global arguments used in the package}
\usage{
global_arguments(m, n, xtrain, xtest, nsamples, features, exact, cov_mat,
  model, nrows, scale, w_threshold, n_threshold, verbose, w, s, D, I,
  cond_approach = "empirical", p_default,
  distance_metric = "Euclidean", kernel_metric = "Gaussian")
}
\arguments{
\item{m}{Integer. Total number of features}

\item{n}{Integer. Number of combinations}

\item{xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{xtest}{Matrix, data.frame or data.table with the features, whose predictions ought to
be explained (test data)}

\item{nsamples}{Integer. Number of samples}

\item{features}{List.}

\item{exact}{Logical. If TRUE, uses the full sum in the Shapley formula, if FALSE, uses a
sampling approach to approximate the sum}

\item{cov_mat}{Covariance matrix of training set}

\item{model}{Model object. Fitted model that is used to produce the predictions}

\item{scale}{Logical}

\item{verbose}{Integer. How much information to print during function execution (in development)}

\item{w}{Matrix}

\item{s}{Matrix}

\item{D}{Matrix}

\item{I}{Matrix}

\item{cond_approach}{String or list. When being a list, the elements in the list refers to the
rows in l$x that ought to be included in each of the approaches!}

\item{p_default}{Numeric}

\item{distance_metric}{String indicating which distance metric should be used in the empirical
conditional distribution. Defaults to "Euclidean", "Mahalanobis" and "Mahalanobis_scaled" being
the other options. "Mahalanobis_scaled" includes the 1/|S| factor in the paper is preferred for
a consistent \eqn{\sigma}.}

\item{kernel_metric}{String indicating which kernel metric should be used in the empirical
conditional distribution. Defaults to "Gaussian" [\eqn{\exp(-D/2\sigma)}], with "independence"
(imputing independently, ignoring any distance) being the second option "Gaussian_old"
[\eqn{\sqrt(\exp(-D/2\sigma))}] is also kept for reproducibility.}

\item{xtrain_mat}{Matrix with the features from the training data}

\item{xtest_mat}{Matrix with the features, whose predictions ought to be explained (test data)}

\item{no_samp}{Integer. How many samples to use when approximating the sum in the Shapley formula
(previously called \code{nrows})}

\item{shapley_weight_inf_replacement}{Numeric. Indicating which weight to use for the full
conditional and unconditional expectations in kernel SHAPs weighted least squares formulation.}

\item{reduce_dim}{Logical. Indicating whether to reduce the dimension of the weighted least
squares problem by merging identical columns and adjusting their weights.}

\item{l}{List. The output from the \code{prepare_kshap} function}

\item{no_samp_mc}{Positive integer. Indicating the maximum number of samples to use in the
Monte Carlo integration for every conditional expectation (previously called \code{n_threshold})}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution.
NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{sigma}{Numeric matrix. (Optional) Containing the covariance matrix of the data generating
distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{verbose}{Logical}

\item{cond_approach}{Either a string indicating which method should be used to estimate all
conditional expectations. Defaults to "empirical_fixed_sigma", with "empirical_AICc_full",
"empirical_AICc_each_k","Gaussian" and "copula" being other alternatives. One can also supply a
named list where the names are one or more of the implemented methods, and the named lists
contains one vector each, each containing the row numbers of the S-matrix computed using
\code{prepare_kshap} that whose corresponding conditional expectations should be computed with
that method. Any number not specified is computed with the default empirical method.}

\item{w_kernel}{Array. Contains all nonscaled weights between training and testing observations
for all combinations.}

\item{xtest_gauss_trans}{Vector with the Gaussian transformed test observations}

\item{mu}{Mean vector of training set}
}
\description{
Set of global arguments used in the package
}
\author{
Nikolai Sellereite
}
