% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forecast.R
\name{group_forecast_setup}
\alias{group_forecast_setup}
\title{Set up user provided groups for explanation in a forecast model.}
\usage{
group_forecast_setup(group, horizon_features)
}
\arguments{
\item{group}{The list of groups to be explained.}

\item{horizon_features}{A list of features per horizon, to split appropriate groups over.}
}
\value{
A list containing
\itemize{
\item group The list group with entries that differ per horizon split accordingly.
\item horizon_group A list of which groups are applicable per horizon.
}
}
\description{
Set up user provided groups for explanation in a forecast model.
}
\keyword{internal}
