% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupVariables.R
\name{group_variables}
\alias{group_variables}
\title{Group variables according to matrix of Kendall's tau}
\usage{
group_variables(X = NULL, corMat = NULL, alpha = 1)
}
\arguments{
\item{X}{optional data matrix, ignores corMat if included}

\item{corMat}{optional correlation matrix, ignores X if included. Either X or corMat must be included}

\item{alpha}{optional tuning parameter for optimal number of clusters}
}
\value{
List
}
\description{
Group variables according to matrix of Kendall's tau
}
\author{
Anders Løland
}
