% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{H.func.new}
\alias{H.func.new}
\title{Most general H-function used in the AICc-function, uses an X data.table where the first column is an .id column identifying how the diagonal should be in the H-matrix.
Need to improve this help file.}
\usage{
H.func.new(h.vec, X, kernel = "Euclidean", scale_var = T, S_scale_dist)
}
\arguments{
\item{h.vec}{Vector of bandwidths of size \code{q=ncol(X)}}
}
\value{
H-function output.
}
\description{
Most general H-function used in the AICc-function, uses an X data.table where the first column is an .id column identifying how the diagonal should be in the H-matrix.
Need to improve this help file.
}
\author{
Martin Jullum
}
