% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{H.func}
\alias{H.func}
\title{Internal  H-function used in the AICc-function. Need to improve this help file.}
\usage{
H.func(h.vec, X, kernel = "Euclidean", scale_var = T, S_scale_dist)
}
\arguments{
\item{h.vec}{Vector of bandwidths of size \code{q=ncol(X)}}
}
\value{
H-function output.
}
\description{
Internal  H-function used in the AICc-function. Need to improve this help file.
}
\author{
Martin Jullum
}
