% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{ifelse0}
\alias{ifelse0}
\title{Conditional Element Selection}
\usage{
ifelse0(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return values for true elements of test.}

\item{no}{return values for false elements of test.}
}
\description{
ifelse returns a value with the same shape as test which is filled with elements selected from either yes or no depending on whether the element of test is TRUE or FALSE.
}
\details{
Copy of base::ifelse, except that the there is no recycling attempt if the length of yes/no is 1.
}
\keyword{internal}
