% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_cpp}
\alias{impute_cpp}
\title{Get imputed data}
\usage{
impute_cpp(ID, Comb, Xtrain, Xtest, S)
}
\arguments{
\item{ID}{Positive integer vector}

\item{Comb}{Positive integer vector}

\item{Xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{Xtest}{Matrix, data.frame or data.table with the features, whose predictions ought to be explained (test data)}

\item{S}{Matrix}
}
\value{
Array of three dimensions
}
\description{
Get imputed data
}
\author{
Nikolai Sellereite
}
