% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{impute_data}
\alias{impute_data}
\title{Get predictions}
\usage{
impute_data(W_kernel, S, Xtrain, Xtest, w_threshold = 0.7,
  noSamp_MC = 1000)
}
\arguments{
\item{W_kernel}{Array containg all unscaled weights between training and testing observations for all combinations.}

\item{S}{Matrix}

\item{Xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{Xtest}{Matrix, data.frame or data.table with the features, whose predictions ought to be explained (test data)}

\item{noSamp_MC}{Postive integer. Indicating the maximum number of samples to use in the Monte Carlo integration for every conditional expectation (previously called n_threshold)}
}
\value{
List
}
\description{
Get predictions
}
\author{
Nikolai Sellereite
}
