% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{inv_gaussian_transform}
\alias{inv_gaussian_transform}
\title{Transforms new data to a standardized normal distribution}
\usage{
inv_gaussian_transform(zx, n_z, type = 7)
}
\arguments{
\item{zx}{Numeric vector. The first \code{n_z} items is the Gaussian data, and last part is
the data with the original transformation.}

\item{n_z}{Positive integer. Number of elements of \code{zx} that belongs to new data.}

\item{type}{Integer between \code{1} and \code{9}. See \code{\link[stats]{quantile}} for more
information.}
}
\value{
Numeric vector of length \code{n_z}
}
\description{
Transforms new data to a standardized normal distribution
}
\examples{
zx <- rnorm(50)
n_z <- 30
x <- shapr:::inv_gaussian_transform(zx, n_z)
str(x)

}
\author{
Martin Jullum
}
\keyword{internal}
