% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_gaussian_transform_cpp}
\alias{inv_gaussian_transform_cpp}
\title{Transforms new data to a standardized normal distribution}
\usage{
inv_gaussian_transform_cpp(z, x)
}
\arguments{
\item{z}{arma::mat.
The data are the Gaussian Monte Carlos samples to transform.}

\item{x}{arma::mat.
The data with the original transformation. Used to conduct the transformation of \code{z}.}
}
\value{
arma::mat of the same dimension as \code{z}
}
\description{
Transforms new data to a standardized normal distribution
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
