% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_copula.R
\name{inv_gaussian_transform_R}
\alias{inv_gaussian_transform_R}
\title{Transform data using the inverse Gaussian transformation.}
\usage{
inv_gaussian_transform_R(z, x)
}
\arguments{
\item{z}{Matrix. The data are the Gaussian Monte Carlos samples to transform.}

\item{x}{Matrix. The data with the original transformation. Used to conduct the transformation of \code{z}.}
}
\value{
Matrix of the same dimension as \code{z}.
}
\description{
Transform data using the inverse Gaussian transformation.
}
\details{
This function is called from \code{prepare_data_copula_cpp()} as the this was faster
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
