% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{k_mahalanobis}
\alias{k_mahalanobis}
\title{The kernel function using the the Mahalanobis distance used within the H-function of the AICc-function. Need to improve this help file.}
\usage{
k_mahalanobis(X, h.vec, Sigma, S_scale_dist)
}
\arguments{
\item{Sigma}{Numeric matrix. (Optional) Containing the covariance matrix of the data generating distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).}
}
\value{
K-function output.
}
\description{
The kernel function using the the Mahalanobis distance used within the H-function of the AICc-function. Need to improve this help file.
}
\author{
Martin Jullum
}
