% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{kl_normal_normal}
\alias{kl_normal_normal}
\title{Compute the KL divergence between Gaussian distributions.}
\usage{
kl_normal_normal(p, q)
}
\arguments{
\item{p}{a Gaussian distribution,}

\item{q}{a Gaussian distribution.}
}
\value{
The KL divergence
}
\description{
Computes the KL divergence between univariate normal distributions using the analytical formula,
see \url{https://en.wikipedia.org/wiki/Kullback\%E2\%80\%93Leibler_divergence#Multivariate_normal_distributions}.
}
