% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forecast.R
\name{lag_data}
\alias{lag_data}
\title{Lag a matrix of variables a specific number of lags for each variables.}
\usage{
lag_data(x, lags)
}
\arguments{
\item{x}{The matrix of variables (one variable per column).}

\item{lags}{A numeric vector denoting how many lags each variable should have.}
}
\value{
A list with two items
\itemize{
\item A matrix, lagged with the lagged data.
\item A list, group, with groupings of the lagged data per variable.
}
}
\description{
Lag a matrix of variables a specific number of lags for each variables.
}
\keyword{internal}
