% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{linters_config}
\alias{linters_config}
\title{Configuration of linting}
\usage{
linters_config()
}
\value{
List
}
\description{
This function is used when linting the package. We're modifying some of
the default settings. Note that the tests, i.e when running \code{devtools::test}
will fail if the package contains lints. At the moment \code{devtools::test} will not
report the different lints, so you'll need to run
\code{lintr::lint_package(linters = linters_config())} to get information of what lines
contains lints.
}
