% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{log_loss}
\alias{log_loss}
\title{Logistic (Cross-Entropy) Loss}
\usage{
log_loss(y, pred)
}
\arguments{
\item{y}{Numeric vector of true binary responses (values in 0/1).}

\item{pred}{Numeric vector (or scalar) of predicted probabilities.}
}
\value{
The mean logistic loss as a single numeric value.
}
\description{
Logistic (Cross-Entropy) Loss
}
\author{
Martin Jullum
}
\keyword{internal}
