% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{make_dummies}
\alias{make_dummies}
\title{Initiate the making of dummy variables}
\usage{
make_dummies(data, newdata)
}
\arguments{
\item{data}{data.table or data.frame. Includes all the features (both factors and possibly others).}

\item{newdata}{data.table or data.frame. New data (features) that has the same
features as \code{data}.}
}
\value{
A list that contains the following entries:
\describe{
\item{obj}{List, Contains \describe{
\item{features}{Vector. Contains the names of all the features in \code{data}.}
\item{factor_features}{Vector. Contains the names of all the factors in \code{data}.}
\item{factor_list}{List. Contains each factor and its vector of levels.}
\item{contrasts_list}{List. Contains all the contrasts of the factors.}
}}
\item{model.matrix}{A data.frame containing all of the factors in \code{new_data} as
one-hot encoded variables.}
}
}
\description{
Initiate the making of dummy variables
}
\examples{

data("Boston", package = "MASS")
x_var <- c("lstat", "rm", "dis", "indus")
y_var <- "medv"
x_train <- as.data.frame(Boston[401:411, x_var])
y_train <- Boston[401:408, y_var]
x_test <- as.data.frame(Boston[1:4, x_var])

# convert to factors for illustational purpose
x_train$rm <- factor(round(x_train$rm))
x_test$rm <- factor(round(x_test$rm), levels = levels(x_train$rm))

dummylist <- make_dummies(data = rbind(x_train, x_test), newdata = x_test)

}
\author{
Annabelle Redelmeier
}
