% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_type}
\alias{model_type}
\alias{model_type.default}
\alias{model_type.lm}
\alias{model_type.glm}
\alias{model_type.ranger}
\alias{model_type.mgcv}
\title{TODO: Add title & description}
\usage{
model_type(x)

\method{model_type}{default}(x)

\method{model_type}{lm}(x)

\method{model_type}{glm}(x)

\method{model_type}{ranger}(x)

\method{model_type}{mgcv}(x)
}
\description{
TODO: Add title & description
}
