% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{mse_loss}
\alias{mse_loss}
\title{Mean Squared Error Loss}
\usage{
mse_loss(y, pred)
}
\arguments{
\item{y}{Numeric vector of true responses.}

\item{pred}{Numeric vector (or scalar) of predictions.}
}
\value{
The mean squared error as a single numeric value.
}
\description{
Mean Squared Error Loss
}
\author{
Martin Jullum
}
\keyword{internal}
