% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{normal_parse_params}
\alias{normal_parse_params}
\title{Creates Normal Distributions}
\usage{
normal_parse_params(params, min_sigma = 0.001)
}
\arguments{
\item{params}{Tensor containing the parameters for the normal distributions.}

\item{min_sigma}{The minimal variance allowed.}
}
\value{
torch::distr_noramal distributions with the provided means and standard deviations.
}
\description{
Function that takes in the a tensor where the first half of the columns contains the means of the
normal distributions, while the latter half of the columns contains the standard deviations. The standard deviations
are clamped with 'min_sigma' to ensure stable results. If 'params' is of dimensions batch_size x 8, the function will
create 4 independent normal distributions for each of the observation ('batch_size' observations in total).
}
\details{
Take a Tensor (e.g. neural network output) and return torch::distr_normal distribution. This Normal distribution
is component-wise independent, and its dimensionality depends on the input shape. First half of channels is mean of the
distribution, the softplus of the second half is std (sigma), so there is no restrictions on the input tensor.
\code{min_sigma} is the minimal value of sigma. I.e., if the above softplus is less than min_sigma, then sigma is clipped
from below with value \code{min_sigma}. This regularization is required for the numerical stability and may be considered
as a neural network architecture choice without any change to the probabilistic model.
}
