% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{num_str}
\alias{num_str}
\title{Convert a character to a numeric class}
\usage{
num_str(x)
}
\arguments{
\item{Character.}{A character that represents a number, or a vector of characters.}
}
\value{
A numeric class object with the value of the string.
}
\description{
To be used in cli calls like \verb{cli_text(\{.val \{num_str("12.10")\}\})} to format a character strings
that typically represent a number like it was a number. May also be used with strings not representing a number.
}
\keyword{internal}
