% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observation_impute}
\alias{observation_impute}
\title{Generate permutations of training data using test observations}
\usage{
observation_impute(W_kernel, S, x_train, x_test, w_threshold = 0.7,
  n_samples = 1000)
}
\arguments{
\item{W_kernel}{Numeric matrix. Contains all nonscaled weights between training and test
observations for all feature combinations. The dimension equals \code{n_train x n_features}.}

\item{S}{Integer matrix of dimension \code{n_combinations x n_features}, where \code{n_combinations}
and \code{n_features} equals the total number of sampled/non-sampled feature combinations and
the total number of unique features, respectively. Note that \code{n_features = ncol(x_train)}.}

\item{x_train}{Numeric matrix}

\item{x_test}{Numeric matrix}

\item{w_threshold}{Numeric vector of length 1. Represents the number of samples to use is chosen to be
the number of the largest weights needed to account for this fraction of the total weight.}
}
\value{
data.table
}
\description{
TODO: Add a description text
}
\examples{
# TODO: Add simple example

}
\author{
Nikolai Sellereite
}
\keyword{internal}
