% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{observation_weights}
\alias{observation_weights}
\title{Calculate Shapley weights}
\usage{
observation_weights(X, m)
}
\arguments{
\item{X}{data.table}
}
\value{
data.table
}
\description{
Calculate Shapley weights
}
\author{
Nikolai Sellereite
}
