% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.shapr}
\alias{plot.shapr}
\title{Plot of the Shapley value explanations}
\usage{
\method{plot}{shapr}(x, digits = 3, plot_phi0 = TRUE,
  index_x_test = NULL, top_k_features = NULL, ...)
}
\arguments{
\item{x}{An \code{shapr} object. See \code{\link{explain}}.}

\item{digits}{Integer. Number of significant digits to use in the feature description}

\item{plot_phi0}{Logical. Whether to include phi0 in the plot}

\item{index_x_test}{Integer vector. Which of the test observations to plot}

\item{top_k_features}{Integer. How many features to include in the plot}

\item{...}{Currently not used.}
}
\value{
ggplot object with plots of the Shapley value explanations
}
\description{
Plots the individual prediction explanations. Uses facet_wrap of ggplot
}
\author{
Martin Jullum
}
