% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kshap}
\alias{plot_kshap}
\title{Plot of the Shapley value explanations}
\usage{
plot_kshap(explanation, Xtest, model, no_desc_digits = 3,
  plot_phi0 = T, plot_which_Xtest = 1:nrow(Xtest),
  top_k_features = ncol(Xtest) + 1)
}
\arguments{
\item{explanation}{The output from compute_kshap}

\item{Xtest}{Data.table. Contains the covariates of the test data}

\item{model}{Model object. Fitted model that is used to produce the predictions}

\item{no_desc_digits}{Integer. Number of significant digits to use in the feature description}

\item{plot_phi0}{Logical. Whether to include phi0 in the plot}

\item{plot_which_Xtest}{Integer vector. Which of the test observations to plot}

\item{top_k_features}{Integer. How many features to include in the plot}
}
\value{
ggplot object with plots of the Shapley value explanations
}
\description{
Plots the individual prediction explanations. Uses facet_wrap of ggplot
}
\author{
Martin Jullum
}
