% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{plot_vaeac_imputed_ggpairs}
\alias{plot_vaeac_imputed_ggpairs}
\title{Plot Pairwise Plots for Imputed and True Data}
\usage{
plot_vaeac_imputed_ggpairs(
  explanation,
  which_vaeac_model = "best",
  true_data = NULL,
  return_figures = TRUE,
  save_figures = FALSE,
  show_figures = FALSE,
  upper_cont = c("cor", "points", "smooth", "smooth_loess", "density", "blank"),
  upper_cat = c("count", "cross", "ratio", "facetbar", "blank"),
  upper_mix = c("box", "box_no_facet", "dot", "dot_no_facet", "facethist",
    "facetdensity", "denstrip", "blank"),
  lower_cont = c("points", "smooth", "smooth_loess", "density", "cor", "blank"),
  lower_cat = c("facetbar", "ratio", "count", "cross", "blank"),
  lower_mix = c("facetdensity", "box", "box_no_facet", "dot", "dot_no_facet",
    "facethist", "denstrip", "blank"),
  diag_cont = c("densityDiag", "barDiag", "blankDiag"),
  diag_cat = c("barDiag", "blankDiag"),
  cor_method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{explanation}{List. The output list from the \code{\link[=explain]{explain()}} function.}

\item{which_vaeac_model}{String. Indicating which vaeac model to use when generating the samples.}

\item{true_data}{Matrix/data.frame containing the data from the distribution that the vaeac model is fitted to.}

\item{return_figures}{Boolean. If we are to return the figures in a list.}

\item{save_figures}{Boolean. If we are to save the figures as png files.}

\item{show_figures}{Boolean. If we are to show the figures while making them.
If TRUE, then we create the figures twice, so the run time will double.}

\item{upper_cont}{Type of plot to use in upper triangle for continuous features.}

\item{upper_cat}{Type of plot to use in upper triangle for categorical features.}

\item{upper_mix}{Type of plot to use in upper triangle for mixed features.}

\item{lower_cont}{Type of plot to use in lower triangle for continuous features.}

\item{lower_cat}{Type of plot to use in lower triangle for categorical features.}

\item{lower_mix}{Type of plot to use in lower triangle for mixed features.}

\item{diag_cont}{Type of plot to use on the diagonal for continuous features.}

\item{diag_cat}{Type of plot to use on the diagonal for categorical features.}

\item{cor_method}{Type of correlation measure.}

\item{...}{Extra parameters sent to the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} function.}
}
\value{
A list containing the figures if \code{return_figures} = \code{TRUE}.
}
\description{
A function that creates a matrix of plots (\code{\link[GGally:ggpairs]{GGally::ggpairs()}}) from
generated imputations from the unconditioned distribution \eqn{p(\boldsymbol{x})} estimated by
a vaeac model, and then compares the imputed values with data from the true distribution (if provided).
}
\details{
See \href{https://www.blopig.com/blog/2019/06/a-brief-introduction-to-ggpairs/}{ggpairs} for an
introduction to \code{\link[GGally:ggpairs]{GGally::ggpairs()}}, and the corresponding
\href{https://ggobi.github.io/ggally/articles/ggally_plots.html}{vignette}.

Each combination of variables are plotted according to whether they are:
\describe{
\item{continuous}{exactly one of ('points' (lower default), 'smooth', 'smooth_loess',
'density', 'cor' (upper default), or 'blank').}
\item{combo}{exactly one of ('box', 'box_no_facet' (upper default), 'dot', 'dot_no_facet',
'facethist', 'facetdensity', 'denstrip', or 'blank').}
\item{discrete}{exactly one of ('ratio', 'facetbar', or 'blank')}
}

Plots on the diagonal can either be:
\describe{
\item{continuous}{exactly one of ('densityDiag', 'barDiag', 'blankDiag').
This option is used for continuous X data.}
\item{discrete}{exactly one of ('barDiag', 'blankDiag'). This option is used for categorical X and Y data.}
\item{na}{exactly one of ('naDiag', 'blankDiag').  This option is used when all X data is \code{NA}.}
}
}
\author{
Lars Henry Berge Olsen
}
