% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{pred_vector}
\alias{pred_vector}
\title{Predict on vector form}
\usage{
pred_vector(model, data)
}
\arguments{
\item{model}{Model object. Fitted model that is used to produce the predictions}

\item{data}{data.table or data.frame with data to perform prediction}
}
\value{
Vector of predictions
}
\description{
Performs prediction of response for model classes lm, glm, ranger and xgboost with binary or continuous response.
Outpus the prediction on vector form. May let the user provide this function to handle any prediction model in the future.
}
\author{
Martin Jullum
}
