% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_model}
\alias{predict_model}
\alias{predict_model.default}
\alias{predict_model.lm}
\alias{predict_model.glm}
\alias{predict_model.ranger}
\alias{predict_model.xgb.Booster}
\alias{predict_model.mgcv}
\title{Predict on vector form}
\usage{
predict_model(x, newdata)

\method{predict_model}{default}(x, newdata)

\method{predict_model}{lm}(x, newdata)

\method{predict_model}{glm}(x, newdata)

\method{predict_model}{ranger}(x, newdata)

\method{predict_model}{xgb.Booster}(x, newdata)

\method{predict_model}{mgcv}(x, newdata)
}
\arguments{
\item{x}{Object of class inheriting from one of the supported models. See details.}

\item{newdata}{A data frame (or matrix) in which to look for variables with which to predict.}
}
\value{
Numeric
}
\description{
Performs prediction of response \code{\link[stats]{lm}}, \code{\link[stats]{glm}},
\code{\link[ranger]{ranger}} and \code{\link[xgboost]{xgboost}} with binary or continuous response.
}
\details{
The following models are currently supported:
\itemize{
\item \code{\link[stats]{lm}}
\item \code{\link[stats]{glm}}
\item \code{\link[ranger]{ranger}}
\item \code{\link[mgcv]{mgcv}}
\item \code{\link[xgboost]{xgboost}}
}
}
\author{
Martin Jullum
}
