% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Prepare and predict in batches}
\usage{
prepare_and_predict(explainer, n_batches, ...)
}
\arguments{
\item{explainer}{An \code{explainer} object to use for explaining the observations.
See \code{\link{shapr}}.}

\item{n_batches}{Positive integer. Specifies how many iterations that should be used to calculate the
contribution function for each test observation. The default value is 1. Setting it to a higher number can
eleviate memory problems for models with many features.}

\item{...}{Additional arguments passed to \code{\link{prepare_data}}}
}
\value{
A list. See \code{\link{explain}} for more information.
}
\description{
Prepare and predict in batches
}
\keyword{internal}
