% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_categorical.R
\name{prepare_data_single_coalition}
\alias{prepare_data_single_coalition}
\title{Compute the conditional probabilities for a single coalition for the categorical approach}
\usage{
prepare_data_single_coalition(internal, index_features)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}
}
\description{
The \code{\link[=prepare_data.categorical]{prepare_data.categorical()}} function is slow when evaluated for a single coalition.
This is a bottleneck for Causal Shapley values which call said function a lot with single coalitions.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
