% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{prepare_kernelShap}
\alias{prepare_kernelShap}
\title{Get shapley weights for test data}
\usage{
prepare_kernelShap(Xtrain, Xtest, exact = TRUE, noSamp = NULL,
  shapley_weight_inf_replacement = 10^6,
  compute_distances_for_no_var = 0:ncol(Xtrain))
}
\arguments{
\item{Xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{Xtest}{Matrix, data.frame or data.table with the features, whose predictions ought to be explained (test data)}

\item{exact}{Logical. If TRUE, uses the full sum in the Shapley formula, if FALSE, uses a sampling approach to approximate the sum}

\item{noSamp}{Integer. How many samples to use when approximating the sum in the Shapley formula
(previously called nrows)}

\item{shapley_weight_inf_replacement}{Numeric. Indicating which weight to use for the full conditional and unconditional expectations in kernel SHAPs WLS formulation.}
}
\value{
Matrix
}
\description{
Get shapley weights for test data
}
\author{
Nikolai Sellereite
}
