% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{prepare_kshap}
\alias{prepare_kshap}
\title{DELETE}
\usage{
prepare_kshap(Xtrain, Xtest, exact = TRUE, noSamp = NULL,
  shapley_weight_inf_replacement = 10^6,
  compute_distances_for_no_var = 0:ncol(Xtrain))
}
\description{
DELETE
}
\keyword{interal}
