% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_next_iteration.R
\name{prepare_next_iteration}
\alias{prepare_next_iteration}
\title{Prepares the next iteration of the iterative sampling algorithm}
\usage{
prepare_next_iteration(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
The (updated) internal list
}
\description{
Prepares the next iteration of the iterative sampling algorithm
}
\keyword{internal}
