% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Process (check and update) data according to specified feature list}
\usage{
preprocess_data(x, feature_list)
}
\arguments{
\item{x}{matrix, data.frame or data.table. The data to check input for and update
according to the specification in \code{feature_list}.}

\item{feature_list}{List. Output from running \code{get_data_specs} or \code{get_model_specs}}
}
\value{
Checked and updated data \code{x} in data.table format.
}
\description{
Process (check and update) data according to specified feature list
}
\keyword{internal}
