% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.shapr}
\alias{print.shapr}
\title{Print method for shapr objects}
\usage{
\method{print}{shapr}(x, digits = 4, ...)
}
\arguments{
\item{x}{A shapr object}

\item{digits}{Scalar Integer.
Number of digits to display to the console}
}
\description{
Print method for shapr objects
}
