% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.shapr}
\alias{print.summary.shapr}
\title{Print Method for summary.shapr Objects}
\usage{
\method{print}{summary.shapr}(x, ...)
}
\arguments{
\item{x}{A summary.shapr object.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns the summary object.
}
\description{
Print Method for summary.shapr Objects
}
