% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_iter.R
\name{print_iter}
\alias{print_iter}
\title{Prints iterative information}
\usage{
print_iter(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\description{
Prints iterative information
}
\keyword{internal}
