% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{process_groups}
\alias{process_groups}
\title{Process (check and update names) the group list}
\usage{
process_groups(group, feature_labels)
}
\arguments{
\item{group}{List. If \code{NULL} regular feature wise Shapley values are computed.
If provided, group wise Shapley values are computed. \code{group} then has length equal to
the number of groups. The list element contains character vectors with the features included
in each of the different groups.}

\item{feature_labels}{Vector of characters. Contains the feature labels used by the model}
}
\value{
List with two named elements: \code{group}: The input, but with group names if non-existing,
\code{group_num} a corresponding group list with names replaced by feature number
}
\description{
Process (check and update names) the group list
}
\details{
This function takes care of all preprocessing and checking of the provided data in \code{x} against
the feature_list which is typically the output from \code{\link[shapr:get_model_specs]{get_model_specs}}
}
\author{
Martin Jullum
}
\keyword{internal}
