% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_copula.R
\name{quantile.type7}
\alias{quantile.type7}
\title{Compute the quantiles using quantile type seven}
\usage{
\method{quantile}{type7}(x, probs)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted.}

\item{probs}{numeric vector of probabilities with values between zero and one.}
}
\value{
A vector of length length(\code{probs}) is returned.
}
\description{
Compute the quantiles using quantile type seven
}
\details{
Using quantile type number 7 from stats::quantile.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
