% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_type7_cpp}
\alias{quantile_type7_cpp}
\title{Compute the quantiles using quantile type seven}
\usage{
quantile_type7_cpp(x, probs)
}
\arguments{
\item{x}{arma::vec.
Numeric vector whose sample quantiles are wanted.}

\item{probs}{arma::vec.
Numeric vector of probabilities with values between zero and one.}
}
\value{
A vector of length \code{length(probs)} with the quantiles is returned.
}
\description{
Compute the quantiles using quantile type seven
}
\details{
Using quantile type number seven from stats::quantile in R.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
