% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression.check_namespaces}
\alias{regression.check_namespaces}
\title{Check that needed libraries are installed}
\usage{
regression.check_namespaces()
}
\description{
This function checks that the \code{parsnip}, \code{recipes}, \code{workflows}, \code{tune}, \code{dials},
\code{yardstick}, \code{hardhat} and \code{rsample}, packages are available.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
