% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression.check_vfold_cv_para}
\alias{regression.check_vfold_cv_para}
\title{Check the parameters that are sent to \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}}
\usage{
regression.check_vfold_cv_para(regression.vfold_cv_para)
}
\arguments{
\item{regression.vfold_cv_para}{Either \code{NULL} (default) or a named list containing
the parameters to be sent to \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}. See the regression vignette for
several examples.}
}
\description{
Check that \code{regression.vfold_cv_para} is either NULL or a named list that only contains recognized parameters.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
