% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression.cv_message}
\alias{regression.cv_message}
\title{Produce message about which batch prepare_data is working on}
\usage{
regression.cv_message(
  regression.results,
  regression.grid,
  n_cv = 10,
  current_comb
)
}
\arguments{
\item{regression.results}{The results of the CV procedures.}

\item{regression.grid}{Object containing the hyperparameter values.}

\item{n_cv}{Integer (default is 10) specifying the number of CV hyperparameter configurations to print.}

\item{current_comb}{Integer vector. The current combination of features, passed to verbosity printing function.}
}
\description{
Produce message about which batch prepare_data is working on
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
