% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression.get_string_to_R}
\alias{regression.get_string_to_R}
\title{Convert the string into an R object}
\usage{
regression.get_string_to_R(string)
}
\arguments{
\item{string}{A character vector/string containing the text to convert into R code.}
}
\description{
Convert the string into an R object
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
