% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression.get_y_hat}
\alias{regression.get_y_hat}
\title{Get the predicted responses}
\usage{
regression.get_y_hat(internal, model, predict_model)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}

\item{model}{Objects.
The model object that ought to be explained.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{predict_model}{Function.
The prediction function used when \code{model} is not natively supported.
See the documentation of \code{\link[=explain]{explain()}} for details.}
}
\value{
The same \code{internal} list, but added vectors \code{internal$data$x_train_y_hat} and
\code{internal$data$x_explain_y_hat} containing the predicted response of the training and explain data.
}
\description{
Get the predicted responses
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
