% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression.prep_message_batch}
\alias{regression.prep_message_batch}
\title{Produce message about which batch prepare_data is working on}
\usage{
regression.prep_message_batch(internal, index_features)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{iter_results}, \code{output}.}

\item{index_features}{Positive integer vector. Specifies the id_coalition to
apply to the present method. \code{NULL} means all coalitions. Only used internally.}
}
\description{
Produce message about which batch prepare_data is working on
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
