% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression_sep_time_message}
\alias{regression_sep_time_message}
\title{Produce time message for separate regression}
\usage{
regression_sep_time_message()
}
\description{
Produce time message for separate regression
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
