% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_surrogate.R
\name{regression_surr_augment}
\alias{regression_surr_augment}
\title{Augment the training data and the explicands}
\usage{
regression_surr_augment(
  internal,
  x,
  y_hat = NULL,
  index_features = NULL,
  augment_masks_as_factor = FALSE,
  augment_include_grand = FALSE,
  augment_add_id_comb = FALSE,
  augment_comb_prob = NULL,
  augment_weights = NULL
)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}

\item{x}{Data.table containing the data. Either the training data or the explicands. If \code{x} is the explicands,
then \code{index_features} must be provided.}

\item{y_hat}{Vector of numerics (optional) containing the predicted responses for the observations in \code{x}.}

\item{index_features}{Array of integers (optional) containing which coalitions to consider. Must be provided if
\code{x} is the explicands.}

\item{augment_masks_as_factor}{Logical (default is \code{FALSE}). If \code{TRUE}, then the binary masks are converted
to factors. If \code{FALSE}, then the binary masks are numerics.}

\item{augment_include_grand}{Logical (default is \code{FALSE}). If \code{TRUE}, then the grand coalition is included.
If \code{index_features} are provided, then \code{augment_include_grand} has no effect.}

\item{augment_add_id_comb}{Logical (default is \code{FALSE}). If \code{TRUE}, an additional column is adding containing
which coalition was applied.}

\item{augment_comb_prob}{Array of numerics (default is \code{NULL}). The length of the array must match the number of
combinations being considered, where each entry specifies the probability of sampling the corresponding coalition.
This is useful if we want to generate more training data for some specific coalitions.}

\item{augment_weights}{String (optional). Specifying which type of weights to add to the observations.
If \code{NULL} (default), then no weights are added. If \code{"Shapley"}, then the Shapley weights for the different
combinations are added to corresponding observations where the coalitions was applied. If \code{uniform}, then
all observations get an equal weight of one.}
}
\value{
A data.table containing the augmented data.
}
\description{
Augment the training data and the explicands
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
