% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{regression_train}
\alias{regression_train}
\title{Train a \code{\link[=tidymodels]{tidymodels()}} model}
\usage{
regression_train(
  x,
  regression_model = parsnip::linear_reg(),
  regression_tune = FALSE,
  regression_tune_values = NULL,
  regression_vfold_cv_para = NULL,
  regression_recipe_func = NULL,
  regression_response_var = "y_hat",
  verbose = 0
)
}
\arguments{
\item{x}{Data.table containing the data. Either the training data or the explicands. If \code{x} is the explicands,
then \code{index_features} must be provided.}

\item{regression_model}{A tidymodels object of class \code{model_specs}. Default is a linear regression model
(\code{\link[parsnip:linear_reg]{parsnip::linear_reg()}}). See \href{https://www.tidymodels.org/find/parsnip/}{tidymodels} for all possible models.}

\item{regression_tune}{Logical (default is \code{FALSE}). If \code{TRUE}, then we are to tune the hyperparemeters based on
the values provided in \code{regression_tune_values}. Note that no checks are conducted as this is checked earlier in
\code{setup_approach.regression_separate} and \code{setup_approach.regression_surrogate}.}

\item{regression_tune_values}{Either NULL (default) or a data.frame/data.table/tibble containing the possible
hyper-parameter value combinations to try. The column names must match the names of the tuneable parameters specified
in \code{regression_model}.}

\item{regression_vfold_cv_para}{Either NULL (default) or a list containing
the parameters to be sent to \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}.}

\item{regression_recipe_func}{Either NULL (default) or a function that must return
the RHS of the formula for arbitrary feature name inputs.}

\item{regression_response_var}{String (default is \code{y_hat}) containing the name of the response variable.}

\item{verbose}{An integer specifying the level of verbosity. If \code{0}, \code{shapr} will stay silent.
If \code{1}, it will print information about performance. If \code{2}, some additional information will be printed out.
Use \code{0} (default) for no verbosity, \code{1} for low verbose, and \code{2} for high verbose.
TODO: Make this clearer when we end up fixing this and if they should force a progressr bar.}
}
\value{
A trained \code{\link[=tidymodels]{tidymodels()}} model based on the provided input parameters.
}
\description{
Function that trains a \code{\link[=tidymodels]{tidymodels()}} model based on the provided input parameters.
This function allows for cross validating the hyperparameters of the model.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
