% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{respects_order}
\alias{respects_order}
\title{Helper function that checks a conditioning index against a particular causal ordering.}
\usage{
respects_order(index, causal_ordering)
}
\arguments{
\item{index}{Integer conditioning index to check against the causal ordering.}

\item{causal_ordering}{List of vectors specifying (partial) causal ordering. Each element in
the list is a component in the order, which can contain one or more variable indices in a vector.
For example, in list(1, c(2, 3)), 2 > 1 and 3 > 1, but 2 and 3 are not comparable.}
}
\description{
Helper function that checks a conditioning index against a particular causal ordering.
}
\author{
Tom Heskes, Ioan Gabriel Bucur
}
\keyword{internal}
