% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{round2}
\alias{round2}
\title{Round numbers to the specified number of decimal places}
\usage{
round2(x, decimal_places = 0)
}
\arguments{
\item{x}{Numeric vector. The numbers to round.}

\item{decimal_places}{Integer. The number of decimal places to round to. Defaults to 0.}
}
\value{
Numeric vector. The rounded numbers.
}
\description{
Round numbers to the specified number of decimal places
}
\keyword{internal}
