% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{round_half_even}
\alias{round_half_even}
\title{Round numbers to the nearest even number at the specified number of digits}
\usage{
round_half_even(x, digits = 2, tol = 10 * .Machine$double.eps)
}
\arguments{
\item{x}{Numeric vector. The numbers to round.}

\item{digits}{Integer.
Number of significant digits to display.
Defaults to 2.}

\item{tol}{Numeric. The tolerance for rounding. Defaults to 10 * .Machine$double.eps.}
}
\value{
Numeric vector. The rounded numbers.
}
\description{
Round numbers to the nearest even number at the specified number of digits
}
\keyword{internal}
