% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_cpp}
\alias{rss_cpp}
\title{sigma_hat_sq-function}
\usage{
rss_cpp(H, y)
}
\arguments{
\item{H}{matrix being the output from H_cpp}

\item{y}{vector with the "response variable"}

\item{ret_log}{logical indicating whether to return the logarithm of the sigma_sq}
}
\value{
Scalar
}
\description{
sigma_hat_sq-function
}
\author{
Martin Jullum
}
