% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_setup.R
\name{sample_coalition_table}
\alias{sample_coalition_table}
\title{Get table with sampled coalitions}
\usage{
sample_coalition_table(
  m,
  n_coalitions = 200,
  weight_zero_m = 10^6,
  paired_shap_sampling = TRUE,
  prev_coal_samples = NULL,
  kernelSHAP_reweighting,
  valid_causal_coalitions = NULL,
  dt_valid_causal_coalitions = NULL
)
}
\arguments{
\item{m}{Positive integer.
Total number of features/groups.}

\item{n_coalitions}{Positive integer.
Note that if \code{exact = TRUE}, \code{n_coalitions} is ignored.}

\item{weight_zero_m}{Numeric.
The value to use as a replacement for infinite coalition weights when doing numerical operations.}

\item{paired_shap_sampling}{Logical.
If \code{TRUE} (default), paired versions of all sampled coalitions are also included in the computation.
That is, if there are 5 features and e.g. coalitions (1,3,5) are sampled, then also coalition (2,4) is used for
computing the Shapley values. This is done to reduce the variance of the Shapley value estimates.}

\item{prev_coal_samples}{List.
A list of previously sampled coalitions.}

\item{kernelSHAP_reweighting}{String.
How to reweight the sampling frequency weights in the kernelSHAP solution after sampling.
The aim of this is to reduce the randomness and thereby the variance of the Shapley value estimates.
The options are one of \code{'none'}, \code{'on_N'}, \code{'on_all'}, \code{'on_all_cond'} (default).
\code{'none'} means no reweighting, i.e. the sampling frequency weights are used as is.
\code{'on_coal_size'} means the sampling frequencies are averaged over all coalitions of the same size.
\code{'on_N'} means the sampling frequencies are averaged over all coalitions with the same original sampling
probabilities.
\code{'on_all'} means the original sampling probabilities are used for all coalitions.
\code{'on_all_cond'} means the original sampling probabilities are used for all coalitions, while adjusting for the
probability that they are sampled at least once.
\code{'on_all_cond'} is preferred as it performs the best in simulation studies, see Olsen & Jullum (2024).}

\item{dt_valid_causal_coalitions}{data.table. Only applicable for asymmetric Shapley
values explanations, and is \code{NULL} for symmetric Shapley values.
The data.table contains information about the coalitions that respects the causal ordering.}
}
\description{
Get table with sampled coalitions
}
\keyword{internal}
