% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_setup.R
\name{sample_coalition_table_determ}
\alias{sample_coalition_table_determ}
\title{Get table with sampled coalitions using the semi-deterministic sampling approach}
\usage{
sample_coalition_table_determ(
  m,
  n_coalitions = 200,
  weight_zero_m = 10^6,
  prev_X,
  kernelSHAP_reweighting,
  dt_coal_determ_info,
  n_samps_scale = 10
)
}
\arguments{
\item{m}{Positive integer.
Total number of features/groups.}

\item{n_coalitions}{Positive integer.
Note that if \code{exact = TRUE}, \code{n_coalitions} is ignored.}

\item{weight_zero_m}{Numeric.
The value to use as a replacement for infinite coalition weights when doing numerical operations.}

\item{prev_X}{data.table. The X data.table from the previous iteration.
Only used for semi-deterministic sampling if \code{semi_deterministic_sampling == TRUE}.}

\item{kernelSHAP_reweighting}{String.
How to reweight the sampling frequency weights in the kernelSHAP solution after sampling.
The aim of this is to reduce the randomness and thereby the variance of the Shapley value estimates.
The options are one of \code{'none'}, \code{'on_N'}, \code{'on_all'}, \code{'on_all_cond'} (default).
\code{'none'} means no reweighting, i.e. the sampling frequency weights are used as is.
\code{'on_N'} means the sampling frequencies are averaged over all coalitions with the same original sampling
probabilities.
\code{'on_all'} means the original sampling probabilities are used for all coalitions.
\code{'on_all_cond'} means the original sampling probabilities are used for all coalitions, while adjusting for the
probability that they are sampled at least once.
\code{'on_all_cond'} is preferred as it performs the best in simulation studies, see
\href{https://arxiv.org/pdf/2410.04883}{Olsen & Jullum (2024)}.}

\item{dt_coal_determ_info}{data.table. Only applicable for semi-deterministic sampling, and is \code{NULL} for
the other sampling strategies. The data.table contains information about the which coalitions should be
deterministically included and which can be sampled.}

\item{n_samps_scale}{Positive integer.
Integer that scales the number of coalitions \code{n_coalitions} to sample as sampling is cheap,
while checking for \code{n_coalitions} unique coalitions is expensive, thus we over sample the
number of coalitions by a factor of \code{n_samps_scale} and determine when we have \code{n_coalitions} unique
coalitions and only use the coalitions up to this point and throw away the remaining coalitions.}
}
\description{
Get table with sampled coalitions using the semi-deterministic sampling approach
}
\keyword{internal}
