% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_coalitions_cpp_str_paired}
\alias{sample_coalitions_cpp_str_paired}
\title{We here return a vector of strings/characters, i.e., a CharacterVector,
where each string is a space-separated list of integers.}
\usage{
sample_coalitions_cpp_str_paired(m, n_coalitions, paired_shap_sampling = TRUE)
}
\arguments{
\item{m}{Positive integer.
Total number of features/groups.}

\item{n_coalitions}{IntegerVector.
The number of features to sample for each feature combination.}

\item{paired_shap_sampling}{Logical.
Whether to do paired sampling of coalitions.}
}
\description{
We here return a vector of strings/characters, i.e., a CharacterVector,
where each string is a space-separated list of integers.
}
\keyword{internal}
