% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_ctree}
\alias{sample_ctree}
\title{Sample ctree variables}
\usage{
sample_ctree(tree, n_samples, x_test, x_train, features, p)
}
\arguments{
\item{features}{List.}

\item{p}{}
}
\value{
data.table with \code{n_samples} (conditional) Gaussian samples
}
\description{
Sample ctree variables
}
\examples{
# TODO: Add simple example

}
\author{
Annabelle Redelmeier
}
\keyword{internal}
