% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_features_cpp_str_paired}
\alias{sample_features_cpp_str_paired}
\title{We here return a vector of strings/characters, i.e., a CharacterVector,
where each string is a space-separated list of integers.}
\usage{
sample_features_cpp_str_paired(m, n_features, paired_shap_sampling = TRUE)
}
\arguments{
\item{m}{Integer The number of elements to sample from, i.e., the number of features.}

\item{n_features}{IntegerVector The number of features to sample for each feature combination.}

\item{paired_shap_sampling}{Logical Should we return both the sampled coalition S and its complement Sbar.}
}
\description{
We here return a vector of strings/characters, i.e., a CharacterVector,
where each string is a space-separated list of integers.
}
\keyword{internal}
