% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{sample_gaussian}
\alias{sample_gaussian}
\title{Sample conditional Gaussian variables}
\usage{
sample_gaussian(given_ind, no_samp_mc, mu, sigma, p, xtest,
  ensure_condcov_symmetry = F)
}
\arguments{
\item{given_ind}{Vector}

\item{no_samp_mc}{Positive integer. Indicating the maximum number of samples to use in the
Monte Carlo integration for every conditional expectation (previously called \code{n_threshold})}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution.
NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{sigma}{Numeric matrix. (Optional) Containing the covariance matrix of the data generating
distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{p}{Positive integer}

\item{xtest}{Matrix, data.frame or data.table with the features, whose predictions ought
to be explained (test data)}
}
\value{
data.table with \code{no_samp_mc} (conditional) Gaussian samples
}
\description{
Sample conditional Gaussian variables
}
\author{
Martin Jullum
}
