% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{scale_data}
\alias{scale_data}
\title{Scale training and test data}
\usage{
scale_data(Xtrain, Xtest, scale = TRUE)
}
\arguments{
\item{Xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{Xtest}{Matrix, data.frame or data.table with the features, whose predictions ought to
be explained (test data)}

\item{scale}{Logical}
}
\value{
List
}
\description{
Scale training and test data
}
\author{
Nikolai Sellereite
}
