% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{set_ind_mc_cores}
\alias{set_ind_mc_cores}
\title{Helper function to set the number of cores for specific function}
\usage{
set_ind_mc_cores(mc_cores = 1, ...)
}
\arguments{
\item{mc_cores}{Integer. The overall number of mc_cores, to be used if nothing else is specified in ...}

\item{...}{mc_cores_simulateAllTrees (Integer) and exist_mc_cores_sample_ctree (Integer) to override mc_cores if specified, otherwise use mc_cores}
}
\description{
Helper function to set the number of cores for specific function
}
