% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_computation.R
\name{setup_computation}
\alias{setup_computation}
\title{Sets up everything for the Shapley values computation in [shapr::explain()]}
\usage{
setup_computation(internal, model, predict_model)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within
\code{\link[shapr:explain]{shapr:::explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data}, \code{functions}, \code{output}.}

\item{model}{Objects.
The model object that ought to be explained}
}
\value{
List \code{internal}
It holds all parameters, data, functions and computed objects used within
\code{\link[shapr:explain]{shapr:::explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data},
\code{functions}, \code{output}. [explain()]
[roxygen2::roxygenize()]
[explain()]
}
\description{
Sets up everything for the Shapley values computation in [shapr::explain()]
}
