% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_setup.R
\name{shapley_setup}
\alias{shapley_setup}
\title{Set up the kernelSHAP framework}
\usage{
shapley_setup(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
The internal list updated with the coalitions to be estimated
}
\description{
Set up the kernelSHAP framework
}
\keyword{internal}
