% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapley_weights}
\alias{shapley_weights}
\title{Calculate Shapley weight}
\usage{
shapley_weights(m, N, s, weight_zero_m = 10^6)
}
\arguments{
\item{m}{Positive integer}

\item{N}{Positive integer}

\item{s}{Positive integer}

\item{weight_zero_m}{Positive integer}
}
\value{
Numeric
}
\description{
Calculate Shapley weight
}
\author{
Nikolai Sellereite
}
