% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapr}
\alias{shapr}
\title{Get Shapley weights for test data}
\usage{
shapr(x, model, n_combinations = NULL)
}
\arguments{
\item{model}{Model object. Fitted model that is used to produce the predictions}

\item{compute_distances_for_no_var}{If equal to \code{NULL} no distances are computed}
}
\value{
Matrix
}
\description{
Get Shapley weights for test data
}
\author{
Nikolai Sellereite
}
