% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{sigma.hat.sq.func}
\alias{sigma.hat.sq.func}
\title{Variance of the local constant regression method. Used in the AICc-function.
Need to improve this help file.}
\usage{
\method{sigma}{hat.sq.func}(y, H)
}
\value{
numeric
}
\description{
Variance of the local constant regression method. Used in the AICc-function.
Need to improve this help file.
}
\author{
Martin Jullum
}
