% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{sigma_hat_squared}
\alias{sigma_hat_squared}
\title{Variance of the local constant regression method. Used in the AICc-function.
Need to improve this help file.}
\usage{
sigma_hat_squared(y, H)
}
\value{
numeric
}
\description{
Variance of the local constant regression method. Used in the AICc-function.
Need to improve this help file.
}
\author{
Martin Jullum
}
