% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\docType{data}
\name{skip_connection}
\alias{skip_connection}
\title{A \code{\link[torch:nn_module]{torch::nn_module()}} Representing a skip connection}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
skip_connection
}
\arguments{
\item{...}{network modules such as, e.g., \code{\link[torch:nn_linear]{torch::nn_linear()}}, \code{\link[torch:nn_relu]{torch::nn_relu()}},
and \code{\link[=memory_layer]{memory_layer()}} objects. See \code{\link[=vaeac]{vaeac()}} for more information.}
}
\description{
Skip connection over the sequence of layers in the constructor. The module passes
input data sequentially through these layers and then adds original data to the result.

This function is only available if the \code{torch} package is installed.
}
\author{
Lars Henry Berge Olsen
}
\keyword{#'}
\keyword{@title}
\keyword{Depends}
\keyword{Function}
\keyword{Torch}
\keyword{internal}
\keyword{on}
\keyword{that}
