% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{SkipConnection}
\alias{SkipConnection}
\title{A torch::nn_module representing a Skip Connection}
\usage{
SkipConnection(...)
}
\arguments{
\item{...}{network modules such as, e.g., \code{\link[torch]{nn_linear}}, \code{\link[torch]{nn_relu}},
and \code{\link{MemoryLayer}} objects. See \code{\link{get_imputation_networks}} for more information.}
}
\description{
Skip-connection over the sequence of layers in the constructor. The module passes
input data sequentially through these layers and then adds original data to the result.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
