% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{snapshot_review_man}
\alias{snapshot_review_man}
\title{Helper function for package development}
\usage{
snapshot_review_man(path, ...)
}
\arguments{
\item{path}{Character
Gives the relative path to the test files to review}
}
\description{
This is a manual extension of [testthat::snapshot_review()] which works for the \code{.rds} files used in
this package.
}
