% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{stable_format}
\alias{stable_format}
\title{Format numbers in a stable, OS-independent way, avoiding rounding surprises}
\usage{
stable_format(
  x,
  digits = 2L,
  justify = "right",
  guard = 2L,
  zero_scale_aware = TRUE
)
}
\arguments{
\item{x}{Numeric vector. The numbers to format.}

\item{digits}{Integer.
Number of significant digits to display.
Defaults to 2.}

\item{justify}{Character. The justification of the formatted numbers. Defaults to "right".}

\item{guard}{Integer. How many extra decimals below the print precision to consider "noise". Defaults to 2.}

\item{zero_scale_aware}{Logical. If TRUE, the function is aware of the scale of the numbers and adjusts the
epsilon accordingly. Defaults to TRUE.}
}
\value{
Character vector. The formatted numbers.
}
\description{
Format numbers in a stable, OS-independent way, avoiding rounding surprises
}
\keyword{internal}
