% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shapr}
\alias{summary.shapr}
\title{Summary method for shapr objects}
\usage{
\method{summary}{shapr}(x, ...)
}
\arguments{
\item{x}{A shapr object.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns a named list of summary components.
See the details section of \code{\link[=get_results]{get_results()}} for details about each component.
}
\description{
Summary method for shapr objects
}
