% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Model testing function}
\usage{
test_model(internal, model)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within
\code{\link[shapr:explain]{shapr:::explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data}, \code{functions}, \code{output}.}

\item{model}{Objects.
The model object that ought to be explained}
}
\description{
Model testing function
}
\keyword{internal}
