% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict_model.R
\name{test_predict_model}
\alias{test_predict_model}
\title{Model testing function}
\usage{
test_predict_model(x_test, predict_model, model, output_size = 1, extra = NULL)
}
\arguments{
\item{predict_model}{Function.
The prediction function used when \code{model} is not natively supported.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{model}{Objects.
The model object that ought to be explained.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{output_size}{TODO: Document}
}
\description{
Model testing function
}
\keyword{internal}
