% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{train_VAEAC_model}
\alias{train_VAEAC_model}
\title{Train the VAEAC Model}
\usage{
train_VAEAC_model(
  training_data,
  model_description = NULL,
  folder_to_save_model = NULL,
  use_cuda = FALSE,
  num_different_vaeac_initiate = 10,
  epochs_initiation_phase = 2,
  epochs = 200,
  save_VAEAC_every_nth_epoch = NULL,
  validation_ratio = 0.25,
  validation_iwae_num_samples = 25,
  depth = 3,
  width = 32,
  latent_dim = 8,
  lr = 0.001,
  batch_size = 64,
  running_avg_num_values = 5,
  activation_function = torch::nn_relu,
  use_skip_connections = TRUE,
  use_skip_connections_between_masked_encoder_and_decoder = TRUE,
  use_batch_normalization = FALSE,
  paired_sampling = TRUE,
  masking_ratio = 0.5,
  mask_generator_only_these_coalitions = NULL,
  mask_generator_only_these_coalitions_probabilities = NULL,
  sigma_mu = 10000,
  sigma_sigma = 1e-04,
  save_data = FALSE,
  transform_all_continuous_features = FALSE,
  verbose = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{training_data}{A matrix or data.frame containing the data. Categorical data must have class names 1,2,...,K.}

\item{model_description}{String containing, e.g., the name of the data distribution or additional parameter information. Used in the save name of the fitted model.}

\item{folder_to_save_model}{String specifying a path to a folder where the function is to save the fitted VAEAC model.}

\item{use_cuda}{Boolean. If we are to use cuda (GPU) if available.}

\item{num_different_vaeac_initiate}{Integer. The number of different VAEAC models to initiate in the start. Pick the best performing one after \code{epochs_initiation_phase } and continue training that one.}

\item{epochs_initiation_phase}{Integer. The number of epochs to run each of the \code{num_different_vaeac_initiate} VAEAC models before only continuing training the best one.}

\item{epochs}{Integer. The number of epochs to train the final VAEAC model. This includes \code{epochs_initiation_phase}.}

\item{save_VAEAC_every_nth_epoch}{Integer. If we are to save the VAEAC model after every nth epoch.}

\item{validation_ratio}{Scalar between 0 and 1 indicating the ratio of instances from data which will be used as validation data.}

\item{validation_iwae_num_samples}{Integer. The number of samples used to compute the IWAE when validating the VAEAC model on the validation data.}

\item{depth}{Integer. The number of hidden layers in the neural networks of the masked encoder, full encoder, and decoder.}

\item{width}{Integer. The number of neurons in each hidden layer in the neural networks of the masked encoder, full encoder, and decoder.}

\item{latent_dim}{Integer. The number of dimensions in the latent space.}

\item{lr}{Numeric. The learning rate used in the ADAM optimizer.}

\item{batch_size}{Integer. The number of samples to include in each batch.}

\item{running_avg_num_values}{Integer. How many of the previous values to include when we compute the running means.}

\item{activation_function}{An torch::nn_module representing an activation function. E.g., torch::nn_relu, torch::nn_leaky_relu, torch::nn_selu, torch::nn_sigmoid.}

\item{use_skip_connections}{Boolean. If we are to use skip connections in each layer. If true, then we add the input to the outcome of each hidden layer, so the output becomes X + activation(WX + b). I.e., identity skip connection.}

\item{use_skip_connections_between_masked_encoder_and_decoder}{Boolean. If we are to apply concatenate skip connections between the layers in the masked encoder and decoder.}

\item{use_batch_normalization}{Boolean. If we are to use batch normalization after the activation function. Note that if \code{use_skip_connections} is TRUE, then the normalization is
done after the adding from the skip connection. I.e, we batch normalize the whole quantity X + activation(WX + b).}

\item{paired_sampling}{Boolean. If we are doing paired sampling. I.e., each batch contains two versions of the same training observation,
but where the first one is masked by S and the second one is masked by \bar{S}, the complement. See FastSHAP by Jethani et al (2022).}

\item{masking_ratio}{Probability of masking a feature in the MCAR mask generator. Default masking scheme which ensures that
VAEAC can do arbitrary conditioning. Is overruled if \code{mask_generator_only_these_coalitions} is specified.}

\item{mask_generator_only_these_coalitions}{Matrix containing the different coalitions to learn.}

\item{mask_generator_only_these_coalitions_probabilities}{Numerics containing the probabilities for sampling each mask in \code{mask_generator_only_these_coalitions}.
Array containing the probabilities for sampling the coalitions in \code{mask_generator_only_these_coalitions}.}

\item{sigma_mu}{Numeric representing a hyperparameter in the normal-gamma prior used on the masked encoder, see Section 3.3.1 in Olsen et al. (2022).}

\item{sigma_sigma}{Numeric representing a hyperparameter in the normal-gamma prior used on the masked encoder, see Section 3.3.1 in Olsen et al. (2022).}

\item{save_data}{Boolean. If we are to save the data together with the model. Useful if one are to continue to train the model later.}

\item{transform_all_continuous_features}{Boolean. If we are to log transform all continuous features before sending the data to VAEAC.
VAEAC creates unbounded values, so if the continuous features are strictly positive, as for Burr and Abalone data, it can be advantageous
to log-transform the data to unbounded form before using VAEAC. If TRUE, then \code{VAEAC_postprocess_data} will take the exp of the results
to get back to strictly positive values when using the VAEAC model to impute missing values.}

\item{verbose}{Boolean. If we are to print the progress of the initialization of different VAEAC models, the training of the final VAEAC model, and summary of the training progress.}

\item{seed}{Integer. Seed for reproducibility.}

\item{...}{List of extra parameters, currently not used.}
}
\value{
A list containing the training/validation errors and paths to where the VAEAC models are saved on the disk.
}
\description{
Function that fits a VAEAC model to the given dataset based on the provided parameters.
}
\details{
One_hot_max_sizes is one for continuous variables and the number of categories for a categorical variable.
First fit several VAEAC models in the initialization phase to reduce the effect of potentially starting with poor value
}
