% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{true_Kshap}
\alias{true_Kshap}
\title{Function to calculate the true Shapley values based on the conditional expectations calculated using \code{cond_expec}}
\usage{
true_Kshap(explainer, cond_expec_mat, x_test)
}
\arguments{
\item{explainer}{explainer object from shapr package.}

\item{cond_expec_mat}{list. Calculated using \code{cond_expec_new} function.}

\item{x_test}{vector of test observations. Has the same dimension as the number of joint Normal random variables calculated in \code{sim_true_Normal} function.}
}
\value{
vector of Shapley values.
}
\description{

}
