% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_activation_func}
\alias{vaeac_check_activation_func}
\title{Function that checks the provided activation function}
\usage{
vaeac_check_activation_func(activation_function)
}
\arguments{
\item{activation_function}{An \code{\link[torch:nn_module]{torch::nn_module()}} representing an activation function such as, e.g.,
\code{\link[torch:nn_relu]{torch::nn_relu()}} (default), \code{\link[torch:nn_leaky_relu]{torch::nn_leaky_relu()}}, \code{\link[torch:nn_selu]{torch::nn_selu()}}, or \code{\link[torch:nn_sigmoid]{torch::nn_sigmoid()}}.}
}
\value{
The function does not return anything.
}
\description{
Function that checks the provided activation function
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
