% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_extra_functions.R
\name{vaeac_check_cuda}
\alias{vaeac_check_cuda}
\title{Function that checks for access to CUDA}
\usage{
vaeac_check_cuda(cuda)
}
\arguments{
\item{cuda}{cuda Logical (default is \code{FALSE}). If \code{TRUE}, then the \code{vaeac} model will be trained using cuda/GPU.
If \code{\link[torch:cuda_is_available]{torch::cuda_is_available()}} is \code{FALSE}, the we fall back to use CPU. If \code{FALSE}, we use the CPU. Often this is
faster for tabular data sets. Note, cuda is not not supported in the current version of the \code{shapr} package.
TODO: Update this when this is done.}
}
\value{
The function does not return anything.
}
\description{
Function that checks for access to CUDA
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
