% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_logicals}
\alias{vaeac_check_logicals}
\title{Function that checks logicals}
\usage{
vaeac_check_logicals(named_list_logicals, return_as_class_logical = FALSE)
}
\arguments{
\item{named_list_logicals}{List containing named entries. I.e., \code{list(a = TRUE, b = FALSE)}.}
}
\value{
The entries in \code{named_list_logicals} as of class numerics if \code{named_list_logicals = TRUE}.
}
\description{
Function that checks logicals
}
\author{
Lars Henry Berge Olsen
}
