% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_mask_gen}
\alias{vaeac_check_mask_gen}
\title{Function that checks the specified masking scheme}
\usage{
vaeac_check_mask_gen(
  mask_gen_these_coalitions,
  mask_gen_these_coalitions_prob,
  training_data
)
}
\arguments{
\item{mask_gen_these_coalitions}{Matrix containing the different coalitions to learn.}

\item{mask_gen_these_coalitions_prob}{Numerics containing the probabilities for
sampling each mask in \code{mask_gen_these_coalitions}.
Array containing the probabilities for sampling the coalitions in \code{mask_gen_these_coalitions}.}

\item{training_data}{A matrix or data.frame containing the data.
Categorical data must have class names \eqn{1,2,\dots,K}.}
}
\value{
The function does not return anything.
}
\description{
Function that checks the specified masking scheme
}
\author{
Lars Henry Berge Olsen
}
