% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_check_functions.R
\name{vaeac_check_masking_ratio}
\alias{vaeac_check_masking_ratio}
\title{Function that checks that the masking ratio argument is valid}
\usage{
vaeac_check_masking_ratio(masking_ratio, n_features)
}
\arguments{
\item{masking_ratio}{Probability of masking a feature in the MCAR mask generator.
Default masking scheme which ensures that vaeac can do arbitrary conditioning.
This is overruled if \code{mask_gen_these_coalitions} is specified.}

\item{n_features}{The number of features, i.e., the number of columns in the training data.}
}
\value{
The function does not return anything.
}
\description{
Function that checks that the masking ratio argument is valid
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
