% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_positive_integers}
\alias{vaeac_check_positive_integers}
\title{Function that checks positive integers}
\usage{
vaeac_check_positive_integers(
  named_list_positive_integers,
  return_as_class_integer = FALSE
)
}
\arguments{
\item{named_list_positive_integers}{List containing named entries. I.e., \code{list(a = 1, b = 2)}.}
}
\value{
The entries in \code{named_list_positive_integers} as of class integer if \code{return_as_class_integer = TRUE}.
}
\description{
Function that checks positive integers
}
\author{
Lars Henry Berge Olsen
}
