% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_positive_numerics}
\alias{vaeac_check_positive_numerics}
\title{Function that checks positive numerics}
\usage{
vaeac_check_positive_numerics(
  named_list_positive_numerics,
  return_as_class_numeric = FALSE
)
}
\arguments{
\item{named_list_positive_numerics}{List containing named entries. I.e., \code{list(a = 0.2, b = 10^3)}.}
}
\value{
The entries in \code{named_list_positive_numerics} as of class integer if \code{return_as_class_numeric = TRUE}.
}
\description{
Function that checks positive numerics
}
\author{
Lars Henry Berge Olsen
}
