% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_save_names}
\alias{vaeac_check_save_names}
\title{Function that checks that the save folder exists and for a valid file name}
\usage{
vaeac_check_save_names(folder_to_save_model, model_description)
}
\arguments{
\item{folder_to_save_model}{String specifying a path to a folder where
the function is to save the fitted vaeac model.}

\item{model_description}{String containing, e.g., the name of the data distribution or
additional parameter information. Used in the save name of the fitted model.}
}
\value{
The function does not return anything.
}
\description{
Function that checks that the save folder exists and for a valid file name
}
\author{
Lars Henry Berge Olsen
}
