% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_check_functions.R
\name{vaeac_check_save_parameters}
\alias{vaeac_check_save_parameters}
\title{Function that gives a warning about disk usage}
\usage{
vaeac_check_save_parameters(save_data, save_every_nth_epoch, x_train_size)
}
\arguments{
\item{save_data}{Boolean. If we are to save the data together with the model. Useful if one are to continue
to train the model later.}

\item{save_every_nth_epoch}{Integer. If we are to save the vaeac model after every nth epoch.}

\item{x_train_size}{The object size of the \code{x_train} object.}
}
\value{
The function does not return anything.
}
\description{
Function that gives a warning about disk usage
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
