% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_extra_functions.R
\name{vaeac_check_verbose}
\alias{vaeac_check_verbose}
\title{Function the checks the verbose parameter}
\usage{
vaeac_check_verbose(verbose)
}
\arguments{
\item{verbose}{Boolean. If we are to print the progress of the initialization of different vaeac models,
the training of the final vaeac model, and summary of the training progress.}
}
\value{
The function does not return anything.
}
\description{
Function the checks the verbose parameter
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
