% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_extra_functions.R
\name{vaeac_check_which_vaeac_model}
\alias{vaeac_check_which_vaeac_model}
\title{Function that checks for valid \code{vaeac} model name}
\usage{
vaeac_check_which_vaeac_model(
  which_vaeac_model,
  epochs,
  save_every_nth_epoch = NULL
)
}
\arguments{
\item{epochs}{Integer. The number of epochs to train the final vaeac model.
This includes \code{epochs_initiation_phase}.}

\item{save_every_nth_epoch}{Integer. If we are to save the vaeac model after every nth epoch.}
}
\value{
The function does not return anything.
}
\description{
Function that checks for valid \code{vaeac} model name
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
