% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_x_colnames}
\alias{vaeac_check_x_colnames}
\title{Function that checks the feature names of data and \code{vaeac} model}
\usage{
vaeac_check_x_colnames(feature_names_vaeac, feature_names_new)
}
\arguments{
\item{feature_names_vaeac}{Array of strings containing the feature names of the \code{vaeac} model.}

\item{feature_names_new}{Array of strings containing the feature names to compare with.}
}
\value{
The function does not return anything.
}
\description{
Function that checks the feature names of data and \code{vaeac} model
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
