% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_continue_train_model}
\alias{vaeac_continue_train_model}
\title{Continue to Train the vaeac Model}
\usage{
vaeac_continue_train_model(
  explanation,
  epochs_new,
  lr_new = NULL,
  x_train = NULL,
  save_data = FALSE,
  verbose = 0,
  seed = 1
)
}
\arguments{
\item{explanation}{A \code{\link[=explain]{explain()}} object and \code{vaeac} must be the used approach.}

\item{epochs_new}{Integer. The number of extra epochs to conduct.}

\item{lr_new}{Numeric. If we are to overwrite the old learning rate in the adam optimizer.}

\item{x_train}{Matrix/data.frame containing new training data. If not present,
then we try to load training data from the vaeac_model.}

\item{save_data}{Boolean. If we are to save the training data.}

\item{verbose}{Boolean. If we are to print out information to the user.}
}
\value{
A list containing the training/validation errors and paths to where the vaeac models are saved on the disk.
}
\description{
Function that loads a previously trained vaeac model and continue the training, either
on new data or on the same dataset as it was trained on before. If we are given a new dataset, then
we assume that new dataset has the same distribution and one_hot_max_sizes as the original dataset.
}
\author{
Lars Henry Berge Olsen
}
