% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{VAEAC_dataset}
\alias{VAEAC_dataset}
\title{Dataset used to load data into the VAEAC model}
\usage{
VAEAC_dataset(X, one_hot_max_sizes)
}
\arguments{
\item{X}{A torch_tensor contain the data}

\item{one_hot_max_sizes}{A torch tensor of dimension p containing the one hot sizes of the p features.
The sizes for the continuous features can either be '0' or '1'.}
}
\description{
Create a new VAEAC_dataset object.

How to fetch a data sample for a given key/index.

Return the size of the dataset
}
\details{
Must create a data set object that represent a map from keys to data samples. It is used by the dataloader()
to load data which should be used to extract the batches for all epochs in the training phase of the neural
network. Note that a dataset object is an R6 instance (https://r6.r-lib.org/articles/Introduction.html)
which is classical object-oriented programming, with self reference. I.e, 'VAEAC_sample_data' is  a class of type 'dataset'.
}
\section{Fields}{

\describe{
\item{\code{A}}{dataset must specify a 'name' for the object}
}}

\examples{
p = 5
N = 14
batch_size = 10
one_hot_max_sizes = rep(1, p)
VAEAC_ds = VAEAC_dataset(torch_tensor(matrix(rnorm(p*N), ncol = p), dtype = torch_float()), one_hot_max_sizes)
VAEAC_ds

VAEAC_dl = dataloader(VAEAC_ds, batch_size = batch_size, shuffle = TRUE, drop_last = FALSE)
VAEAC_dl$.length()
VAEAC_dl$.iter()

VAEAC_iterator = VAEAC_dl$.iter()
VAEAC_iterator$.next() # batch1
VAEAC_iterator$.next() # batch2
VAEAC_iterator$.next() # Empty
}
