% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_get_current_save_state}
\alias{vaeac_get_current_save_state}
\title{Function that extracts additional objects from the environment into the state list}
\usage{
vaeac_get_current_save_state(environment)
}
\arguments{
\item{environment}{The \code{\link[base:environment]{base::environment()}} where the objects are stored.}
}
\value{
List containing the values of \code{epoch}, \code{train_vlb}, \code{val_iwae}, \code{val_iwae_running},
and the \code{state_dict()} of the vaeac model and optimizer.
}
\description{
The function extract the objects that we are going to save together with the \code{vaeac} model to make it possible to
train the model further and to evaluate it.
The environment should be the local environment inside the \code{\link[=vaeac_train_model_auxiliary]{vaeac_train_model_auxiliary()}} function.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
