% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_get_evaluation_criteria}
\alias{vaeac_get_evaluation_criteria}
\title{Extract the Training VLB and Validation IWAE from a List of Explanations Objects Using the \code{vaeac} Approach}
\usage{
vaeac_get_evaluation_criteria(explanation_list)
}
\arguments{
\item{explanation_list}{A list of \code{\link[=explain]{explain()}} objects applied to the same data, model, and
\code{vaeac} must be the used approach. If the entries in the list is named, then the function use
these names. Otherwise, it defaults to the approach names (with integer suffix for duplicates)
for the explanation objects in \code{explanation_list}.}
}
\value{
A data.table containing the training VLB, validation IWAE, and running validation IWAE at each epoch for
each vaeac model.
}
\description{
Extract the Training VLB and Validation IWAE from a List of Explanations Objects Using the \code{vaeac} Approach
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
