% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_get_n_decimals}
\alias{vaeac_get_n_decimals}
\title{Function to get string of values with specific number of decimals}
\usage{
vaeac_get_n_decimals(value, n_decimals = 3)
}
\arguments{
\item{value}{The number to get \code{n_decimals} for.}

\item{n_decimals}{Positive integer. The number of decimals. Default is three.}
}
\value{
String of \code{value} with \code{n_decimals} decimals.
}
\description{
Function to get string of values with specific number of decimals
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
