% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_extra_functions.R
\name{vaeac_get_optimizer}
\alias{vaeac_get_optimizer}
\title{Function to create the optimizer used to train \code{vaeac}}
\usage{
vaeac_get_optimizer(vaeac_model, optimizer_name = "adam")
}
\arguments{
\item{vaeac_model}{A \code{vaeac} model created using \code{\link[=vaeac]{vaeac()}}.}

\item{optimizer_name}{String containing the name of the \code{\link[torch:optimizer]{torch::optimizer()}} to use.}
}
\value{
Array of string containing the save files to use when training the \code{vaeac} model. The first three names
corresponds to the best, best_running, and last epochs, in that order.
}
\description{
Only \code{\link[torch:optim_adam]{torch::optim_adam()}} is currently supported. But it is easy to add an additional option later.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
