% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_get_x_explain_extended}
\alias{vaeac_get_x_explain_extended}
\title{Function to extend the explicands and apply all relevant masks/coalitions}
\usage{
vaeac_get_x_explain_extended(x_explain, S, index_features)
}
\arguments{
\item{x_explain}{Matrix or data.frame/data.table.
Features for which predictions should be explained.}

\item{S}{The \code{internal$objects$S} matrix containing the possible coalitions.}

\item{index_features}{Positive integer vector. Specifies the id_coalition to
apply to the present method. \code{NULL} means all coalitions. Only used internally.}
}
\value{
The extended version of \code{x_explain} where the masks from \code{S} with indices \code{index_features} have been applied.
}
\description{
Function to extend the explicands and apply all relevant masks/coalitions
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
