% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_impute_missing_entries}
\alias{vaeac_impute_missing_entries}
\title{Impute Missing Values Using Vaeac}
\usage{
vaeac_impute_missing_entries(
  x_explain_with_NaNs,
  path_vaeac_model,
  n_samples,
  use_cuda = FALSE,
  sample_random = TRUE,
  convert_to_2D = TRUE,
  return_as_postprocessed_dt = TRUE,
  batch_size = NULL,
  verbose = FALSE,
  seed = NULL,
  index_features = NULL
)
}
\arguments{
\item{x_explain_with_NaNs}{A 2D matrix, where the missing entries to impute are represented by \code{NaN}.}

\item{path_vaeac_model}{String. The location of the saved vaeac model.}

\item{n_samples}{Integer. The number of imputed versions we create for each row in \code{x_explain_with_NaNs}.}

\item{use_cuda}{Boolean. If we are to use cuda. NOTE PROPERLY TESTED IN CURRENT VERSION.}

\item{sample_random}{Boolean. If we are to generate random samples from the inferred generative distributions,
or if we are to sample the most likely values (mean for continuous and class with highest prob for categorical).}

\item{convert_to_2D}{Boolean. If the returned results should be of shape
[\code{nrow(x_explain_with_NaNs)}, \code{n_samples}, \code{n_features}] or
[(\code{nrow(x_explain_with_NaNs)} \eqn{\times} \code{n_samples}), \code{n_features}].}

\item{return_as_postprocessed_dt}{Boolean. If we are to postprocess the data, i.e.,
convert categorical features to factors with correct level names (and transform continuous features
back to original scale). The returned object will then be a \code{\link[data.table]{data.table}}.}

\item{batch_size}{Integer. The number of samples in each batch.
If \code{NULL}, then use the same value saved in the vaeac object.
We recommend a large number because of a lot of overhead for small batch sizes, that is,
\code{batch_size} should be much larger than the batch size used during the training.}

\item{verbose}{Boolean. If we are to print the progress to the user.}

\item{seed}{Integer. Seed used before generating the MC samples.}

\item{index_features}{Optional integer vector. Used internally in shapr package to index the coalitions.}
}
\value{
A 2D or 3D array or 2D data.table where the missing values (\code{NaN}) in \code{x_explain_with_NaNs}
have been imputed \code{n_samples} times. The dimension of the returned object depends on the
boolean parameters \code{convert_to_2D} and \code{return_as_postprocessed_dt}.
}
\description{
Impute Missing Values Using Vaeac
}
\details{
Function that imputes the missing values in 2D matrix where each row constitute an individual.
The values are sampled from the conditional distribution estimated by a vaeac model.
}
\author{
Lars Henry Berge Olsen
}
