% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{VAEAC_impute_values}
\alias{VAEAC_impute_values}
\title{Impute Missing Values Using VAEAC}
\usage{
VAEAC_impute_values(
  instances_to_impute,
  path_VAEAC_model,
  num_imputations,
  use_cuda = FALSE,
  sample_random = TRUE,
  convert_to_2D = TRUE,
  return_as_postprocessed_data_table = TRUE,
  batch_size = NULL,
  verbose = FALSE
)
}
\arguments{
\item{instances_to_impute}{A 2^p (maybe 2^p-2) x p matrix. Should be 'NaN' where we want to impute the missing values.}

\item{path_VAEAC_model}{String containing the location of the saved VAEAC model}

\item{num_imputations}{the number of imputed versions we create for each row in 'instances_to_impute'}

\item{use_cuda}{Boolean. If we are to use cuda.}

\item{sample_random}{Boolean. If we are to generate random samples from the inferred generative distributions, or if we are to sample the most likely values (mean for cont, class with highest prob for cat).}

\item{convert_to_2D}{Boolean. If the returned results should be of shape \link{num_instances_to_impute, num_imputations, num_features} or
\link{(num_instances_to_impute * num_imputations), num_features}.}

\item{return_as_postprocessed_data_table}{Boolean. If we are to postprocess the data, i.e., convert categorical features to factors
with correct level names (and transform continuous features back to original scale). The returned object will then be a data table.}

\item{batch_size}{Integer. The number of samples in each batch. We recommend a large number because of a lot of overhead for small
batch sizes, that is, much larger than the batch size used during training. If NULL, then use the same value saved in the VAEAC object.}

\item{verbose}{Boolean. If we are to print the progress to the user.}
}
\value{
An array where the missing values (NaN) in \code{instances_to_impute} have been imputed \code{num_imputations} times.
}
\description{
Impute Missing Values Using VAEAC
}
\details{
Function that imputes the missing values in 2D matrix where each row constitute an individual.
The values are sampled from the conditional distribution estimated by a VAEAC model. Note that this function
does not preprocess the data.
}
