% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{vaeac_kl_normal_normal}
\alias{vaeac_kl_normal_normal}
\title{Compute the KL Divergence Between Two Gaussian Distributions.}
\usage{
vaeac_kl_normal_normal(p, q)
}
\arguments{
\item{p}{A \code{\link[torch:distr_normal]{torch::distr_normal()}} object.}

\item{q}{A \code{\link[torch:distr_normal]{torch::distr_normal()}} object.}
}
\value{
The KL divergence between the two Gaussian distributions.
}
\description{
Computes the KL divergence between univariate normal distributions using the analytical formula,
see \url{https://en.wikipedia.org/wiki/Kullback\%E2\%80\%93Leibler_divergence#Multivariate_normal_distributions}.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
