% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{VAEAC_postprocess_data}
\alias{VAEAC_postprocess_data}
\title{Postprocess Data Generated by a VAEAC Model}
\usage{
VAEAC_postprocess_data(data, VAEAC_model_state_list)
}
\arguments{
\item{data}{Matrix/data.frame/data.table containing the data generated by a VAEAC model}

\item{VAEAC_model_state_list}{List. The returned list from the \code{VAEAC_preprocess_data} function.}
}
\value{
data.table with the generated data from a VAEAC model where the categorical features
now have the original class names.
}
\description{
VAEAC generates numerical values. This function converts categorical features
to from numerics with class labels 1,2,...,K, to factors with the original and class labels.
}
\examples{
preprocessed = VAEAC_preprocess_data(abalone[,-1])
preprocessed$x_train
postprocessed = VAEAC_postprocess_data(preprocessed$x_train, preprocessed)
postprocessed
all.equal(preprocessed$x_train_org, postprocessed)
}
