% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_save_state}
\alias{vaeac_save_state}
\title{Function that saves the state list and the current save state of the \code{vaeac} model}
\usage{
vaeac_save_state(state_list, file_name, return_state = FALSE)
}
\arguments{
\item{state_list}{List containing all the parameters in the state.}

\item{file_name}{String containing the file path.}

\item{return_state}{Logical if we are to return the state list or not.}
}
\value{
This function does not return anything
}
\description{
Function that saves the state list and the current save state of the \code{vaeac} model
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
