% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_extra_functions.R
\name{vaeac_train_model_auxiliary}
\alias{vaeac_train_model_auxiliary}
\title{Function used to train a \code{vaeac} model}
\usage{
vaeac_train_model_auxiliary(
  vaeac_model,
  optimizer,
  train_dataloader,
  val_dataloader,
  val_iwae_n_samples,
  running_avg_n_values,
  verbose,
  cuda,
  epochs,
  save_every_nth_epoch,
  epochs_early_stopping,
  epochs_start = 1,
  progressr_bar = NULL,
  vaeac_save_file_names = NULL,
  state_list = NULL,
  initialization_idx = NULL,
  n_vaeacs_initialize = NULL,
  train_vlb = NULL,
  val_iwae = NULL,
  val_iwae_running = NULL
)
}
\arguments{
\item{vaeac_model}{A \code{\link[=vaeac]{vaeac()}} object. The \code{vaeac} model this function is to train.}

\item{optimizer}{A \code{\link[torch:optimizer]{torch::optimizer()}} object. See \code{\link[=vaeac_get_optimizer]{vaeac_get_optimizer()}}.}

\item{train_dataloader}{A \code{\link[torch:dataloader]{torch::dataloader()}} containing the training data for the \code{vaeac} model.}

\item{val_dataloader}{A \code{\link[torch:dataloader]{torch::dataloader()}} containing the validation data for the \code{vaeac} model.}

\item{val_iwae_n_samples}{Integer. The number of samples used to compute the
IWAE when validating the vaeac model on the validation data.}

\item{running_avg_n_values}{Integer. How many of the previous values to include when we compute the running means.}

\item{verbose}{Boolean. If we are to print the progress of the initialization of different vaeac models,
the training of the final vaeac model, and summary of the training progress.}

\item{cuda}{Boolean. If we are to use cuda (GPU) if available. STILL IN DEVELOPMENT!}

\item{epochs}{Integer. The number of epochs to train the final vaeac model.
This includes \code{epochs_initiation_phase}.}

\item{save_every_nth_epoch}{Integer. If we are to save the vaeac model after every nth epoch.}

\item{epochs_early_stopping}{Integer. The training stops if there has been no improvement in the validation IWAE
for \code{epochs_early_stopping} epochs. If the user wants the training process to be solely based on this, then \code{epochs}
should be set to a large number.}

\item{epochs_start}{Positive integer (default is \code{1}). At which epoch the training is starting at.}

\item{progressr_bar}{A \code{\link[progressr:progressor]{progressr::progressor()}} object (default is \code{NULL}) to keep track of progress.}

\item{vaeac_save_file_names}{Array of strings containing the save file names for the \code{vaeac} model.}

\item{state_list}{Named list containing the objects returned from \code{\link[=vaeac_get_full_state_list]{vaeac_get_full_state_list()}}.}

\item{initialization_idx}{Positive integer (default is \code{NULL}). The index
of the current \code{vaeac} model in the initialization phase.}

\item{n_vaeacs_initialize}{Integer. The number of different vaeac models to initiate in the start.
Pick the best performing one after \code{epochs_initiation_phase} and continue training that one.}

\item{train_vlb}{A \code{\link[torch:torch_tensor]{torch::torch_tensor()}} (default is \code{NULL})
of one dimension containing previous values for the training VLB.}

\item{val_iwae}{A \code{\link[torch:torch_tensor]{torch::torch_tensor()}} (default is \code{NULL})
of one dimension containing previous values for the validation IWAE.}

\item{val_iwae_running}{A \code{\link[torch:torch_tensor]{torch::torch_tensor()}} (default is \code{NULL})
of one dimension containing previous values for the running validation IWAE.}
}
\value{
Depending on if we are in the initialization phase or not. Then either the trained \code{vaeac} model, or
a list of where the \code{vaeac} models are stored on disk and the parameters of the model.
}
\description{
This function can be applied both in the initialization phase when, we train several initiated \code{vaeac} models, and
to keep training the best performing \code{vaeac} model for the remaining number of epochs. We are in the former setting
when \code{initialization_idx} is provided and the latter when it is \code{NULL}. When it is \code{NULL}, we save the \code{vaeac} models
with lowest VLB, IWAE, running IWAE, and the epochs according to \code{save_every_nth_epoch} to disk.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
