% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_check_functions.R
\name{vaeac_train_print_summary}
\alias{vaeac_train_print_summary}
\title{Function to printout a training summary for the \code{vaeac} model}
\usage{
vaeac_train_print_summary(best_state, best_state_running, last_state)
}
\arguments{
\item{best_state}{The state list (i.e., the saved \code{vaeac} object) of the \code{vaeac}
model at the epoch with the lowest IWAE.}

\item{best_state_running}{The state list of (i.e., the saved \code{vaeac} object)
the \code{vaeac} model at the epoch with the lowest running IWAE.}

\item{last_epoch}{The state list (i.e., the saved \code{vaeac} object)
of \code{vaeac} model at the epoch with the lowest IWAE.}
}
\value{
This function only prints out a message.
}
\description{
Function to printout a training summary for the \code{vaeac} model
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
