% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{VAEAC_training_vlb_and_validation_iwae_shapr}
\alias{VAEAC_training_vlb_and_validation_iwae_shapr}
\title{Extract the Training VLB and Validation IWAE from the Provided VAEAC Model}
\usage{
VAEAC_training_vlb_and_validation_iwae_shapr(
  explanation,
  plot_figure = TRUE,
  plot_from_nth_epoch = 1,
  return_figure = FALSE,
  return_training_validation_errors = FALSE
)
}
\arguments{
\item{explanation}{List. The output list from the shapr::explanation function.}

\item{plot_from_nth_epoch}{Integer. From which epoch to plot from. First epoch can be large in absolute value and make the rest of the plot difficult to interpret.}

\item{plot}{Boolean. If we are to plot the train VLB and validation IWAE, and add figure to return list.}
}
\value{
A list containing the training VLB and validation IWAE at each epoch, and the total number of epochs.
}
\description{
Extract the Training VLB and Validation IWAE from the Provided VAEAC Model
}
