% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_update_para_locations}
\alias{vaeac_update_para_locations}
\title{Move \code{vaeac} parameters to correct location}
\usage{
vaeac_update_para_locations(parameters)
}
\arguments{
\item{parameters}{List. The \code{internal$parameters} list created inside the \code{\link[=explain]{explain()}} function.}
}
\value{
Updated version of \code{parameters} where all \code{vaeac} parameters are located at the correct location.
}
\description{
This function ensures that the main and extra parameters for the \code{vaeac}
approach is located at their right locations.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
