% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{w_shapley}
\alias{w_shapley}
\title{Calculate Shapley weight}
\usage{
w_shapley(m, N, s)
}
\arguments{
\item{m}{Integer. Total number of features}

\item{N}{Integer. Number of combinations}

\item{s}{Integer. Number of chosen}
}
\value{
Numeric
}
\description{
Calculate Shapley weight
}
\author{
Nikolai Sellereite
}
