% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{weight_matrix}
\alias{weight_matrix}
\title{Calculate weighted matrix}
\usage{
weight_matrix(X, use_shapley_weights_in_W = TRUE,
  normalize_W_weights = TRUE)
}
\arguments{
\item{X}{data.table}
}
\value{
Matrix
}
\description{
Calculate weighted matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
