% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weight_matrix_cpp}
\alias{weight_matrix_cpp}
\title{Get distance}
\usage{
weight_matrix_cpp(features, m, n, w)
}
\arguments{
\item{features}{List.}

\item{m}{Integer. Total number of features}

\item{n}{Positive integer. Number of combinations}

\item{w}{Matrix}
}
\value{
Matrix of dimension n x m + 1
}
\description{
Get distance
}
\author{
Nikolai Sellereite
}
