% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_matrix}
\alias{weighted_matrix}
\title{Get distance}
\usage{
weighted_matrix(features, m, n, w)
}
\arguments{
\item{features}{List.}

\item{m}{Integer. Total number of features}

\item{n}{Positive integer. Number of combinations}
}
\value{
Matrix of dimension n x m + 1
}
\description{
Get distance
}
\author{
Nikolai Sellereite
}
