% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{additional_regression_setup}
\alias{additional_regression_setup}
\title{Additional Setup for Regression-Based Methods}
\usage{
additional_regression_setup(internal, model, predict_model)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
The (updated) internal list
}
\description{
Additional Setup for Regression-Based Methods
}
\keyword{internal}
