% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aicc_full_cpp}
\alias{aicc_full_cpp}
\title{AICc formula for several sets, alternative definition}
\usage{
aicc_full_cpp(h, X_list, mcov_list, S_scale_dist, y_list, negative)
}
\arguments{
\item{h}{numeric specifying the scaling (sigma)}

\item{X_list}{List.
Contains matrices with the appropriate features of the training data}

\item{mcov_list}{List.
Contains the covariance matrices of the matrices in X_list}

\item{S_scale_dist}{Logical.
Indicates whether Mahalanobis distance should be scaled with the number of variables.}

\item{y_list}{List.
Contains the appropriate (temporary) response variables.}

\item{negative}{Logical.
Whether to return the negative of the AICc value.}
}
\value{
Scalar with the numeric value of the AICc formula
}
\description{
AICc formula for several sets, alternative definition
}
\author{
Martin Jullum
}
\keyword{internal}
