% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aicc_full_single_cpp}
\alias{aicc_full_single_cpp}
\title{Temp-function for computing the full AICc with several X's etc}
\usage{
aicc_full_single_cpp(X, mcov, S_scale_dist, h, y)
}
\arguments{
\item{X}{matrix with "covariates"}

\item{mcov}{covariance matrix}

\item{S_scale_dist}{logical indicating whether the Mahalanobis distance should be scaled with the number of variables}

\item{h}{numeric specifying the scaling (sigma)}

\item{y}{vector with the "response variable"}
}
\value{
Scalar with the numeric value of the AICc formula
}
\description{
Temp-function for computing the full AICc with several X's etc
}
\author{
Martin Jullum
}
