% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{apply_dummies}
\alias{apply_dummies}
\title{Make dummy variables - this is an internal function intended only to be used in
predict_model.xgb.Booster()}
\usage{
apply_dummies(feature_list, testdata)
}
\arguments{
\item{feature_list}{List. Output of \code{make_dummies$feature_list}.}

\item{testdata}{data.table or data.frame. New data that has the same
feature names, types, and levels as \code{obj$data}.}
}
\value{
A data.frame containing all of the factors in \code{testdata} as
one-hot encoded variables.
}
\description{
Make dummy variables - this is an internal function intended only to be used in
predict_model.xgb.Booster()
}
\author{
Annabelle Redelmeier
}
\keyword{internal}
