% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{apply_dummies}
\alias{apply_dummies}
\title{Make dummy variables - this is an internal function intended only to be used in
predict_model.xgb.Booster()}
\usage{
apply_dummies(obj, newdata)
}
\arguments{
\item{obj}{List. Output of \code{make_dummies}.}

\item{newdata}{data.table or data.frame. New data (features) that has the same
features as the data used in \code{make_dummies}.}
}
\value{
A data.frame containing all of the factors in \code{new_data} as
one-hot encoded variables.
}
\description{
Make dummy variables - this is an internal function intended only to be used in
predict_model.xgb.Booster()
}
\author{
Annabelle Redelmeier
}
\keyword{internal}
