% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_groups}
\alias{check_groups}
\title{Check that the group parameter has the right form and content}
\usage{
check_groups(feature_names, group)
}
\description{
Check that the group parameter has the right form and content
}
\keyword{internal}
