% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_compute_vS}
\alias{cli_compute_vS}
\title{Print Messages in Compute_vS with CLI}
\usage{
cli_compute_vS(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
No return value (but prints compute_vS messages with cli)
}
\description{
Print Messages in Compute_vS with CLI
}
\keyword{internal}
