% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_features}
\alias{cluster_features}
\title{Group variables according to matrix of Kendall's tau}
\usage{
cluster_features(x_mat = NULL, cor_mat = NULL, alpha = 1)
}
\arguments{
\item{x_mat}{optional data matrix, ignores \code{cor_mat} if included}

\item{cor_mat}{optional correlation matrix, ignores \code{x_mat} if included. Either \code{x_mat} or
\code{cor_mat} must be included}

\item{alpha}{optional tuning parameter for optimal number of clusters}
}
\value{
List
}
\description{
Group variables according to matrix of Kendall's tau
}
\author{
Anders Løland
}
