% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_features}
\alias{cluster_features}
\title{Group variables according to matrix of Kendall's tau}
\usage{
cluster_features(x_mat = NULL, cor_mat = NULL, alpha = 1)
}
\arguments{
\item{x_mat}{Numeric matrix. Ignored if \code{cor_mat} is included.}

\item{cor_mat}{Numeric matrix. Note that \code{x_mat} or \code{cor_mat} must be included,
otherwise the function will return an error message.}

\item{alpha}{Numeric. Weight for number of clusters. See \code{\link[maptree]{kgs}}.}
}
\value{
List
}
\description{
Group variables according to matrix of Kendall's tau
}
\details{
The returned list contains the following items
\describe{
  \item{cor_mat}{TODO: Add information}
  \item{which_k}{TODO: Add information}
  \item{cluster}{TODO: Add information}
}
}
\author{
Anders Løland
}
