% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coalition_matrix_cpp}
\alias{coalition_matrix_cpp}
\title{Get coalition matrix}
\usage{
coalition_matrix_cpp(coalitions, m)
}
\arguments{
\item{coalitions}{List.
Each of the elements equals an integer vector representing a valid combination of features/feature groups.}

\item{m}{Integer.
Number of features/feature groups.}
}
\value{
Matrix
}
\description{
Get coalition matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
