% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{compute_kshap}
\alias{compute_kshap}
\title{Note that this function is deprecated, but we'll keep it for a week
to check that results are stable.}
\usage{
compute_kshap(model, l, noSamp_MC = 1000, verbose = FALSE,
  cond_approach = "empirical", empirical_settings = list(type =
  "fixed_sigma", fixed_sigma_vec = 0.1, n_samples_aicc = 1000,
  eval_max_aicc = 20, start_aicc = 0.1, w_threshold = 0.95), pred_zero,
  mu = NULL, Sigma = NULL)
}
\description{
TODO: Delete this function from the codebase
}
\keyword{internal}
