% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_shapley}
\alias{compute_shapley}
\title{Compute Shapley Values}
\usage{
compute_shapley(internal, dt_vS)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}

\item{dt_vS}{The contribution matrix.}
}
\value{
A \code{data.table} with Shapley values for each test observation.
}
\description{
Compute Shapley Values
}
\keyword{internal}
