% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_explanation.R
\name{compute_shapley_new}
\alias{compute_shapley_new}
\title{Compute shapley values}
\usage{
compute_shapley_new(internal, dt_vS)
}
\arguments{
\item{dt_vS}{The contribution matrix.}

\item{explainer}{An \code{explain} object.}
}
\value{
A \code{data.table} with shapley values for each test observation.
}
\description{
Compute shapley values
}
\keyword{internal}
