% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_explanation.R
\name{compute_shapley_new}
\alias{compute_shapley_new}
\title{Compute shapley values}
\usage{
compute_shapley_new(internal, dt_vS)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}

\item{dt_vS}{The contribution matrix.}
}
\value{
A \code{data.table} with Shapley values for each test observation.
}
\description{
Compute shapley values
}
\keyword{internal}
