% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{compute_time}
\alias{compute_time}
\title{Gather and Compute the Timing of the Different Parts of the Explain Function}
\usage{
compute_time(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
A list with reformatted timing information.
}
\description{
Gather and Compute the Timing of the Different Parts of the Explain Function
}
\keyword{internal}
