% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vS.R
\name{compute_vS}
\alias{compute_vS}
\title{Computes \code{v(S)} for all features subsets \code{S}.}
\usage{
compute_vS(internal, model, method = "lapply")
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within
\code{\link[shapr:explain]{shapr:::explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data}, \code{functions}, \code{output}.}

\item{model}{Objects.
The model object that ought to be explained}

\item{method}{Character
Indicates whether the lappy method (default) or loop method should be used.}
}
\description{
Computes \code{v(S)} for all features subsets \code{S}.
}
