% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_ctree.R
\name{create_ctree}
\alias{create_ctree}
\title{Make all conditional inference trees}
\usage{
create_ctree(
  given_ind,
  x_train,
  mincriterion,
  minsplit,
  minbucket,
  use_partykit = "on_error"
)
}
\arguments{
\item{given_ind}{Numeric value. Indicates which features are conditioned on.}

\item{mincriterion}{Numeric scalar or vector. (default = 0.95)
Either a scalar or vector of length equal to the number of features in the model.
Value is equal to 1 - \eqn{\alpha} where \eqn{\alpha} is the nominal level of the conditional independence tests.
If it is a vector, this indicates which value to use when conditioning on various numbers of features.}

\item{minsplit}{Numeric scalar. (default = 20)
Determines minimum value that the sum of the left and right daughter nodes required for a split.}

\item{minbucket}{Numeric scalar. (default = 7)
Determines the minimum sum of weights in a terminal node required for a split}

\item{use_partykit}{String. In some semi-rare cases \code{partyk::ctree} runs into an error related to the LINPACK
used by R. To get around this problem, one may fall back to using the newer (but slower) \code{partykit::ctree}
function, which is a reimplementation of the same method. Setting this parameter to \code{"on_error"} (default)
falls back to  \code{partykit::ctree}, if \code{party::ctree} fails. Other options are \code{"never"}, which always
uses \code{party::ctree}, and \code{"always"}, which always uses \code{partykit::ctree}. A warning message is
created whenever \code{partykit::ctree} is used.}
}
\value{
List with conditional inference tree and the variables conditioned/not conditioned on.
}
\description{
Make all conditional inference trees
}
\author{
Annabelle Redelmeier, Martin Jullum
}
\keyword{internal}
