% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{create_S_batch}
\alias{create_S_batch}
\title{Create a list of indexes used to run batches}
\usage{
create_S_batch(explainer, n_batches)
}
\arguments{
\item{explainer}{The binary matrix `S` returned from `shapr`.}

\item{n_batches}{Numeric value specifying how many batches `S` should be split into.}
}
\value{
A list of length `n_batches`.
}
\description{
Create a list of indexes used to run batches
}
\keyword{internal}
