% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{default_doc_explain}
\alias{default_doc_explain}
\title{Exported documentation helper function.}
\usage{
default_doc_explain()
}
\arguments{
\item{internal}{Not used.}

\item{index_features}{Positive integer vector. Specifies the indices of combinations to
apply to the present method. \code{NULL} means all combinations. Only used internally.}
}
\description{
Exported documentation helper function.
}
\keyword{internal}
