% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{default_doc_export}
\alias{default_doc_export}
\title{Exported documentation helper function.}
\usage{
default_doc_export(internal, iter, index_features, digits)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}

\item{iter}{Integer.
The iteration number. Only used internally.}

\item{index_features}{Positive integer vector. Specifies the id_coalition to
apply to the present method. \code{NULL} means all coalitions. Only used internally.}

\item{digits}{Integer.
(Maximum) number of digits to be displayed after the decimal point.
Defaults to 2.}
}
\description{
Exported documentation helper function.
}
\keyword{internal}
