% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{explain}
\alias{explain}
\alias{explain.empirical}
\alias{explain.gaussian}
\alias{explain.copula}
\alias{explain.combined}
\title{Explaining the output of machine learning models with more accurately estimated Shapley values}
\usage{
explain(x, explainer, approach, prediction_zero, ...)

\method{explain}{empirical}(x, explainer, approach, prediction_zero,
  index_features, type = "fixed_sigma", fixed_sigma_vec = 0.1,
  AICc_no_samp_per_optim = 1000, AIC_optim_max_eval = 20,
  AIC_optim_startval = 0.1, w_threshold = 0.95)

\method{explain}{gaussian}(x, explainer, approach, prediction_zero,
  mu = NULL, cov_mat = NULL, n_samples = 1000)

\method{explain}{copula}(x, explainer, approach, prediction_zero,
  n_samples = 1000)

\method{explain}{combined}(x, explainer, approach, prediction_zero, ...)
}
\arguments{
\item{x}{1}

\item{explainer}{2}

\item{approach}{3}

\item{prediction_zero}{4}

\item{...}{Soething}
}
\description{
Explaining the output of machine learning models with more accurately estimated Shapley values
}
