% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Main function}
\usage{
explain(
  x_train,
  x_explain,
  model = NULL,
  approach,
  prediction_zero,
  n_combinations = NULL,
  group = NULL,
  n_samples = 1000,
  n_batches = 1,
  seed = 1,
  keep_samp_for_vS = FALSE,
  ...
)
}
\description{
Main function
}
