% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{explain}
\alias{explain}
\alias{explain.empirical}
\alias{explain.gaussian}
\alias{explain.copula}
\alias{explain.combined}
\alias{explain.ctree}
\title{Explain the output of machine learning models with more accurately estimated Shapley values}
\usage{
explain(x, explainer, approach, prediction_zero, ...)

\method{explain}{empirical}(x, explainer, approach, prediction_zero,
  type = "fixed_sigma", fixed_sigma_vec = 0.1,
  AICc_no_samp_per_optim = 1000, AIC_optim_max_eval = 20,
  AIC_optim_startval = 0.1, w_threshold = 0.95, ...)

\method{explain}{gaussian}(x, explainer, approach, prediction_zero,
  mu = NULL, cov_mat = NULL, ...)

\method{explain}{copula}(x, explainer, approach, prediction_zero, ...)

\method{explain}{combined}(x, explainer, approach, prediction_zero,
  mu = NULL, cov_mat = NULL, ...)

\method{explain}{ctree}(x, explainer, approach, prediction_zero,
  comb_indici = NULL, comb_mincriterion = NULL, mincriterion = 0.95,
  minsplit = 20, minbucket = 7, sample = TRUE, ...)
}
\arguments{
\item{x}{A matrix or data.frame. Contains the the features, whose
predictions ought to be explained (test data).}

\item{explainer}{An \code{explainer} object to use for explaining the observations.
See \code{\link{shapr}}.}

\item{approach}{Character vector of length \code{1} or \code{n_features}.
\code{n_features} equals the total number of features in the model. All elements should
either be \code{"gaussian"}, \code{"copula"} or \code{"empirical"}. See details for more information.}

\item{prediction_zero}{The prediction value for unseen data, typically equal to the mean of
the response.}

\item{...}{Additional arguments passed to \code{\link{prepare_data}}}

\item{type}{Character. Should be equal to either \code{"independence"},
\code{"fixed_sigma"}, \code{"AICc_each_k"} or \code{"AICc_full"}.}

\item{fixed_sigma_vec}{Vector or numeric. Only applicable when \code{approach='empirical'} and
\code{type='fixed_sigma'}. The bandwidth to use. Default value \code{0.1}}

\item{AICc_no_samp_per_optim}{Positive integer. Only applicable when
\code{approach='empirical'} and \code{type='AICc_each_k'} or
\code{type='AICc_full'}. Number of samples to consider in AICc optimization.}

\item{AIC_optim_max_eval}{Positive integer. Only applicable when \code{approach='empirical'}
and \code{type='AICc_each_k'} or \code{type='AICc_full'}. Numeric. Maximum value when
optimizing the AICc.}

\item{AIC_optim_startval}{Numeric. Only applicable when \code{approach='empirical'} and
\code{type='AICc_each_k'} or \code{type='AICc_full'}. Starting value when optimizing the AICc.}

\item{w_threshold}{Positive integer between 0 and 1.}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution.
If \code{NULL} the expected values are estimated from the data. Note that this is only used
when \code{approach = "gaussian"}.}

\item{cov_mat}{Numeric matrix. (Optional) Containing the covariance matrix of the data
generating distribution. \code{NULL} means it is estimated from the data if needed
(in the Gaussian approach).}

\item{comb_indici}{Numeric value. (Optional) Contains the splitting point corresponding to where to change the
\code{comb_mincriterion}. Right now, this is only implemented for one splitting value. Could potentially make more splits later.
If \code{NULL}, the \code{mincriterion} is constant for every combination.}

\item{comb_mincriterion}{Numeric vector. (Optional) Contains the different mincriterions to use for each
combination.
If \code{NULL}, the \code{mincriterion} is constant for every combination.}

\item{mincriterion}{equal to 1 - alpha where alpha is the nominal level of the conditional independence tests.
If \code{comb_indici} and \code{comb_mincriterion} are both not \code{NULL}, then \code{mincriterion} can be set
to \code{NULL}. Otherwise, it needs to be filled out.}

\item{minsplit}{is the value that the sum of the left and right daughter nodes need to exceed.}

\item{minbucket}{is equal to the minimum sum of weights in a terminal node.}

\item{sample}{whether to sample from the node or just take all observations}
}
\value{
data.frame. Contains the estimated Shapley values for the test data. Note that
the dimensions of the data.frame equals \code{n x (p+1)}, where \code{n} equals the number
of test observations, and \code{p} equals the total number of features.
}
\description{
TODO: Add a more detailed description
}
\details{
TODO: Add information about approach.
TODO: Some additional details about the returned object
}
\author{
Camilla Lingjaerde
}
