% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_combinations}
\alias{feature_combinations}
\title{Get combinations}
\usage{
feature_combinations(m, exact = TRUE, n_combinations = 200,
  shapley_weight_inf_replacement = 10^6, reduce_dim = TRUE)
}
\value{
data.table
}
\description{
Get combinations
}
\details{
The returned data.table contains the following columns
\describe{
\item{ID}{Positive integer. Unique key for combination}
\item{features}{List of integer vectors}
\item{nfeatures}{Positive integer}
\item{N}{Positive integer}
}
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
