% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_combinations}
\alias{feature_combinations}
\title{Get combinations}
\usage{
feature_combinations(m, exact = TRUE, n_combinations = 200,
  shapley_weight_inf_replacement = 10^6, reduce_dim = TRUE)
}
\arguments{
\item{m}{Integer. Total number of features}

\item{exact}{Logical. If TRUE, uses the full sum in the Shapley formula, if FALSE, uses a
sampling approach to approximate the sum}

\item{shapley_weight_inf_replacement}{Numeric. Indicating which weight to use for the full
conditional and unconditional expectations in kernel SHAPs weighted least squares formulation.}

\item{reduce_dim}{Logical. Indicating whether to reduce the dimension of the weighted least
squares problem by merging identical columns and adjusting their weights.}
}
\value{
data.table
}
\description{
Get combinations
}
\details{
The returned data.table contains the following columns
\describe{
\item{ID}{Positive integer. Unique key for combination}
\item{features}{List of integer vectors}
\item{nfeatures}{Positive integer}
\item{N}{Positive integer}
}
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
