% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{features}
\alias{features}
\alias{features.default}
\alias{features.lm}
\alias{features.glm}
\alias{features.ranger}
\alias{features.gam}
\alias{features.xgb.Booster}
\title{Fetch feature labels from model object}
\usage{
features(x, feature_labels = NULL)

\method{features}{default}(x, feature_labels)

\method{features}{lm}(x)

\method{features}{glm}(x)

\method{features}{ranger}(x)

\method{features}{gam}(x)

\method{features}{xgb.Booster}(x)
}
\arguments{
\item{x}{Numeric matrix or data.frame. Contains the data used for training the model.}

\item{feature_labels}{Character. The label of the features used for training the model. If \code{NULL} it is
assumed that the features are represented by all columns in \code{x}.}
}
\description{
Fetch feature labels from model object
}
\keyword{internal}
