% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{features}
\alias{features}
\alias{features.default}
\alias{features.lm}
\alias{features.glm}
\alias{features.ranger}
\alias{features.gam}
\alias{features.xgb.Booster}
\title{Fetches feature labels from a given model object}
\usage{
features(x, cnms, feature_labels = NULL)

\method{features}{default}(x, cnms, feature_labels = NULL)

\method{features}{lm}(x, cnms, feature_labels = NULL)

\method{features}{glm}(x, cnms, feature_labels = NULL)

\method{features}{ranger}(x, cnms, feature_labels = NULL)

\method{features}{gam}(x, cnms, feature_labels = NULL)

\method{features}{xgb.Booster}(x, cnms, feature_labels = NULL)
}
\arguments{
\item{x}{Model object for the model to be explained.}

\item{cnms}{Character vector. Represents the names of the columns in the data used for training/explaining.}

\item{feature_labels}{Character vector. Represents the labels of the features used for prediction.}
}
\description{
Fetches feature labels from a given model object
}
\keyword{internal}
