% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_explanation.R
\name{finalize_explanation}
\alias{finalize_explanation}
\title{Gather the final output to create the explanation object}
\usage{
finalize_explanation(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
List of reformatted output information extracted from \code{internal}.
}
\description{
Gather the final output to create the explanation object
}
\keyword{internal}
