% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{process_data}
\alias{process_data}
\title{TODO: Add header, parameter info, return and example here}
\usage{
process_data(
  x,
  feature_list,
  name_1 = "model",
  name_2 = "data",
  use_feature_list_as_truth = T
)
}
\description{
TODO: Add header, parameter info, return and example here
}
\keyword{internal}
