% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_round}
\alias{format_round}
\title{Format numbers with rounding}
\usage{
format_round(x, digits = 2L)
}
\arguments{
\item{x}{Numeric vector. The numbers to format.}

\item{digits}{Integer.
(Maximum) number of digits to be displayed after the decimal point.
Defaults to 2.}
}
\value{
Character vector. The formatted numbers.
}
\description{
Format numbers with rounding
}
\keyword{internal}
