% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_shapley_info}
\alias{format_shapley_info}
\title{Internal function to extract the formatted Shapley value table}
\usage{
format_shapley_info(internal, iter, digits = 2L)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}

\item{iter}{Integer.
The iteration number. Only used internally.}

\item{digits}{Integer.
(Maximum) number of digits to be displayed after the decimal point.
Defaults to 2.}
}
\description{
To be used by both \code{\link[=print_iter]{print_iter()}} and \code{\link[=summary.shapr]{summary.shapr()}}
}
\keyword{internal}
