% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{gaussian_transform}
\alias{gaussian_transform}
\title{Transforms a sample to standardized normal distribution}
\usage{
gaussian_transform(x)
}
\arguments{
\item{x}{Numeric vector.The data which should be transformed to a standard normal distribution.}
}
\value{
Numeric vector of length \code{length(x)}
}
\description{
Transforms a sample to standardized normal distribution
}
\author{
Martin Jullum
}
\keyword{internal}
