% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_copula.R
\name{gaussian_transform_separate}
\alias{gaussian_transform_separate}
\title{Transforms new data to standardized normal (dimension 1) based on other data transformations}
\usage{
gaussian_transform_separate(yx, n_y)
}
\arguments{
\item{yx}{Numeric vector. The first \code{n_y} items is the data that is transformed, and last
part is the data with the original transformation.}

\item{n_y}{Positive integer. Number of elements of \code{yx} that belongs to the Gaussian data.}
}
\value{
Vector of back-transformed Gaussian data
}
\description{
Transforms new data to standardized normal (dimension 1) based on other data transformations
}
\author{
Martin Jullum
}
\keyword{internal}
