% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{get_list_approaches}
\alias{get_list_approaches}
\title{Helper function used in \code{\link{explain.combined}}}
\usage{
get_list_approaches(n_features, approach)
}
\arguments{
\item{n_features}{Integer vector. Note that
\code{length(n_features) <= 2^m}, where \code{m} equals the number
of features.}

\item{approach}{Character vector of length \code{m}. All elements should be
either \code{"empirical"}, \code{"gaussian"} or \code{"copula"}.}
}
\value{
List
}
\description{
Helper function used in \code{\link{explain.combined}}
}
\author{
Nikolai Sellereite
}
\keyword{internal}
