% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model_specs}
\alias{get_model_specs}
\alias{get_model_specs.default}
\alias{get_model_specs.lm}
\alias{get_model_specs.glm}
\alias{get_model_specs.gam}
\alias{get_model_specs.ranger}
\alias{get_model_specs.xgb.Booster}
\title{Fetches feature labels from a given model object}
\usage{
get_model_specs(x, feature_labels = NULL)

\method{get_model_specs}{default}(x, feature_labels = NULL)

\method{get_model_specs}{lm}(x, feature_labels = NULL)

\method{get_model_specs}{glm}(x, feature_labels = NULL)

\method{get_model_specs}{gam}(x, feature_labels = NULL)

\method{get_model_specs}{ranger}(x, feature_labels = NULL)

\method{get_model_specs}{xgb.Booster}(x, feature_labels = NULL)
}
\arguments{
\item{x}{Model object for the model to be explained.}

\item{feature_labels}{Character vector. Represents the labels of the features used for prediction.}
}
\description{
Fetches feature labels from a given model object
}
\examples{
# Load example data
data("Boston", package = "MASS")
# Split data into test- and training data
x_train <- head(Boston, -3)
# Fit a linear model
model <- lm(medv ~ lstat + rm + dis + indus, data = x_train)

cnms <- c("lstat", "rm", "dis", "indus")

# Checking that features used by the model corresponds to cnms
get_model_specs(x = model, feature_labels = NULL)
}
\keyword{internal}
