% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model_specs}
\alias{get_model_specs}
\alias{get_model_specs.default}
\alias{get_model_specs.lm}
\alias{get_model_specs.glm}
\alias{get_model_specs.gam}
\alias{get_model_specs.ranger}
\alias{get_model_specs.xgb.Booster}
\title{Fetches feature labels from a given model object}
\usage{
get_model_specs(model, feature_labels = NULL)

\method{get_model_specs}{default}(model, feature_labels = NULL)

\method{get_model_specs}{lm}(model, feature_labels = NULL)

\method{get_model_specs}{glm}(model, feature_labels = NULL)

\method{get_model_specs}{gam}(model, feature_labels = NULL)

\method{get_model_specs}{ranger}(model, feature_labels = NULL)

\method{get_model_specs}{xgb.Booster}(model, feature_labels = NULL)
}
\arguments{
\item{feature_labels}{Character vector. Represents the labels of the features used for prediction.}
}
\description{
Fetches feature labels from a given model object
}
\examples{
 # Load example data
data("Boston", package = "MASS")
# Split data into test- and training data
x_train <- data.table::as.data.table(head(Boston))
x_train[,rad:=as.factor(rad)]
model <- lm(medv ~ lstat + rm + rad + indus, data = x_train)

get_model_specs(model, feature_labels = NULL)
}
\keyword{internal}
