% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict_model.R
\name{get_predict_model}
\alias{get_predict_model}
\title{Get predict_model function}
\usage{
get_predict_model(predict_model, model)
}
\arguments{
\item{predict_model}{Function.
The prediction function used when \code{model} is not natively supported.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{model}{Objects.
The model object that ought to be explained.
See the documentation of \code{\link[=explain]{explain()}} for details.}
}
\description{
Get predict_model function
}
\keyword{internal}
