% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_separate.R
\name{get_regression_tune}
\alias{get_regression_tune}
\title{Get if model is to be tuned}
\usage{
get_regression_tune(regression_model, regression_tune_values)
}
\arguments{
\item{regression_model}{A tidymodels object of class \code{model_specs}. Default is a linear regression model
(\code{\link[parsnip:linear_reg]{parsnip::linear_reg()}}). See \href{https://www.tidymodels.org/find/parsnip/}{tidymodels} for all possible models.}

\item{regression_tune_values}{Either NULL (default) or a data.frame/data.table/tibble containing the possible
hyper-parameter value combinations to try. The column names must match the names of the tuneable parameters specified
in \code{regression_model}.}
}
\description{
That is, if the regression model contains hyper-parameters we are to tune using cross validation.
See \href{https://www.tidymodels.org/find/parsnip/#model-args}{tidymodels} for default model hyper-parameters.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
