% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_supported_models}
\alias{get_supported_models}
\title{Provides a data.table with the supported models}
\usage{
get_supported_models()
}
\description{
Provides a data.table with the supported models
}
\keyword{internal}
