% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{inv_gaussian_transform}
\alias{inv_gaussian_transform}
\title{Transforms new data to standardized normal (dimension 1) based on other data transformations}
\usage{
inv_gaussian_transform(zx, n_z, type = 7)
}
\arguments{
\item{zx}{Vector where the first part is the Gaussian data, and last part is
the data with the original transformation}

\item{type}{The quantile type used when back-transforming. 7 (default) is the default in stats::quantile().}

\item{n_y}{How many elements of \code{yx} that belongs to the y-part (new data)}
}
\value{
Vector of transformed new data
}
\description{
Handled in this way in order to allow using the apply function over this function
}
\author{
Martin Jullum
}
