% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{inv_gaussian_transform}
\alias{inv_gaussian_transform}
\title{Transforms new data to a standardized normal distribution}
\usage{
inv_gaussian_transform(zx, n_z)
}
\arguments{
\item{zx}{Numeric vector. The first \code{n_z} items are the Gaussian data, and the last part is
the data with the original transformation.}

\item{n_z}{Positive integer. Number of elements of \code{zx} that belongs to new data.}
}
\value{
Numeric vector of length \code{n_z}
}
\description{
Transforms new data to a standardized normal distribution
}
\author{
Martin Jullum
}
\keyword{internal}
