% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{make_dummies}
\alias{make_dummies}
\title{Initiate the making of dummy variables}
\usage{
make_dummies(data, ...)
}
\arguments{
\item{data}{data.table or data.frame. Includes all the features (both factors and possibly others).}

\item{...}{Currently not used.}
}
\value{
A list that contains the following entries:
\describe{
\item{features}{Vector. Contains the names of all the features in \code{data}.}
\item{factor_features}{Vector. Contains the names of all the factors in \code{data}.}
\item{factor_features}{List. Contains each factor and its vector of levels.}
\item{contrasts_list}{List. Contains all the contrasts of the factors.}
}
}
\description{
Initiate the making of dummy variables
}
\examples{
data("Boston", package = "MASS")
x_var <- c("lstat", "chas", "rad", "indus")
y_var <- "medv"
# convert to factors
Boston$rad = as.factor(Boston$rad)
Boston$chas = as.factor(Boston$chas)
x_train <- Boston[-1:-6, x_var]
y_train <- Boston[-1:-6, y_var]
x_test <- Boston[1:6, x_var]

dummyfunc <- make_dummies(data = rbind(x_train, x_test))

}
\author{
Annabelle Redelmeier
}
