% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{make_dummies}
\alias{make_dummies}
\title{Initiate the making of dummy variables}
\usage{
make_dummies(traindata, testdata)
}
\arguments{
\item{traindata}{data.table or data.frame.}

\item{testdata}{data.table or data.frame. New data that has the same
feature names, types, and levels as \code{traindata}.}
}
\value{
A list that contains the following entries:
\describe{
\item{feature_spec}{List. Output from \code{check_features}}
\item{train_dummies}{A data.frame containing all of the factors in \code{traindata} as
one-hot encoded variables.}
\item{test_dummies}{A data.frame containing all of the factors in \code{testdata} as
one-hot encoded variables.}
\item{traindata_new}{Original traindata with correct column ordering and factor levels. To be passed to
shapr}
\item{testdata_new}{Original testdata with correct column ordering and factor levels. To be passed to
\code{\link[shapr:explain]{explain}.}}
}
}
\description{
Initiate the making of dummy variables
}
\examples{
# Load example data
data("airquality")
airquality <- airquality[complete.cases(airquality), ]
x_var <- c("Solar.R", "Wind", "Temp", "Month")
y_var <- "Ozone"
x_train <- as.data.frame(airquality[100:110, x_var])
y_train <- airquality[100:110, y_var]
x_explain <- as.data.frame(airquality[1:4, x_var])

# convert to factors for illustational purpose
x_train$Temp <- factor(round(x_train$Temp, -1))
x_explain$Temp <- factor(round(x_explain$Temp, -1), levels = levels(x_train$Temp))

dummylist <- make_dummies(traindata = x_train, testdata = x_explain)
}
\author{
Annabelle Redelmeier, Martin Jullum
}
