% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{num_str}
\alias{num_str}
\title{Convert a character to a numeric class}
\usage{
num_str(x)
}
\arguments{
\item{x}{Character. A single character that represents a number, or a vector of characters.}
}
\value{
A numeric class object with the value of the string.
}
\description{
To be used in cli calls like \code{cli::cli_text("{.val {shapr:::num_str('12.10')}}")} to format character strings
that typically represent a number as if it were numeric. May also be used with strings not representing a number.
}
\keyword{internal}
