% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{observation_impute}
\alias{observation_impute}
\title{Get predictions}
\usage{
observation_impute(w_kernel, s, xtrain, xtest, w_threshold = 0.7,
  no_samp_mc = 1000)
}
\arguments{
\item{w_kernel}{Array. Contains all nonscaled weights between training and testing observations
for all combinations.}

\item{s}{Matrix}

\item{xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{xtest}{Matrix, data.frame or data.table with the features, whose predictions ought
to be explained (test data)}

\item{no_samp_mc}{Positive integer. Indicating the maximum number of samples to use in the
Monte Carlo integration for every conditional expectation (previously called \code{n_threshold})}
}
\value{
List
}
\description{
Get predictions
}
\author{
Nikolai Sellereite
}
