% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_empirical.R
\name{observation_impute}
\alias{observation_impute}
\title{Generate permutations of training data using test observations}
\usage{
observation_impute(
  W_kernel,
  S,
  x_train,
  x_explain,
  empirical.eta = 0.7,
  n_samples = 1000
)
}
\arguments{
\item{W_kernel}{Numeric matrix. Contains all nonscaled weights between training and test
observations for all feature combinations. The dimension equals \verb{n_train x m}.}

\item{S}{Integer matrix of dimension \verb{n_combinations x m}, where \code{n_combinations}
and \code{m} equals the total number of sampled/non-sampled feature combinations and
the total number of unique features, respectively. Note that \code{m = ncol(x_train)}.}

\item{x_train}{Numeric matrix}

\item{x_explain}{Numeric matrix}

\item{n_samples}{Positive integer.
Indicating the maximum number of samples to use in the
Monte Carlo integration for every conditional expectation. See also details.}
}
\value{
data.table
}
\description{
Generate permutations of training data using test observations
}
\author{
Nikolai Sellereite
}
\keyword{internal}
