% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{observation_impute_cpp}
\alias{observation_impute_cpp}
\title{Get imputed data}
\usage{
observation_impute_cpp(id, Comb, xtrain, xtest, s)
}
\arguments{
\item{id}{Positive integer vector}

\item{Comb}{Positive integer vector}

\item{xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{xtest}{Matrix, data.frame or data.table with the features, whose predictions ought to
be explained (test data)}

\item{s}{Matrix}
}
\value{
Array of three dimensions
}
\description{
Get imputed data
}
\author{
Nikolai Sellereite
}
