% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{paired_sampler}
\alias{paired_sampler}
\title{Sampling Paired Observations}
\usage{
paired_sampler(vaeac_dataset_object, shuffle = FALSE)
}
\arguments{
\item{vaeac_dataset_object}{A \code{\link[=vaeac_dataset]{vaeac_dataset()}} object containing the data.}

\item{shuffle}{Boolean. If \code{TRUE}, then the data is shuffled. If \code{FALSE},
then the data is returned in chronological order.}
}
\description{
A sampler used to samples the batches where each instances is sampled twice
}
\details{
A sampler object that allows for paired sampling by always including each observation from the
\code{\link[=vaeac_dataset]{vaeac_dataset()}} twice. A \code{\link[torch:sampler]{torch::sampler()}} object can be used with \code{\link[torch:dataloader]{torch::dataloader()}} when creating
batches from a torch dataset \code{\link[torch:dataset]{torch::dataset()}}. See \url{https://rdrr.io/cran/torch/src/R/utils-data-sampler.R} for
more information. This function does not use batch iterators, which might increase the speed.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
