% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.shapr}
\alias{plot.shapr}
\title{Plot of the Shapley value explanations}
\usage{
\method{plot}{shapr}(
  x,
  plot_type = "bar",
  digits = 3,
  plot_phi0 = TRUE,
  index_x_explain = NULL,
  top_k_features = NULL,
  col = c("#00BA38", "#F8766D"),
  plot_order = "largest_first",
  features_to_plot = NULL,
  histogram = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{shapr} object.
The output from [explain()].}

\item{plot_type}{Character.
Specifies the type of plot to produce.
\code{"bar"} (the default) gives a regular horizontal bar plot of the Shapley value magnitudes.
\code{"waterfall"} gives a waterfall plot indicating the changes in the prediction score due to each features
contribution (their Shapley values).}

\item{digits}{Integer.
Number of significant digits to use in the feature description}

\item{plot_phi0}{Logical.
Whether to include \code{phi0} in the plot.}

\item{index_x_explain}{Integer vector.
Which of the test observations to plot. E.g. if you have
explained 10 observations using \code{\link{explain}}, you can generate a plot for the first 5
observations by setting \code{index_x_explain = 1:5}.}

\item{top_k_features}{Integer.
How many features to include in the plot. E.g. if you have 15
features in your model you can plot the 5 most important features, for each explanation, by setting
\code{top_k_features = 1:5}.}

\item{col}{Character vector (or length 2).
The color codes (hex codes or other names understood by [ggplot2::ggplot()]) for positive and negative
Shapley values, respectively.}

\item{plot_order}{Character.
Specifies what order to plot the features with respect to the magnitude of the shapley values.
 \code{"largest_first"} (the default) plots the features ordered from largest to smallest absolute Shapley value.
 \code{"smallest_first"} plots the features ordered from smallest to largest absolute Shapley value.
 \code{"original"} plots the features in the original order of the data table.}

\item{features_to_plot}{Integer or character vector.
Specifies what features to include in (scatter) plot. Can be a numerical vector indicating feature index, or a character
vector, indicating the name(s) of the feature(s) to plot.}

\item{histogram}{Logical.
Whether to include a histogram indicating the distribution of the data when making the scatter plot. Note that the
bins are scaled so that when all the bins are stacked they fit the span of the y-axis of the plot.}

\item{...}{Currently not used.}
}
\value{
ggplot object with plots of the Shapley value explanations
}
\description{
Plots the individual prediction explanations.
}
\details{
See the examples below, or \code{vignette("understanding_shapr", package = "shapr")} for an examples of
how you should use the function.
}
\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
  #' # Load example data
  data("Boston", package = "MASS")

  # Split data into test- and training data
  x_train <- head(Boston, -3)
  x_explain <- tail(Boston, 3)

  # Fit a linear model
  model <- lm(medv ~ lstat + rm + dis + indus, data = x_train)

  # Define unconditional expectation
  p <- mean(x_train$medv)

  # Explain predictions
  explanation <- explain(x_train,x_explain,model,approach="empirical",prediction_zero=p)

  # Plot the explantion (this function)
  if (requireNamespace("ggplot2", quietly = TRUE)) {
   plot(explanation) # Regular bar plot

   plot(explanation, plot_type = "waterfall") # Waterfall plot
  }
}
}
\author{
Martin Jullum, Vilde Ung
}
