% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.shapr}
\alias{plot.shapr}
\title{Plot of the Shapley value explanations}
\usage{
\method{plot}{shapr}(x, digits = 3, plot_phi0 = TRUE,
  index_x_test = NULL, top_k_features = NULL)
}
\arguments{
\item{digits}{Integer. Number of significant digits to use in the feature description}

\item{plot_phi0}{Logical. Whether to include phi0 in the plot}

\item{top_k_features}{Integer. How many features to include in the plot}

\item{explanation}{The output from compute_kshap}

\item{Xtest}{Data.table. Contains the covariates of the test data}

\item{plot_which_Xtest}{Integer vector. Which of the test observations to plot}
}
\value{
ggplot object with plots of the Shapley value explanations
}
\description{
Plots the individual prediction explanations. Uses facet_wrap of ggplot
}
\author{
Martin Jullum
}
