% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kshap}
\alias{plot_kshap}
\title{Plot of the Shapley value explanations}
\usage{
plot_kshap(explanation, l, no_desc_digits = 3, plot_phi0 = T,
  plot_which_xtest = 1:nrow(l$xtest), top_k_features = ncol(l$xtest) +
  1)
}
\arguments{
\item{explanation}{The output from compute_kshap}

\item{l}{List. The output from the \code{prepare_kshap} function}

\item{no_desc_digits}{Integer. Number of significant digits to use in the feature description}

\item{plot_phi0}{Logical. Whether to include phi0 in the plot}

\item{plot_which_xtest}{Integer vector. Which of the test observations to plot}

\item{top_k_features}{Integer. How many features to include in the plot}
}
\value{
ggplot object with plots of the Shapley value explanations
}
\description{
Plots the individual prediction explanations. Uses facet_wrap of ggplot
}
\author{
Martin Jullum
}
