% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/model_glm.R, R/model_lm.R,
%   R/model_mgcv_gam.R, R/model_ranger.R, R/model_xgboost.R
\name{predict_model}
\alias{predict_model}
\alias{predict_model.default}
\alias{predict_model.glm}
\alias{predict_model.lm}
\alias{predict_model.gam}
\alias{predict_model.ranger}
\alias{predict_model.xgb.Booster}
\title{Generate predictions for different model classes}
\usage{
predict_model(x, newdata)

\method{predict_model}{default}(x, newdata)

\method{predict_model}{glm}(x, newdata)

\method{predict_model}{lm}(x, newdata)

\method{predict_model}{gam}(x, newdata)

\method{predict_model}{ranger}(x, newdata)

\method{predict_model}{xgb.Booster}(x, newdata)
}
\arguments{
\item{x}{Model object for the model to be explained.}

\item{newdata}{A data frame (or matrix) in which to look for variables with which to predict.}
}
\value{
Numeric
}
\description{
Performs prediction of response \code{\link[stats]{lm}}, \code{\link[stats]{glm}},
\code{\link[ranger]{ranger}},  \code{\link[mgcv:gam]{mgcv::gam}} and
\code{\link[xgboost:xgb.train]{xgboost::xgb.train}} with binary or continuous
response. See details for more information.
}
\details{
The following models are currently supported:
\itemize{
\item \code{\link[stats:lm]{stats::lm}}
\item \code{\link[stats:glm]{stats::glm}}
\item \code{\link[ranger:ranger]{ranger::ranger}}
\item \code{\link[mgcv:gam]{mgcv::gam}}
\item \code{\link[xgboost:xgb.train]{xgboost::xgb.train}}
}

The returned object \code{p} always satisfies the following properties:
\itemize{
\item \code{is.atomic(p)} equals \code{TRUE}
\item \code{is.double(p)} equals \code{TRUE}
}

If you have a binary classification model we'll always return the probability prediction
for a single class.

For more details on how to explain other types of models (i.e. custom models), see the Advanced usage section
of the vignette: \cr
From R: \code{vignette("understanding_shapr", package = "shapr")}  \cr
Web: \url{https://norskregnesentral.github.io/shapr/articles/understanding_shapr.html#explain-custom-models}
}
\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
  # Load example data
  data("Boston", package = "MASS")
  # Split data into test- and training data
  x_train <- head(Boston, -3)
  x_explain <- tail(Boston, 3)
  # Fit a linear model
  model <- lm(medv ~ lstat + rm + dis + indus, data = x_train)

  # Predicting for a model with a standardized format
  predict_model(x = model, newdata = x_explain)
}
}
\author{
Martin Jullum
}
\keyword{internal}
