% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_model}
\alias{predict_model}
\title{Predict on vector form}
\usage{
predict_model(x, newdata)
}
\arguments{
\item{data}{data.table or data.frame with data to perform prediction}
}
\value{
Vector of predictions
}
\description{
Performs prediction of response \code{\link[stats]{lm}}, \code{\link[stats]{glm}},
\code{\link[ranger]{ranger}} and \code{\link[xgboost]{xgboost}} with binary or continuous response.
Output the prediction on vector form. May let the user provide this function to handle any
prediction model in the future.
}
\author{
Martin Jullum
}
