% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{prediction}
\alias{prediction}
\title{Calculate shapley weights}
\usage{
prediction(dt, prediction_zero, explainer)
}
\arguments{
\item{dt}{data.table}

\item{prediction_zero}{Positive integer between 0 & 1.}

\item{explainer}{An object of class \code{explainer}}
}
\description{
TODO: Write a better description
}
\details{
TODO: Write details about how this is done (reference to paper)
}
\examples{
TODO: Add simple examples

}
\author{
Nikolai Sellereite
}
