% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Note that this function is deprecated, but we'll keep it for a week
to check that results are stable.}
\usage{
predictions(model, D, h_optim_vec, kernel_metric, S, Xtrain, Xtest,
  w_threshold = 0.7, noSamp_MC = 1000, verbose = FALSE,
  cond_approach_list, feature_list, pred_zero, mu, Sigma,
  mu_Gauss_trans = mu_Gauss_trans,
  Sigma_Gauss_trans = Sigma_Gauss_trans, Xtest_Gauss_trans)
}
\description{
TODO: Delete this function from the codebase
}
\keyword{internal}
