% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Calculate Shapley values}
\usage{
prepare_and_predict(
  explainer,
  n_batches,
  prediction_zero,
  only_return_dt_mat,
  ...
)
}
\arguments{
\item{explainer}{An \code{explainer} object to use for explaining the observations.
See \code{\link{shapr}}.}

\item{n_batches}{Positive integer. Specifies how many iterations that should be used to calculate the
contribution function for each test observation. The default value is 1. Setting it to a higher number can
alleviate memory problems for models with many features.}

\item{prediction_zero}{Numeric. The prediction value for unseen data, typically equal to the mean of
the response.}

\item{only_return_dt_mat}{Logical. Used internally in `explain.combined`. If `TRUE` the `data.table` from
\code{\link{prediction}} is returned, else an object of class `shapr`.}

\item{...}{Additional arguments passed to \code{\link{prepare_data}}}
}
\value{
A list. See \code{\link{explain}} for more information.
}
\description{
Sample covariate values, predict and calculate Shapley values. The sampling and prediction can be done in batches
if `n_batches` is greater than 1.
}
\keyword{internal}
