% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Prepare and predict in batches}
\usage{
prepare_and_predict(explainer, n_batches, ...)
}
\arguments{
\item{explainer}{An \code{explainer} object.}

\item{n_batches}{Integer specifying number of batches.}

\item{...}{Passed to \code{\link{prepare_data}}.}
}
\value{
A list. See \code{\link{explain}} for more information.
}
\description{
Prepare and predict in batches
}
\keyword{internal}
