% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Calculate Shapley values}
\usage{
prepare_and_predict(explainer, n_batches, prediction_zero, ...)
}
\arguments{
\item{explainer}{An \code{explainer} object to use for explaining the observations.
See \code{\link{shapr}}.}

\item{n_batches}{Positive integer.
Specifies how many batches the total number of feature combinations should be split into when calculating the
contribution function for each test observation.
The default value is 1.
Increasing the number of batches may significantly reduce the RAM allocation for models with many features.
This typically comes with a small increase in computation time.}

\item{prediction_zero}{Numeric. The prediction value for unseen data, typically equal to the mean of
the response.}

\item{...}{Arguments passed to \code{\link{prepare_data}} with exception of \code{only_return_contrib_dt},
which is only passed to explain. If \code{TRUE} the
\code{data.table} from \code{\link{prediction}} is returned, else an object of class \code{shapr}.
Each column (except for \code{row_id}) correspond to the vector \code{v_D} in Equation 7 in the reference.
The Shapley values can be calculated by \code{t(explainer$W \%*\% dt_contrib[, -"row_id"]))}}
}
\value{
A list. See \code{\link{explain}} for more information.
}
\description{
Sample covariate values, predict and calculate Shapley values. The sampling and prediction can be done in batches
if \code{n_batches} is greater than 1.
}
\keyword{internal}
