% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{prepare_data}
\alias{prepare_data}
\alias{prepare_data.empirical}
\alias{prepare_data.gaussian}
\alias{prepare_data.copula}
\alias{prepare_data.ctree}
\title{Generate data used for predictions}
\usage{
prepare_data(x, ...)

\method{prepare_data}{empirical}(x, seed = 1, n_samples = 1000, index_features = NULL, ...)

\method{prepare_data}{gaussian}(x, seed = 1, n_samples = 1000, index_features = NULL, ...)

<<<<<<< HEAD
\method{prepare_data}{copula}(x, x_test_gaussian = 1, seed = 1,
  n_samples = 1000, index_features = NULL, ...)

\method{prepare_data}{ctree}(x, seed = 1, n_samples = 1000,
  index_features = NULL, mc_cores = 1,
  mc_cores_simulateAllTrees = mc_cores,
  mc_cores_sample_ctree = mc_cores, ...)
=======
\method{prepare_data}{copula}(
  x,
  x_test_gaussian = 1,
  seed = 1,
  n_samples = 1000,
  index_features = NULL,
  ...
)
>>>>>>> master
}
\arguments{
\item{x}{Explainer object. See \code{\link{explain}} for more information.}

\item{...}{Currently not used.}

\item{seed}{Positive integer. If \code{NULL} the seed will be inherited from the calling environment.}

\item{n_samples}{Positive integer. Indicating the maximum number of samples to use in the
Monte Carlo integration for every conditional expectation.}

<<<<<<< HEAD
\item{index_features}{Positive integer vector. Only used internally.}

\item{mc_cores}{Integer. Only for class \code{ctree}. The number of cores to use in paralellization of the
tree building and tree prediction Defaults to 1. Uses parallel::mclapply which relies on forking, i.e. does not
work on Windows systems.}

\item{mc_cores_simulateAllTrees}{Integer. As for \code{mc_cores}, but specific for the tree building function
#' Defaults to \code{mc_cores}.}

\item{mc_cores_simulateAllTrees}{Integer. As for \code{mc_cores}, but specific for the tree building prediction function.
Defaults to \code{mc_cores}.}
=======
\item{index_features}{Positive integer vector. Specifies the indices of combinations to apply to the present method.
\code{NULL} means all combinations. Only used internally.}

\item{x_test_gaussian}{Matrix. Test data quantile-transformed to standard Gaussian variables. Only applicable if
\code{approach = "empirical"}.}
>>>>>>> master
}
\description{
Generate data used for predictions
}
