% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach.R, R/approach_categorical.R,
%   R/approach_copula.R, R/approach_ctree.R, R/approach_empirical.R,
%   R/approach_gaussian.R, R/approach_independence.R
\name{prepare_data}
\alias{prepare_data}
\alias{prepare_data.categorical}
\alias{prepare_data.copula}
\alias{prepare_data.ctree}
\alias{prepare_data.empirical}
\alias{prepare_data.gaussian}
\alias{prepare_data.independence}
\title{Generate data used for predictions}
\usage{
prepare_data(internal, ...)

\method{prepare_data}{categorical}(internal, index_features = NULL, ...)

\method{prepare_data}{copula}(internal, index_features = NULL, ...)

\method{prepare_data}{ctree}(internal, index_features = NULL, ...)

\method{prepare_data}{empirical}(internal, index_features = NULL, ...)

\method{prepare_data}{gaussian}(internal, index_features = NULL, ...)

\method{prepare_data}{independence}(internal, index_features = NULL, ...)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}

\item{...}{Currently not used.}

\item{index_features}{Positive integer vector. Specifies the indices of combinations to apply to the present method.
\code{NULL} means all combinations. Only used internally.}

\item{x}{Explainer object. See \code{\link[=explain]{explain()}} for more information.}

\item{seed}{Positive integer. If \code{NULL} the seed will be inherited from the calling environment.}
}
\value{
A data.table containing simulated data passed to prediction().
}
\description{
Generate data used for predictions
}
\keyword{internal}
