% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{prepare_data}
\alias{prepare_data}
\alias{prepare_data.empirical}
\alias{prepare_data.gaussian}
\alias{prepare_data.copula}
\title{Generate data used for predictions}
\usage{
prepare_data(x, ...)

\method{prepare_data}{empirical}(x, seed = 1, n_samples = 1000,
  index_features = NULL, ...)

\method{prepare_data}{gaussian}(x, seed = 1, n_samples = 1000,
  index_features = NULL, ...)

\method{prepare_data}{copula}(x, x_test = 1, seed = 1,
  n_samples = 1000, index_features = NULL, ...)
}
\description{
Generate data used for predictions
}
