% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach.R, R/approach_copula.R,
%   R/approach_ctree.R, R/approach_empirical.R, R/approach_gaussian.R,
%   R/approach_independence.R
\name{prepare_data}
\alias{prepare_data}
\alias{prepare_data.copula}
\alias{prepare_data.ctree}
\alias{prepare_data.empirical}
\alias{prepare_data.gaussian}
\alias{prepare_data.independence}
\title{Generate data used for predictions}
\usage{
prepare_data(internal, ...)

\method{prepare_data}{copula}(internal, index_features = NULL, ...)

\method{prepare_data}{ctree}(internal, index_features = NULL, ...)

\method{prepare_data}{empirical}(internal, index_features = NULL, ...)

\method{prepare_data}{gaussian}(internal, index_features = NULL, ...)

\method{prepare_data}{independence}(internal, index_features = NULL, ...)
}
\arguments{
\item{...}{Currently not used.}

\item{index_features}{List. Default is NULL but if either various methods are being used or various mincriterion are
used for different numbers of conditioned features, this will be a list with the features to pass.}

\item{x}{Explainer object. See \code{\link[=explain]{explain()}} for more information.}

\item{seed}{Positive integer. If \code{NULL} the seed will be inherited from the calling environment.}

\item{mc_cores}{Integer. Only for class \code{ctree} currently. The number of cores to use in paralellization of the
tree building (\code{create_ctree}) and tree sampling (\code{sample_ctree}). Defaults to 1. Note: Uses
parallel::mclapply which relies on forking, i.e. uses only 1 core on Windows systems.}

\item{mc_cores_create_ctree}{Integer. Same as \code{mc_cores}, but specific for the tree building function
#' Defaults to \code{mc_cores}.}

\item{mc_cores_sample_ctree}{Integer. Same as \code{mc_cores}, but specific for the tree building prediction
function.
Defaults to \code{mc_cores}.}
}
\value{
A data.table containing simulated data passed to prediction().
}
\description{
Generate data used for predictions
}
\keyword{internal}
