% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach.R, R/approach_categorical.R,
%   R/approach_copula.R, R/approach_ctree.R, R/approach_empirical.R,
%   R/approach_gaussian.R, R/approach_independence.R, R/approach_timeseries.R
\name{prepare_data}
\alias{prepare_data}
\alias{prepare_data.categorical}
\alias{prepare_data.copula}
\alias{prepare_data.ctree}
\alias{prepare_data.empirical}
\alias{prepare_data.gaussian}
\alias{prepare_data.independence}
\alias{prepare_data.timeseries}
\title{Generate data used for predictions and Monte Carlo integration}
\usage{
prepare_data(internal, index_features = NULL, ...)

\method{prepare_data}{categorical}(internal, index_features = NULL, ...)

\method{prepare_data}{copula}(internal, index_features = NULL, ...)

\method{prepare_data}{ctree}(internal, index_features = NULL, ...)

\method{prepare_data}{empirical}(internal, index_features = NULL, ...)

\method{prepare_data}{gaussian}(internal, index_features = NULL, ...)

\method{prepare_data}{independence}(internal, index_features = NULL, ...)

\method{prepare_data}{timeseries}(internal, index_features = NULL, ...)
}
\arguments{
\item{internal}{Not used.}

\item{index_features}{Positive integer vector. Specifies the indices of combinations to
apply to the present method. \code{NULL} means all combinations. Only used internally.}

\item{...}{Currently not used.}
}
\value{
A data.table containing simulated data used to estimate
the contribution function by Monte Carlo integration.
}
\description{
Generate data used for predictions and Monte Carlo integration
}
\keyword{internal}
