% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach.R, R/approach_categorical.R,
%   R/approach_copula.R, R/approach_ctree.R, R/approach_empirical.R,
%   R/approach_gaussian.R, R/approach_independence.R,
%   R/approach_regression_separate.R, R/approach_regression_surrogate.R,
%   R/approach_timeseries.R, R/approach_vaeac.R
\name{prepare_data}
\alias{prepare_data}
\alias{prepare_data.categorical}
\alias{prepare_data.copula}
\alias{prepare_data.ctree}
\alias{prepare_data.empirical}
\alias{prepare_data.gaussian}
\alias{prepare_data.independence}
\alias{prepare_data.regression_separate}
\alias{prepare_data.regression_surrogate}
\alias{prepare_data.timeseries}
\alias{prepare_data.vaeac}
\title{Generate data used for predictions and Monte Carlo integration}
\usage{
prepare_data(internal, index_features = NULL, ...)

\method{prepare_data}{categorical}(internal, index_features = NULL, ...)

\method{prepare_data}{copula}(internal, index_features, ...)

\method{prepare_data}{ctree}(internal, index_features = NULL, ...)

\method{prepare_data}{empirical}(internal, index_features = NULL, ...)

\method{prepare_data}{gaussian}(internal, index_features, ...)

\method{prepare_data}{independence}(internal, index_features = NULL, ...)

\method{prepare_data}{regression_separate}(internal, index_features = NULL, ...)

\method{prepare_data}{regression_surrogate}(internal, index_features = NULL, ...)

\method{prepare_data}{timeseries}(internal, index_features = NULL, ...)

\method{prepare_data}{vaeac}(internal, index_features = NULL, ...)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}

\item{index_features}{Positive integer vector. Specifies the id_coalition to
apply to the present method. \code{NULL} means all coalitions. Only used internally.}

\item{...}{Currently not used.}
}
\value{
A data.table containing simulated data used to estimate
the contribution function by Monte Carlo integration.
}
\description{
Generate data used for predictions and Monte Carlo integration
}
\author{
Martin Jullum

Annabelle Redelmeier and Lars Henry Berge Olsen

Lars Henry Berge Olsen

Martin Jullum,
}
\keyword{internal}
