% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{prepare_kshap}
\alias{prepare_kshap}
\title{Get Shapley weights for test data}
\usage{
prepare_kshap(x, x_test, n_combinations = NULL)
}
\arguments{
\item{compute_distances_for_no_var}{If equal to \code{NULL} no distances are computed}
}
\value{
Matrix
}
\description{
Get Shapley weights for test data
}
\author{
Nikolai Sellereite
}
