% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Process (check and update) data according to specified feature list}
\usage{
preprocess_data(
  x,
  feature_list,
  name_1 = "model",
  name_2 = "data",
  use_feature_list_as_truth = T
)
}
\arguments{
\item{x}{matrix, data.frame or data.table. The data to check input for and update
according to the specification in \code{feature_list}.}

\item{feature_list}{List. Output from running \code{get_data_specs} or \code{get_model_specs}}

\item{name_1, name_2}{Character. Passed directly on to \code{check_features}}

\item{use_feature_list_as_truth}{Logical. Passed directly on to \code{use_1_as_truth} in \code{check_features}.}
}
\value{
Checked and updated data \code{x} in data.table format.
}
\description{
Process (check and update) data according to specified feature list
}
\keyword{internal}
