% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.shapr}
\alias{print.shapr}
\title{Print method for shapr objects}
\usage{
\method{print}{shapr}(
  x,
  what = c("shapley_est", "shapley_sd", "MSEv", "MSEv_explicand", "MSEv_coalition",
    "timing_summary"),
  digits = 4,
  nsmall = max(0, digits - 2),
  ...
)
}
\arguments{
\item{x}{A shapr object}

\item{what}{Character. Which component to print.
Options are "shapley_est", "shapley_sd", "MSEv", "MSEv_explicand", "MSEv_coalition" & "timing_summary.
Defaults to "shapley_est".
Only one component can be printed at a time.
See the details section of \code{\link[=get_results]{get_results()}} for details about each component.}

\item{digits}{Integer.
Number of significant digits to display.
Defaults to 4.}

\item{nsmall}{Integer.
Minimum digits to the right of the decimal point.
Defaults to 2 less than \code{digits}.}

\item{...}{Further arguments passed to \code{\link[data.table:print.data.table]{data.table::print.data.table()}}.}
}
\value{
The object is returned invisibly after printing selected output.
}
\description{
Print method for shapr objects
}
