% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{print_iter}
\alias{print_iter}
\title{Print Iterative Information}
\usage{
print_iter(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
No return value (but prints iterative information)
}
\description{
Print Iterative Information
}
\keyword{internal}
