% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_surrogate.R
\name{regression.check_sur_n_comb}
\alias{regression.check_sur_n_comb}
\title{Check the \code{regression.surrogate_n_comb} parameter}
\usage{
regression.check_sur_n_comb(regression.surrogate_n_comb, used_n_combinations)
}
\arguments{
\item{regression.surrogate_n_comb}{Integer (default is \code{internal$parameters$used_n_combinations}) specifying the
number of unique combinations/coalitions to apply to each training observation. Maximum allowed value is
"\code{internal$parameters$used_n_combinations} - 2". By default, we use all coalitions, but this can take a lot of memory
in larger dimensions. Note that by "all", we mean all coalitions chosen by \code{shapr} to be used. This will be all
\eqn{2^{n_{\text{features}}}} coalitions (minus empty and grand coalition) if \code{shapr} is in the exact mode. If the
user sets a lower value than \code{internal$parameters$used_n_combinations}, then we sample this amount of unique
coalitions separately for each training observations. That is, on average, all coalitions should be equally trained.}

\item{used_n_combinations}{Integer. The number of used combinations (including the empty and grand coalitions).}
}
\description{
Check that \code{regression.surrogate_n_comb} is either NULL or a valid integer.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
