% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_regression_surrogate.R
\name{regression.check_sur_n_comb}
\alias{regression.check_sur_n_comb}
\title{Check the \code{regression.surrogate_n_comb} parameter}
\usage{
regression.check_sur_n_comb(regression.surrogate_n_comb, n_coalitions)
}
\arguments{
\item{regression.surrogate_n_comb}{Positive integer.
Specifies the number of unique coalitions to apply to each training observation.
The default is the number of sampled coalitions in the present iteration.
Any integer between 1 and the default is allowed.
Larger values requires more memory, but may improve the surrogate model.
If the user sets a value lower than the maximum, we sample this amount of unique coalitions
separately for each training observations.
That is, on average, all coalitions should be equally trained.}

\item{n_coalitions}{Integer. The number of used coalitions (including the empty and grand coalition).}
}
\description{
Check that \code{regression.surrogate_n_comb} is either NULL or a valid integer.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
