% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapr-package.R
\name{release_questions}
\alias{release_questions}
\title{Auxiliary function for the vaeac vignette}
\usage{
release_questions()
}
\description{
Function that question if the main and vaeac vignette has been built using the
\code{rebuild-long-running-vignette.R} function. This is only useful when using devtools to release
\code{shapr} to cran. See \code{\link[devtools:release]{devtools::release()}} for more information.
}
