% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapr-package.R
\name{release_questions}
\alias{release_questions}
\title{Auxiliary function for the vignettes}
\usage{
release_questions()
}
\description{
Ask whether the vignettes have been built using the
\code{rebuild-long-running-vignette.R} script.
This is only useful when using devtools to release \code{shapr} to CRAN.
See \code{\link[devtools:release]{devtools::release()}} for more information.
}
\keyword{internal}
