% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_cpp}
\alias{rss_cpp}
\title{Function for computing sigma_hat_sq}
\usage{
rss_cpp(H, y)
}
\arguments{
\item{H}{Matrix.
Output from \code{\link[=hat_matrix_cpp]{hat_matrix_cpp()}}}

\item{y}{Vector
Representing the (temporary) response variable}
}
\value{
Scalar
}
\description{
Function for computing sigma_hat_sq
}
\author{
Martin Jullum
}
\keyword{internal}
