% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_cpp}
\alias{rss_cpp}
\title{sigma_hat_sq-function}
\usage{
rss_cpp(H, y)
}
\arguments{
\item{H}{Matrix. Output from \code{\link{hat_matrix_cpp}}}

\item{y}{Vector, i.e. representing the response variable}
}
\value{
Scalar
}
\description{
sigma_hat_sq-function
}
\author{
Martin Jullum
}
