% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_copula}
\alias{sample_copula}
\title{Sample conditional variables using the Gaussian copula approach}
\usage{
sample_copula(given_ind, noSamp_MC, mu, Sigma, p, Xtest_Gauss_trans,
  Xtrain, Xtest)
}
\arguments{
\item{given_ind}{Vector. The indices of the features to condition upon.}

\item{noSamp_MC}{Positive integer. Indicating the maximum number of samples to use in the
Monte Carlo integration for every conditional expectation (previously called \code{n_threshold})}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution.
NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{Sigma}{Numeric matrix. (Optional) Containing the covariance matrix of the data generating
distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{p}{Positive integer. The number of features.}

\item{Xtest_Gauss_trans}{Vector with the Gaussian transformed features of the observation whose predictions ought to be explained (test data). Dimension \code{1xp} or \code{px1}.}

\item{Xtrain}{Matrix, data.frame or data.table with the features from the training data}

\item{Xtest}{Matrix, data.frame or data.table with the features of the observation whose predictions ought to be explained (test data). Dimension \code{1xp} or \code{px1}.}
}
\value{
data.table with \code{noSamp_MC} (conditional) Gaussian samples
}
\description{
Sample conditional variables using the Gaussian copula approach
}
\author{
Martin Jullum
}
