% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_copula}
\alias{sample_copula}
\title{Sample conditional variables using the Gaussian copula approach}
\usage{
sample_copula(index_given, n_samples, mu, cov_mat, p, x_test_gaussian,
  x_train, x_test)
}
\arguments{
\item{index_given}{Vector. The indices of the features to condition upon.}

\item{p}{Positive integer. The number of features.}

\item{x_test_gaussian}{Vector with the Gaussian transformed features of the observation whose
predictions ought to be explained (test data). Dimension \code{1xp} or \code{px1}.}

\item{x_test}{Matrix, data.frame or data.table with the features of the observation whose
predictions ought to be explained (test data). Dimension \code{1xp} or \code{px1}.}
}
\value{
data.table with \code{n_samples} (conditional) Gaussian samples
}
\description{
Sample conditional variables using the Gaussian copula approach
}
\examples{
# TODO: Add simple example

}
\author{
Martin Jullum
}
\keyword{internal}
