% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_ctree}
\alias{sample_ctree}
\title{Sample ctree variables from a given conditional inference tree}
\usage{
sample_ctree(tree, n_samples, x_test, x_train, p, sample)
}
\arguments{
\item{tree}{List. Contains tree which is an object of type ctree built from the party package.
Also contains given_ind, the features to condition upon.}

\item{n_samples}{Numeric. Indicates how many samples to use for MCMC.}

\item{x_test}{Matrix, data.frame or data.table with the features of the observation whose
predictions ought to be explained (test data). Dimension \code{1xp} or \code{px1}.}

\item{x_train}{Matrix, data.frame or data.table with training data.}

\item{p}{Positive integer. The number of features.}

\item{sample}{Boolean. True indicates that the method samples from the terminal node
of the tree whereas False indicates that the method takes all the samples if it is less than n_samples.}
}
\value{
data.table with \code{n_samples} (conditional) Gaussian samples
}
\description{
Sample ctree variables from a given conditional inference tree
}
\examples{
# TODO: Add simple example
}
\author{
Annabelle Redelmeier
}
\keyword{internal}
