% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_gaussian}
\alias{sample_gaussian}
\title{Sample conditional Gaussian variables}
\usage{
sample_gaussian(index_mu, n_samples, mu, sigma, p, x_test,
  ensure_condcov_symmetry = F)
}
\arguments{
\item{index_mu}{Vector. The indices of the features to condition upon.}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution.
NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{sigma}{Numeric}

\item{p}{Positive integer. The number of features.}

\item{x_test}{Matrix, data.frame or data.table with the features of the observation whose
predictions ought to be explained (test data). Dimension \code{1xp} or \code{px1}.}

\item{ensure_condcov_symmetry}{Logical. If \code{true}, \code{symmpart} is used on the
conditional covariance matrix to ensure symmetry.}
}
\value{
data.table with \code{n_samples} (conditional) Gaussian samples
}
\description{
Sample conditional Gaussian variables
}
\examples{
# TODO: Add simple example

}
\author{
Martin Jullum
}
\keyword{internal}
