% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_gaussian}
\alias{sample_gaussian}
\title{Sample conditional Gaussian variables}
\usage{
sample_gaussian(index_given, n_samples, mu, cov_mat, p, x_test)
}
\arguments{
\item{index_given}{Vector. The indices of the features to condition upon.}

\item{mu}{Numeric vector. (Optional) Containing the mean of the data generating distribution.
NULL means it is estimated from the data if needed (in the Gaussian approach).}

\item{p}{Positive integer. The number of features.}

\item{x_test}{Numeric matrix of dimension 1 x p.}
}
\value{
data.table with \code{n_samples} (conditional) Gaussian samples
}
\description{
Sample conditional Gaussian variables
}
\examples{
# TODO: Add simple example

}
\author{
Martin Jullum
}
\keyword{internal}
