% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Save the Intermediate Results to Disk}
\usage{
save_results(internal)
}
\arguments{
\item{internal}{List.
Not used directly, but passed through from \code{\link[=explain]{explain()}}.}
}
\value{
No return value (but saves the intermediate results to disk)
}
\description{
Save the Intermediate Results to Disk
}
\keyword{internal}
