% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach.R, R/approach_copula.R,
%   R/approach_ctree.R, R/approach_empirical.R, R/approach_gaussian.R,
%   R/approach_independence.R
\name{setup_approach}
\alias{setup_approach}
\alias{setup_approach.copula}
\alias{setup_approach.ctree}
\alias{setup_approach.empirical}
\alias{setup_approach.gaussian}
\alias{setup_approach.independence}
\title{Set up the framework chosen approach}
\usage{
setup_approach(internal, ...)

\method{setup_approach}{copula}(internal, ...)

\method{setup_approach}{ctree}(
  internal,
  mincriterion = 0.95,
  minsplit = 20,
  minbucket = 7,
  sample = TRUE,
  ...
)

\method{setup_approach}{empirical}(
  internal,
  explain.type = "fixed_sigma",
  explain.eta = 0.95,
  explain.fixed_sigma_vec = 0.1,
  explain.n_samples_aicc = 1000,
  explain.eval_max_aicc = 20,
  explain.start_aicc = 0.1,
  explain.cov_mat = NULL,
  model = NULL,
  predict_model = NULL
)

\method{setup_approach}{gaussian}(internal, mu = NULL, cov_mat = NULL, ...)

\method{setup_approach}{independence}(internal, ...)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}

\item{explain.type}{Character. (default = \code{"fixed_sigma"})
Should be equal to either \code{"independence"},\code{"fixed_sigma"}, \code{"AICc_each_k"} \code{"AICc_full"}. TODO: Describe better what the methods do here.}

\item{explain.eta}{Numeric. (default = 0.95)
Needs to be \verb{0 < eta <= 1}.
Represents the minimum proportion of the total empirical weight that data samples should use.
If e.g. \code{eta = .8} we will choose the \code{K} samples with the largest weight so that the sum of the weights
accounts for 80\\% of the total weight.
\code{eta} is the \eqn{\eta} parameter in equation (15) of Aas et al (2021).}

\item{explain.fixed_sigma_vec}{Numeric. (default = 0.1)
Represents the kernel bandwidth in the distance computation. TODO: What length should it have? 1 or what?
Only used when \code{explain.type = "fixed_sigma"}}

\item{explain.n_samples_aicc}{Positive integer. (default = 1000)
Number of samples to consider in AICc optimization.
Only used for \code{explain.type} is either \code{"AICc_each_k"} or \code{"AICc_full"}.}

\item{explain.eval_max_aicc}{Positive integer. (default = 20)
Maximum number of iterations when optimizing the AICc.
Only used for \code{explain.type} is either \code{"AICc_each_k"} or \code{"AICc_full"}.}

\item{explain.start_aicc}{Numeric. (default = 0.1)
Start value of the \code{sigma} parameter when optimizing the AICc.
Only used for \code{explain.type} is either \code{"AICc_each_k"} or \code{"AICc_full"}.}

\item{explain.cov_mat}{Numeric matrix. (Optional, default = NULL)
Containing the covariance matrix of the data generating distribution used to define the Mahalanobis distance.
\code{NULL} means it is estimated from \code{x_train}.}

\item{model}{Objects.
The model object that ought to be explained}
}
\description{
The different choices of \code{approach} takes different (optional) parameters, which are forwarded from \code{\link[=explain]{explain()}}.
See the various input
}
