% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_computation.R
\name{setup_computation}
\alias{setup_computation}
\title{Sets up everything for the Shapley values computation in \code{\link[=explain]{explain()}}}
\usage{
setup_computation(internal, model, predict_model)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}

\item{model}{Objects.
The model object that ought to be explained.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{predict_model}{Function.
The prediction function used when \code{model} is not natively supported.
See the documentation of \code{\link[=explain]{explain()}} for details.}
}
\value{
List \code{internal}
It holds all parameters, data, and computed objects used within \code{\link[=explain]{explain()}}.
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.
}
\description{
Sets up everything for the Shapley values computation in \code{\link[=explain]{explain()}}
}
