% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapley_weights}
\alias{shapley_weights}
\title{Calculate Shapley weight}
\usage{
shapley_weights(m, N, s, weight_zero_m)
}
\arguments{
\item{m}{Integer. Total number of features}

\item{N}{Integer. Number of combinations}

\item{s}{Integer. Number of chosen}
}
\value{
Numeric
}
\description{
Calculate Shapley weight
}
\author{
Nikolai Sellereite
}
