% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapr}
\alias{shapr}
\title{Create an explainer object with Shapley weights for test data.}
\usage{
shapr(x, model, n_combinations = NULL)
}
\arguments{
\item{x}{An \code{ntrain x p} numeric matrix or data.frame, where \code{p = ncol(x)} (total
number of explanatory variables).Contains the variables used for training the model
(i.e. the explanatory variables). Note that the response variable should not be part of
\code{x}.}

\item{model}{The model whose predictions we want to explain.}

\item{n_combinations}{Integer. The number of feature combinations to sample. If \code{NULL},
the exact method is used and all combinations are considered. The maximum number of
combinations equals \code{2^p}.}
}
\value{
Named list that contains the follwing items:
\describe{
  \item{model}{First item}
  \item{n_combinations}{Second item}
  \item{exact}{Second item}
  \item{n_features}{Second item}
  \item{model_type}{Second item}
  \item{S}{Second item}
  \item{W}{Second item}
  \item{X}{Second item}
  \item{x_train}{Second item}
}
}
\description{
Create an explainer object with Shapley weights for test data.
}
\details{
TODO: Write details about n_combinations
}
\examples{
# Load example data
data("Boston", package = "MASS")
df <- Boston

# Example using the exact method
x_var <- c("lstat", "rm", "dis", "indus")
y_var <- "medv"
df1 <- df[, x_var]
model <- lm(medv ~ lstat + rm + dis + indus, data = df)
explainer <- shapr(df1, model)
print(nrow(explainer$X))
# 16 (which equals 2^4)

# Example using approximation
y_var <- "medv"
x_var <- setdiff(colnames(df), y_var)
model <- lm(medv ~ ., data = df)
df2 <- df[, x_var]
explainer <- shapr(df2, model, n_combinations = 1e3)
print(nrow(explainer$X))

# Example using approximation where n_combinations > 2^m
x_var <- c("lstat", "rm", "dis", "indus")
y_var <- "medv"
df3 <- df[, x_var]
model <- lm(medv ~ lstat + rm + dis + indus, data = df)
explainer <- shapr(df1, model, n_combinations = 1e3)
print(nrow(explainer$X))
# 16 (which equals 2^4)
}
\author{
Nikolai Sellereite
}
