% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapr}
\alias{shapr}
\title{Get Shapley weights.}
\usage{
shapr(x, model, n_combinations = NULL)
}
\arguments{
\item{x}{\code{ntrain x p} Matrix, data.frame or data.table with the features from the training data.}

\item{model}{Model object. Fitted model that is used to produce the predictions}

\item{n_combinations}{Integer. The number of feature combinations to sample. If ´\code{NULL}, the exact method is used and all combinations are considered.}
}
\value{
List
}
\description{
Get Shapley weights.
}
\examples{
TODO: Add simple example
}
\author{
Nikolai Sellereite
}
