% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapr}
\alias{shapr}
\title{Create an explainer object with Shapley weights for test data.}
\usage{
shapr(x, model, n_combinations = NULL, group = NULL)
}
\arguments{
\item{x}{Numeric matrix or data.frame/data.table. Contains the data used to estimate the (conditional)
distributions for the features needed to properly estimate the conditional expectations in the Shapley formula.}

\item{model}{The model whose predictions we want to explain. Run
\code{\link[shapr:get_supported_models]{shapr:::get_supported_models()}}
for a table of which models \code{shapr} supports natively.}

\item{n_combinations}{Integer. The number of feature combinations to sample. If \code{NULL},
the exact method is used and all combinations are considered. The maximum number of
combinations equals \code{2^ncol(x)}.}

\item{group}{List. If \code{NULL} regular feature wise Shapley values are computed.
If provided, group wise Shapley values are computed. \code{group} then has length equal to
the number of groups. The list element contains character vectors with the features included
in each of the different groups.}
}
\value{
Named list that contains the following items:
\describe{
  \item{exact}{Boolean. Equals \code{TRUE} if \code{n_combinations = NULL} or
  \code{n_combinations < 2^ncol(x)}, otherwise \code{FALSE}.}
  \item{n_features}{Positive integer. The number of columns in \code{x}}
  \item{S}{Binary matrix. The number of rows equals the number of unique combinations, and
  the number of columns equals the total number of features. I.e. let's say we have a case with
  three features. In that case we have \code{2^3 = 8} unique combinations. If the j-th
  observation for the i-th row equals \code{1} it indicates that the j-th feature is present in
  the i-th combination. Otherwise it equals \code{0}.}
  \item{W}{Matrix. This matrix is equal to the matrix \code{R_D} in Equation 7 in the reference
  of \code{link{explain}}. The Shapley value for a test observation will be equal to the matrix-vector product
  of \code{W} and the contribution vector.}
  \item{X}{data.table. Returned object from \code{\link{feature_combinations}}}
  \item{x_train}{data.table. Transformed \code{x} into a data.table.}
  \item{feature_list}{List. The \code{updated_feature_list} output from
  \code{\link[shapr:preprocess_data]{preprocess_data}}}
}

In addition to the items above, \code{model} and \code{n_combinations} are also present in the returned object.
}
\description{
Create an explainer object with Shapley weights for test data.
}
\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
  # Load example data
  data("Boston", package = "MASS")
  df <- Boston

  # Example using the exact method
  x_var <- c("lstat", "rm", "dis", "indus")
  y_var <- "medv"
  df0 <- df[, x_var]
  model <- lm(medv ~ lstat + rm + dis + indus, data = df)
  explainer <- shapr(df0, model)

  print(nrow(explainer$X))
  # 16 (which equals 2^4)

  # Example using approximation
  y_var <- "medv"
  model <- lm(medv ~ ., data = df)
  explainer <- shapr(df, model, n_combinations = 1e3)

  print(nrow(explainer$X))

  # Example using approximation where n_combinations > 2^m
  x_var <- c("lstat", "rm", "dis", "indus")
  y_var <- "medv"
  model <- lm(medv ~ lstat + rm + dis + indus, data = df)
  explainer <- shapr(df0, model, n_combinations = 1e3)

  print(nrow(explainer$X))
  # 16 (which equals 2^4)

  # Example using groups
  group <- list(A=x_var[1:2], B=x_var[3:4])

  explainer_group <- shapr(df0, model, group = group)
  print(nrow(explainer_group$X))
  # 4 (which equals 2^(#groups))
}
}
\author{
Nikolai Sellereite
}
