% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{simulateAllTrees}
\alias{simulateAllTrees}
\title{Make all conditional inference trees}
\usage{
simulateAllTrees(given_ind, x_train, comb, minbucket, mincriterion,
  minsplit)
}
\arguments{
\item{minsplit}{}
}
\value{
data.table with \code{n_samples} (conditional) Gaussian samples
}
\description{
Make all conditional inference trees
}
\examples{
# TODO: Add simple example

}
\author{
Annabelle Redelmeier
}
\keyword{internal}
