% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{simulateAllTrees}
\alias{simulateAllTrees}
\title{Make all conditional inference trees}
\usage{
simulateAllTrees(
  given_ind,
  x_train,
  comb_indici,
  comb_mincriterion,
  mincriterion,
  minsplit,
  minbucket,
  use_partykit = "on_error"
)
}
\arguments{
\item{given_ind}{Numeric value. Indicates which features are conditioned on.}

\item{x_train}{Numeric vector. Indicates the specific values of features for individual i.}

\item{comb_indici}{Numeric value. (Optional) Contains the splitting point corresponding to where to change the
\code{comb_mincriterion}.
If \code{NULL}, the \code{mincriterion} is constant for every combination.
This is depreciated and will be deleted soon.}

\item{comb_mincriterion}{Numeric vector. (Optional) Contains the different mincriterions to use for each
combination.
If \code{NULL}, the \code{mincriterion} is constant for every combination.
This is depreciated and will be deleted soon.}

\item{mincriterion}{Numeric value or vector equal to 1 - alpha where alpha is the nominal level of the conditional
independence tests.
Can also be a vector equal to the length of the number of features indicating which mincriterion to use
when conditioning on various numbers of features.}

\item{minsplit}{Numeric value. Equal to the value that the sum of the left and right daughter nodes need to exceed.}

\item{minbucket}{Numeric value. Equal to the minimum sum of weights in a terminal node.}

\item{use_partykit}{String. In some semi-rare cases \code{partyk::ctree} runs into an error related to the LINPACK
used by R. To get around this problem, one may fall back to using the newer (but slower) \code{partykit::ctree}
function, which is a reimplementation of the same method. Setting this parameter to \code{"on_error"} (default)
falls back to  \code{partykit::ctree}, if \code{party::ctree} fails. Other options are \code{"never"}, which always
uses \code{party::ctree}, and \code{"always"}, which alsways uses \code{partykit::ctree}. A warning message is
created whenever \code{partykit::ctree} is used.}
}
\value{
List with conditional inference tree and the variables conditioned/not conditioned on.
}
\description{
Make all conditional inference trees
}
\examples{
# TODO: Add simple example
}
\author{
Annabelle Redelmeier, Martin Jullum
}
\keyword{internal}
