% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{SkipConnection}
\alias{SkipConnection}
\title{A torch::nn_module representing a Skip Connection}
\usage{
SkipConnection(...)
}
\arguments{
\item{...}{network modules such as torch::nn_linear, activation_function(), and memory layers. See \code{get_imputation_networks}.}
}
\description{
Skip-connection over the sequence of layers in the constructor. The module passes
input data sequentially through these layers and then adds original data to the result.
}
