% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shapr}
\alias{summary.shapr}
\title{Summary method for shapr objects}
\usage{
\method{summary}{shapr}(object, digits = 2L, nsmall = 0L, ...)
}
\arguments{
\item{object}{A shapr object.}

\item{digits}{Integer.
Number of significant digits to display.
Defaults to 2.}

\item{nsmall}{Integer.
Minimum digits to the right of the decimal point.
Defaults to 0.}

\item{...}{Currently unused.}
}
\value{
Prints a formatted summary of the shapr object,
and invisibly returns a named list of summary components.
See the details section of \code{\link[=get_results]{get_results()}} for details about each component.
}
\description{
Summary method for shapr objects
}
