% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shapr}
\alias{summary.shapr}
\title{Summary method for shapr objects}
\usage{
\method{summary}{shapr}(x, digits = 4, ...)
}
\arguments{
\item{x}{A shapr object.}

\item{digits}{Integer. Number of digits for printed numbers.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns a list of key summary components.
}
\description{
Summary method for shapr objects
}
