% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shapr}
\alias{summary.shapr}
\title{Summary method for shapr objects}
\usage{
\method{summary}{shapr}(x, digits = 4, ...)
}
\arguments{
\item{digits}{Integer. Number of digits to round numerical outputs.}

\item{...}{Currently unused.}

\item{object}{A shapr object.}
}
\value{
Invisibly returns a named list of summary components.
}
\description{
Summary method for shapr objects
}
