% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shapr}
\alias{summary.shapr}
\title{Summary Method for Shapr Objects}
\usage{
\method{summary}{shapr}(object, digits = 2L, ...)
}
\arguments{
\item{object}{A shapr object.}

\item{digits}{Integer.
(Maximum) number of digits to be displayed after the decimal point.
Defaults to 2.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{summary.shapr}, which is a named list
with the same accessible components as returned by \code{\link[=get_results]{get_results()}}.
See \code{\link[=get_results]{get_results()}} for details about each component.
}
\description{
Provides a formatted summary of a shapr object and returns an object of class
\code{summary.shapr} containing the same information as returned by \code{\link[=get_results]{get_results()}}.
}
