% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict_model.R
\name{test_predict_model}
\alias{test_predict_model}
\title{Model testing function}
\usage{
test_predict_model(x_test, predict_model, model, internal)
}
\arguments{
\item{predict_model}{Function.
The prediction function used when \code{model} is not natively supported.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{model}{Objects.
The model object that ought to be explained.
See the documentation of \code{\link[=explain]{explain()}} for details.}

\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{iter_list}, \code{timing_list},
\code{main_timing_list}, \code{output}, and \code{iter_timing_list}.}
}
\description{
Model testing function
}
\keyword{internal}
