% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{testing_cleanup}
\alias{testing_cleanup}
\title{Cleans out certain output arguments to allow perfect reproducability of the output}
\usage{
testing_cleanup(output)
}
\description{
Cleans out certain output arguments to allow perfect reproducability of the output
}
\author{
Lars Henry Berge Olsen, Martin Jullum
}
\keyword{internal}
