% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{testing_cleanup}
\alias{testing_cleanup}
\title{Clean Out Certain Output Arguments to Allow Perfect Reproducibility of the Output}
\usage{
testing_cleanup(output)
}
\value{
Cleaned up version of the output list used for testthat testing
}
\description{
Clean Out Certain Output Arguments to Allow Perfect Reproducibility of the Output
}
\author{
Lars Henry Berge Olsen, Martin Jullum
}
\keyword{internal}
