% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_cuda}
\alias{vaeac_check_cuda}
\title{Function that checks for access to CUDA}
\usage{
vaeac_check_cuda(cuda)
}
\arguments{
\item{cuda}{Logical (default is \code{FALSE}). If \code{TRUE}, then the \code{vaeac} model will be trained using cuda/GPU.
If \code{\link[torch:cuda_is_available]{torch::cuda_is_available()}} is \code{FALSE}, the we fall back to use CPU. If \code{FALSE}, we use the CPU. Using a GPU
for smaller tabular dataset often do not improve the efficiency.
See \code{vignette("installation", package = "torch")} fo help to enable running on the GPU (only Linux and Windows).}
}
\value{
The function does not return anything.
}
\description{
Function that checks for access to CUDA
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
