% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_epoch_values}
\alias{vaeac_check_epoch_values}
\title{Function that checks provided epoch arguments}
\usage{
vaeac_check_epoch_values(
  epochs,
  epochs_initiation_phase,
  epochs_early_stopping
)
}
\arguments{
\item{epochs}{Integer. The number of epochs to train the final vaeac model.
This includes \code{epochs_initiation_phase}.}

\item{epochs_initiation_phase}{Integer. The number of epochs to run each of the \code{num_vaeacs_initiate}
vaeac models before only continuing training the best one.}

\item{epochs_early_stopping}{Integer. The training stops if there has been no improvement in the validation IWAE
for \code{epochs_early_stopping} epochs. If the user wants the training process to be solely based on this, then \code{epochs}
should be set to a large number.}
}
\value{
The function does not return anything.
}
\description{
Function that checks provided epoch arguments
}
\author{
Lars Henry Berge Olsen
}
