% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_logicals}
\alias{vaeac_check_logicals}
\title{Function that checks logicals}
\usage{
vaeac_check_logicals(named_list_logicals)
}
\arguments{
\item{named_list_logicals}{List containing named entries. I.e., \code{list(a = TRUE, b = FALSE)}.}
}
\value{
The function does not return anything.
}
\description{
Function that checks logicals
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
