% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_masking_ratio}
\alias{vaeac_check_masking_ratio}
\title{Function that checks that the masking ratio argument is valid}
\usage{
vaeac_check_masking_ratio(masking_ratio, n_features)
}
\arguments{
\item{masking_ratio}{Numeric (default is \code{0.5}). Probability of masking a feature in the
\code{\link[=mcar_mask_generator]{mcar_mask_generator()}} (MCAR = Missing Completely At Random). The MCAR masking scheme ensures that \code{vaeac}
model can do arbitrary conditioning as all coalitions will be trained. \code{masking_ratio} will be overruled if
\code{mask_gen_coalitions} is specified.}

\item{n_features}{The number of features, i.e., the number of columns in the training data.}
}
\value{
The function does not return anything.
}
\description{
Function that checks that the masking ratio argument is valid
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
