% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_positive_integers}
\alias{vaeac_check_positive_integers}
\title{Function that checks positive integers}
\usage{
vaeac_check_positive_integers(named_list_positive_integers)
}
\arguments{
\item{named_list_positive_integers}{List containing named entries. I.e., \code{list(a = 1, b = 2)}.}
}
\value{
The function does not return anything.
}
\description{
Function that checks positive integers
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
