% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_probabilities}
\alias{vaeac_check_probabilities}
\title{Function that checks probabilities}
\usage{
vaeac_check_probabilities(named_list_probabilities)
}
\arguments{
\item{named_list_probabilities}{List containing named entries. I.e., \code{list(a = 0.2, b = 0.9)}.}
}
\value{
The function does not return anything.
}
\description{
Function that checks probabilities
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
