% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_extra_functions.R
\name{vaeac_check_use_cuda}
\alias{vaeac_check_use_cuda}
\title{Function that checks for access to CUDA}
\usage{
vaeac_check_use_cuda(use_cuda)
}
\arguments{
\item{use_cuda}{Boolean. If we are to use cuda (GPU) if available. STILL IN DEVELOPMENT!}
}
\value{
The function does not return anything.
}
\description{
Function that checks for access to CUDA
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
