% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_check_verbose}
\alias{vaeac_check_verbose}
\title{Function that checks the verbose parameter}
\usage{
vaeac_check_verbose(verbose)
}
\arguments{
\item{verbose}{Boolean. An integer specifying the level of verbosity. Use \code{0} (default) for no verbosity,
\code{1} for low verbose, and \code{2} for high verbose.}
}
\value{
The function does not return anything.
}
\description{
Function that checks the verbose parameter
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
