% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac_torch_modules.R
\name{vaeac_postprocess_data}
\alias{vaeac_postprocess_data}
\title{Postprocess Data Generated by a vaeac Model}
\usage{
vaeac_postprocess_data(data, vaeac_model_state_list)
}
\arguments{
\item{data}{data.table containing the data generated by a vaeac model}

\item{vaeac_model_state_list}{List. The returned list from the \code{vaeac_preprocess_data} function or
a loaded checkpoint list of a saved vaeac object.}
}
\value{
data.table with the generated data from a vaeac model where the categorical features
now have the original class names.
}
\description{
vaeac generates numerical values. This function converts categorical features
to from numerics with class labels 1,2,...,K, to factors with the original and class labels.
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
