% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_prep_message_batch}
\alias{vaeac_prep_message_batch}
\title{Produce message about which batch prepare_data is working on}
\usage{
vaeac_prep_message_batch(internal, index_features)
}
\arguments{
\item{internal}{List.
Holds all parameters, data, functions and computed objects used within \code{\link[=explain]{explain()}}
The list contains one or more of the elements \code{parameters}, \code{data}, \code{objects}, \code{output}.}

\item{index_features}{Positive integer vector. Specifies the indices of combinations to
apply to the present method. \code{NULL} means all combinations. Only used internally.}
}
\description{
Produce message about which batch prepare_data is working on
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
