% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_train_model_continue}
\alias{vaeac_train_model_continue}
\title{Continue to Train the \code{vaeac} Model}
\usage{
vaeac_train_model_continue(
  explanation,
  epochs_new,
  lr_new = NULL,
  x_train = NULL,
  save_data = FALSE,
  verbose = NULL,
  seed = 1
)
}
\arguments{
\item{explanation}{A \code{\link[=explain]{explain()}} object and \code{vaeac} must be the used approach.}

\item{epochs_new}{Positive integer. The number of extra epochs to conduct.}

\item{lr_new}{Positive numeric. If we are to overwrite the old learning rate in the adam optimizer.}

\item{x_train}{A data.table containing the training data. Categorical data must have class names \eqn{1,2,\dots,K}.}

\item{save_data}{Logical (default is \code{FALSE}). If \code{TRUE}, then the data is stored together with
the model. Useful if one are to continue to train the model later using \code{\link[=vaeac_train_model_continue]{vaeac_train_model_continue()}}.}

\item{verbose}{String vector or NULL.
Controls verbosity (printout detail level) via one or more of \code{"basic"}, \code{"progress"},
\code{"convergence"}, \code{"shapley"} and \code{"vS_details"}.
\code{"basic"} (default) displays basic information about the computation and messages about parameters/checks.
\code{"progress"} displays where in the calculation process the function currently is.
\code{"convergence"} displays how close the Shapley value estimates are to convergence
(only when \code{iterative = TRUE}).
\code{"shapley"} displays intermediate Shapley value estimates and standard deviations (only when \code{iterative = TRUE}),
and the final estimates.
\code{"vS_details"} displays information about the v(S) estimates,
most relevant for \code{approach \%in\% c("regression_separate", "regression_surrogate", "vaeac")}.
\code{NULL} means no printout.
Any combination can be used, e.g., \code{verbose = c("basic", "vS_details")}.}

\item{seed}{Positive integer (default is \code{1}). Seed for reproducibility. Specifies the seed before any randomness
based code is being run.}
}
\value{
A list containing the training/validation errors and paths to where the vaeac models are saved on the disk.
}
\description{
Function that loads a previously trained vaeac model and continue the training, either
on new data or on the same dataset as it was trained on before. If we are given a new dataset, then
we assume that new dataset has the same distribution and one_hot_max_sizes as the original dataset.
}
\references{
\itemize{
\item \href{https://www.jmlr.org/papers/volume23/21-1413/21-1413.pdf}{
Olsen, L. H., Glad, I. K., Jullum, M., & Aas, K. (2022). Using Shapley values and variational autoencoders to
explain predictive models with dependent mixed features. Journal of machine learning research, 23(213), 1-51}
}
}
\author{
Lars Henry Berge Olsen
}
