% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_vaeac.R
\name{vaeac_update_pretrained_model}
\alias{vaeac_update_pretrained_model}
\title{Function that checks and adds a pre-trained \code{vaeac} model}
\usage{
vaeac_update_pretrained_model(parameters)
}
\arguments{
\item{parameters}{List containing the parameters used within \code{\link[=explain]{explain()}}.}
}
\value{
This function adds a valid pre-trained vaeac model to the \code{parameter}.
}
\description{
Function that checks and adds a pre-trained \code{vaeac} model
}
\author{
Lars Henry Berge Olsen
}
\keyword{internal}
