% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{weight_matrix}
\alias{weight_matrix}
\title{Get weighted matrix}
\usage{
weight_matrix(X, use_shapley_weights_in_W = T, normalize_W_weights = T)
}
\arguments{
\item{X}{data.table}

\item{use_shapley_weights_in_W}{Logical}

\item{normalize_W_weights}{Logical}
}
\value{
Matrix
}
\description{
Get weighted matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
