% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{weight_matrix}
\alias{weight_matrix}
\title{Get weighted matrix}
\usage{
weight_matrix(dt, use_shapley_weights_in_w = T,
  normalize_w_weights = T)
}
\arguments{
\item{x}{data.table}
}
\value{
Matrix
}
\description{
Get weighted matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
