% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_setup.R
\name{weight_matrix}
\alias{weight_matrix}
\title{Calculate Weighted Matrix}
\usage{
weight_matrix(X, normalize_W_weights = TRUE)
}
\arguments{
\item{X}{data.table.
Output from \code{\link[=create_coalition_table]{create_coalition_table()}}.}

\item{normalize_W_weights}{Logical. Whether to normalize the coalition weights to sum to 1 for
increased numerical stability before solving the WLS (weighted least squares). Applies to all coalitions
except coalitions \code{1} and \code{2^m}.}
}
\value{
Numeric matrix. See \code{\link[=weight_matrix_cpp]{weight_matrix_cpp()}} for more information.
}
\description{
Calculate Weighted Matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
