% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{weight_matrix}
\alias{weight_matrix}
\title{Calculate weighted matrix}
\usage{
weight_matrix(X, normalize_W_weights = TRUE)
}
\arguments{
\item{X}{data.table}

\item{normalize_W_weights}{Logical. Whether to normalize the weights for the combinations to sum to 1 for
increased numerical stability before solving the WLS (weighted least squares). Applies to all combinations
except combination \code{1} and \code{2^m}.}
}
\value{
Numeric matrix. See \code{\link{weight_matrix_cpp}} for more information.
}
\description{
Calculate weighted matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
