% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weight_matrix_cpp}
\alias{weight_matrix_cpp}
\title{Calculate weight matrix}
\usage{
weight_matrix_cpp(coalitions, m, n, w)
}
\arguments{
\item{coalitions}{List.
Each of the elements equals an integer vector representing a valid combination of features/feature groups.}

\item{m}{Integer.
Number of features/feature groups.}

\item{n}{Integer.
Number of combinations.}

\item{w}{Numeric vector
Should have length \code{n}. \code{w[i]} equals the Shapley weight of feature/feature group combination \code{i},
represented by \code{coalitions[[i]]}.}
}
\value{
Matrix of dimension n x m + 1
}
\description{
Calculate weight matrix
}
\author{
Nikolai Sellereite, Martin Jullum
}
\keyword{internal}
