% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weight_matrix_cpp}
\alias{weight_matrix_cpp}
\title{Calculate weight matrix}
\usage{
weight_matrix_cpp(subsets, m, n, w)
}
\arguments{
\item{subsets}{List. Each of the elements equals an integer
vector representing a valid combination of features/feature groups.}

\item{m}{Integer. Number of features/feature groups}

\item{n}{Integer. Number of combinations}

\item{w}{Numeric vector of length \code{n}, i.e. \code{w[i]} equals
the Shapley weight of feature/feature group combination \code{i}, represented by
\code{subsets[[i]]}.}
}
\value{
Matrix of dimension n x m + 1
}
\description{
Calculate weight matrix
}
\author{
Nikolai Sellereite
}
\keyword{internal}
