% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weight_matrix_cpp}
\alias{weight_matrix_cpp}
\title{Calculate weight matrix}
\usage{
weight_matrix_cpp(features, m, n, w)
}
\arguments{
\item{features}{List. Each of the elements equals an integer
vector representing a valid combination of features.}

\item{m}{Integer. Number of features}

\item{n}{Integer. Number of combinations}

\item{w}{Numeric vector of length \code{n}, i.e. \code{w[i]} equals
the shapley weight of feature combination \code{i}, represented by
\code{features[[i]]}.}
}
\value{
Matrix of dimension n x m + 1
}
\description{
Calculate weight matrix
}
\author{
Nikolai Sellereite
}
